[clang frontend] Pre-process AST locations to make them explicit and easier to deal with. Don't dive into include files.
Summary:
@public
The clang location information is described in an incremental way: each location information is a delta with respect to the previous one in the AST. This is based on a the visit of the AST nodes which corresponds to the order in which the lines are printed with the standard clang AST dump:
  clang -cc1 -ast-dump filename.c

This diff adds a preprocessing phase to the front-end so that location information is composed during a visit, and explicit location information is used instead.

In the case of include files, we report the last known location before including the file.

The current file for a function is the file where it is defined. So if a function is entirely defined in a .h file, then the location information will consistently be about the .h file. If instead a function is defined in the source file being analyzed, and some AST nodes come from macro expansion, line information will refer to the original file.

The front-end tests reveal that the location information was incorrect in a few dot files.

Test Plan: arc unit, after having fixed the wrong location in the existing .dot files