Add @import support for Xcode projects
Summary:
@public
This is a workaround to a clang crash that happens whenever `-fmodules` and `YojsonASTExporter` are used together.
This workaround, uses the `-plugin` argument instead of `-add-plugin` one for the clang frontend, and as a result of that, it overrides the default action of clang, which means no object files are emitted, but just the AST.
To generate the missing data needed by the subsequent building phases of xcodebuild, we run Apple's clang.

Test Plan:
Compiled project containing Pods and `@import`, through a command of the form:
  infer -- xcodebuild -workspace project_name.xcworkspace -scheme project_name -sdk iphonesimulator