[infer][Java] Deal with empty strings '' in the classpath
Summary:
@public
The empty string '' is a valid classpath information. This diff parses it and skips it when detecting the parts of the classpath. This case is happening when analyzing Buck. This one be one possibility why Infer does not load any bytecode in some cases, leading the

  TODO: print error message

failure happening with the release (now `Failed to load any Java source code`).

Test Plan: Infer CI.