[clang] Removing the case of flagging npes in initialisers. This causes many FPs.
Summary:
@public
There are many FPs of the form init method that contains

if ((self = [super initWithFrame:frame])) {
...
}
return self;

then an object being initialised with that constructor and added to an array or dictionary.
There we flag NPE and very likely that won't be a bug. So I'm removing the option for self
to be nil in the constructor, which should solve the problem.

Test Plan: Changed the relevant test.