[infer][java] fix model of InputStreamReader to take into account charset given in lowercase
Summary:
The models for InputStreamReader and OutputStreamWriter are taking into consideration the charset passed as parameter in order to follow the exception branch when the charset is not valid. However, the previsous models were only considering encoding literals with uppercase letters. This diff adds the lowercase encoding names to the list.

Closes https://github.com/facebook/infer/issues/127