Fixing Symexec_memory error that occurs when dereferencing the return value of an undefined function
Summary: @​public Infer previously did not work correctly when a function returns the result of a skip function:

```
retUndef() {
  x = undefined();
  return x;
}

derefUndef() {
  y = retUndef();
  y.doSomething(); // Symexec_memory_error here, prevents spec inference
}
```

The problem is that angelic mode did not know to add the return value of `retUndef()` to the footprint.
This diff fixes the problem by adding return values marked with the `Aundef` attribute to the footprint.
This is done lazily (e.g., a value only gets added to the footprint when you try to deref it).

Reviewed By: @jvillard

Differential Revision: D2444929