Add pointer type to argument of method call
Summary: public
Type of `this` argument to c++ method call has pointer type.
It's either raw pointer (for `->`) or reference (for `.`).

It was already correct in method declaration and method parameters, but it wasn't correct in method calls. Same thing will apply to constructor expressions.

As a result of this change, we won't expand type when calling methods.
Change to ast_expressions.ml fixes problem with autogenerating getters/setters that produced lvalue types after LValueToRValue cast.

Reviewed By: dulmarod

Differential Revision: D2605756

fb-gh-sync-id: 1027600