Adding progress bar to the analysis
Summary: public
Adding a progress bar to the analysis phase to check progress.
The progress bar for the analysing procedures part is just printing dots at the moment and will be
improved soon when we do all the multithreading inside Ocaml.
There is also a no_progress_bar option in the scripts to hide it, for instance in CI.

Reviewed By: cristianoc

Differential Revision: D2625763

fb-gh-sync-id: daf6f96