Moving the property frontend checkers to a dummy procdesc
Summary:
public This puts the bugs found with the checker strong delegate to a dummy method.
The error message will appear in the line of the class implementation definition, since the properties
are likely to be defined in the h file, and getting the reporting in a file that is not the current is
difficult.

Reviewed By: ddino

Differential Revision: D2718016

fb-gh-sync-id: 66273a4