Remove the concept of generated method from the frontend and proc attributes
Summary:
public The concept of generated method is not used anymore, so it is removed.
In general,  I think we have noticed that generating code in the frontend is not the best
idea because it decreases performance, it's better to do it in the backend like with the
getters/setters.

Reviewed By: akotulski

Differential Revision: D2739142

fb-gh-sync-id: 0cf9535