Do not reuse Undef function values
Summary:
Change eradicate handling of complex values so that an unknown function that
has an existing mapping to Undef is treated as if there was no existing
mapping.

Without this change, joining control-flow branches where one called a function
and the other did not resulted in a mapping to Undef. Later calls to the
function would then reuse the Undef mapping.

public

Reviewed By: cristianoc

Differential Revision: D2695548

fb-gh-sync-id: ab69c47