No longer enforce the subtyping rule for the @PerformanceCritical annotation
Summary:
public
The contravariant subtyping rule for the PerformanceCritial annotation was meant to document the code but can be very too verbose on exisiting project. It is also not necessary as we can get this annotation from the supertypes. I am disabling it for now, but keep the code in case we want to revive it at some point in the future.

Reviewed By: sblackshear

Differential Revision: D2750212

fb-gh-sync-id: 2424281