Generalize the detection of assertion failures to any kind of custom errors defined programatically in the orginal source code
Summary:
public
This diff cleans up the detection of assertion failures in C, C++ and Objective C which was previously hacked on top of the tracing mode for Java. The code is also generalized to detect any custom errors which can be defined using the `__infer_fail` builtin, and the case of assertion failure is now just the specific case of translating `assert` using `__infer_fail` directly in the clang frontend.

Reviewed By: jberdine

Differential Revision: D2786574

fb-gh-sync-id: dd1e1cf