Revive the "call by procedure name" mechanism for Java
Summary:
public
In the case of Java, the name of the method is enough to lookup the summary of the callees and run the symbolic execution. This revision separates the case of method call in Java and in C, C++ and Objective C. Most of the code for executing method calls was Clang specific and this is an intermediate step to be able to run the capture and analysis on demand.

Reviewed By: sblackshear

Differential Revision: D2809171

fb-gh-sync-id: da62dce