Perf critical checker: save the chains of calls to expensive methods in the error traces
Summary:
public
Use Infer traces to follow the chain of calls from methods annotated PerformanceCritical to the methods annotated as Expensive.

This contains direct jumps from method definition to method definition. So the traces are of the form:

  method definition of m1 -> definition of m2 -> definition of m3 -> ... -> definition of mN annotated as Expensive

In the next diff, I will make the checker create traces of the form:

  definition of m1 -> call to m2 -> definition of m2 -> call to m3 -> ... -> call to mN -> definition of mN annotated as Expensive

and then simplify the error message whenever the chain of call is longer than, say, 5 calls.

Reviewed By: jberdine

Differential Revision: D2818398

fb-gh-sync-id: c566a44