Always use the current directory as the default classpath
Summary:
public
When no classpath was specified, Infer were considering it to be None instead of the current directory. The consequence is that we were replacing the current directory with the annotation processor for suppress warnings, leading to cases where `javac` was compiling fine but `infer -- javac` was failing with classes not found compilation issues. This diff fixes by always having at least "." in the classpath.

Reviewed By: sblackshear

Differential Revision: D2853035

fb-gh-sync-id: e69db7c