Refactor the worklist data structures of the intraprocedural algorithm as non-global state.
Summary:
public
The Worklist module currently encapsulates a global data structure. Likewise for the global variables for the join state and pathset todo and visited.
This diff refactors the Worklist module into an API where instances of the record can be created and passed around. All the global state is included in the record.
The Worklist data structure never escapes the Interproc module.

Reviewed By: sblackshear

Differential Revision: D2887674

fb-gh-sync-id: 65cb234