Treat xvalues as they were lvalues
Summary:
public
xvalues is concept introduced in C++11. While they are not same as lvalues, they have one common trait:
They have identity which means that:
> it's possible to determine whether the expression refers to the same entity as another expression, such as by comparing addresses of the objects or the functions they identify (obtained directly or indirectly);

It means that as far as backend is concerned, they should be treated in same way. Right now there is no concept of "move" in the backend and so we don't have
to differentiate between them.

Reference:
http://en.cppreference.com/w/cpp/language/value_category

Reviewed By: cristianoc

Differential Revision: D2895593

fb-gh-sync-id: 5101e28