remove subtyping rule for the @PerformanceCritical annotation
Summary:public
Is seems that automatically inheriting annotations like `PerformanceCritical` or `NoAllocation` is the right thing to do in general. Otherwise, we need to enforce sub-typing rules which in the best case just adds a little bit of documentation, but could miss important issues when the code is not fully annotated. I am simplifying this part to avoid adding boilerplate code for the  `NoAllocation` case.

Reviewed By: sblackshear

Differential Revision: D2938627

fb-gh-sync-id: ddb668b
shipit-source-id: ddb668b