Add command line arguments for reactive mode.
Summary:public
Add command-line argument --reactive to enable reactive propagation mode.
When the mode is active, the files changed during compilation are detected, and the analysis propagates reactively starting from the modified files.
The reactive mode allows to analyze a subset of the files in a project and follow their dependencies, without storing the results of previous analyses (specs files). Captured files are preserved from previous runs of the analysis (for example, when the previous analysis was the initial capture), so the mode can be used repeatedly while changing code.

Reviewed By: jvillard

Differential Revision: D2931697

fb-gh-sync-id: 9d6dda0
shipit-source-id: 9d6dda0