Translate ImplicitValueInitExpr
Summary:public This also required a refactoring of InitListExpr.
The idea is that ImplicitValueInitExpr can stand for initialising a whole struct,
so we translating as a list of zero expressions, according to the struct's fields,
which is then paired with a list of field expressions, such that one get a list of
assignment instructions.

Reviewed By: ddino

Differential Revision: D2999875

fb-gh-sync-id: 7f609a0
shipit-source-id: 7f609a0