Prevent assert false from occurring on Java functions in Procname.
Summary:public
Assert false have been observed in Procname when analyzing some C projects.
This diff changes the Procname API to make it safe for Java: the java functions in the module don't assert false now. This takes care of the errors observed in C projects.
The new API forces changes throughout the codebase. In particular, the constant propagation module was making assumptions that it would only be executed on Java code, triggering assert false on C. Now it is safe.
For the remaining functions in the Procname module, those for other languages, a special assert false in Utils is used to print stack traces. This is for future debugging.

Reviewed By: sblackshear

Differential Revision: D3054077

fb-gh-sync-id: a77f1d7
shipit-source-id: a77f1d7