Fix cases of resource leaks not detected when the resource indirectly implements Closeable
Summary:public
Before this diff, the Java frontend was not adding the definition of the inherited interfaces to the type environment, thus failing to answer questions like "does type X implements Closeable". Infer was therefore missing to detect resource leaks when the resource was indirectly implementing Closeable via an intermediate interface.

Reviewed By: sblackshear

Differential Revision: D3067555

fb-gh-sync-id: 86d0760
shipit-source-id: 86d0760