create "install" target
Summary:public
This create `make install` and `make uninstall` targets that can be used to
install Infer on a system. All the files needed are installed under
"/usr/local/lib/infer" (by default, but can be configured with ./configure
flags as usual), pretty much like their git infer/ counterparts. Moreover,
"infer" and "inferTraceBugs" are installed in /usr/local/bin (similarly
configurable).

As a bonus, `opam install infer` and `opam remove infer` now work. To use
these, it's better to pin the git repo with "-k git":

  cd ~/infer/ # where the repo is
  opam pin add -k git --yes infer .

The reason is that otherwise, and only if the repo is not clean, opam will
convert symlinks into regular files, which is not good (compilation will fail).

Reviewed By: jberdine

Differential Revision: D3109337

fb-gh-sync-id: 96c9fd1
fbshipit-source-id: 96c9fd1