Create symlinks in near future when merging
Summary:public
It could happen that the modification times of symbolic links and
original files checked by reactive mode could be exactly the same.  In
this case, some files would perpetually be re-analyzed.

This diff creates symlinks and sets their accessed and modified times
to 1 second in the future so that strict timestamp checking is robust.

Reviewed By: cristianoc

Differential Revision: D3098451

fb-gh-sync-id: 3724468
fbshipit-source-id: 3724468