Remove source file filtering in clang_wrapper
Summary:public
Remove option to filter source files from being analysed from clang wrapper.
The option is unused and there should be better mechanism to do so if we need it.
This will slightly simplify the script and allow to analyze files that are from
outside of "blessed" extensions (such as .cxx or .hpp)

Reviewed By: dulmarod

Differential Revision: D3126330

fb-gh-sync-id: 1628913
fbshipit-source-id: 1628913