modify API to be able to hook extra infer options easily
Summary:
To specify a new flag to run some test using our framework, one currently has
to add a new argument to the method that runs Infer and modify existing methods
to give a default value to that argument everywhere. With this diff, it's
possible to just specify extra arguments as parameters to
`createCInferCommand()` et al.

Reviewed By: martinoluca

Differential Revision: D3522161

fbshipit-source-id: 1499fc2