Encode polarity of predicates into atom tag
Summary:
Change representation of pure predicate applications to distinguish
between positive and negative literals using the Apred and Anpred
constructors instead of a boolean field.

This representation is more compact, and is uniform with the treatment
of equalities and disequalities.  Some code is simpler, but there isn't
much in it.

Reviewed By: cristianoc

Differential Revision: D3669387

fbshipit-source-id: 07cdea6