Adding new mode linters
Summary:
Adding a new mode linters. Now if the analyzer is linters, we do the linters and don't translate,
then, if the analyzer is Infer, we do the translation and the backend and not the linters checks, and the
default is that we do capture, backend and lint checks.

Made the tests separated, which saves time and also shows that the linters mode works.

Reviewed By: jvillard

Differential Revision: D3723472

fbshipit-source-id: 9d828d8