Address capture flakiness when a symbol is multiply defined.
Summary:
When a project defines the same symbol in different files,
the results of the capture will be different depending on
the order in which files are captured.

In particular, if a procedure with the same name is defined in different files,
it will be associated with the last file being captured.
So the results can be different if the files are processed in different order.
In case of parallel capture, this has in effect a non-deterministic behaviour.

This diff defines a canonical file (alphabetically) when a procedure is multiply
defined, so that the procedure is associated to the same file whatever the order.

Reviewed By: jeremydubreil

Differential Revision: D3805002

fbshipit-source-id: 2c561f4