[backend] Remove the global reference DB.current_source and clean up the flow of information about the current source file.
Summary:
The global reference `DB.current_source` is used internally in the module DB, by all the front-ends, and directly and indirectly by the back-end, including saving and restoring the state in case of on-demand procedure calls. In particular, it is heavily used in printing functions.
This diff cleans up the flow of information about what the current file is, making it explicit, and removes the reference.

Reviewed By: jberdine

Differential Revision: D3901247

fbshipit-source-id: ef596bd