[infer][java] When using Buck to analyze Java projects, only report the errors found on the targets passed on the command line
Summary:
Before this diff, Infer was simply going through the list of jar files found in `buck-out` and was loading all the `report.json` files found in those jar files in order to merge them into a final report. The main drawback of this was that removing `buck-out` was mandatory to get accurate results when switching between targets to analyze.

With this diff, we now use the `buck audit classpath` option to get from a list of targets, the list of jar files to load the `report.json` files from. This allows to more easily use Infer from the command line when switch branches from the repositories or when switching between targets to analyze.

Reviewed By: martinoluca

Differential Revision: D3922548

fbshipit-source-id: ec550fa