[travis] use new reason version
Summary:
We updated the version of Reason, which removes the need to pin
reason/merlin/merlin_extend. In fact, the pinning we used to do breaks the
compilation as the version of reason is too old. We don't need to pin anymore
to get the right version of Reason, so simply remove the `opam pin ...` setup
commands (as was done for ./build-infer.sh).

Reviewed By: yunxing

Differential Revision: D4030365

fbshipit-source-id: 19c9a39