[infer][config] fix resolution of symbolic links
Summary: Fix the resolution of symbolic links. The previous version did not work if the path itself was not a symbolic link, but there was a symbolic link somewhere up in the path tree. For example: `path/to/file` where `file` is not a symbolic link, but `to`  is a symbolic link.

Reviewed By: jberdine

Differential Revision: D4062947

fbshipit-source-id: 394221d