[Merge] Use multilink files to speed up merge with buck projects.
Summary:
Merging the results directories of targets on buck projects involved creating symbolic links into buck-out.
The bulk of files are .attr files: one per procedure. Creating these links can be a bottleneck, and the merge phase can be slower than the analysis phases on projects with many procedures.
This diff introduces multilinks to speed up merge.
A multilink is a file `multilink.txt` containing a sequence of paths
```
path/to/file1.ext
path/to/file2.ext
...
```
A multilink file is a compact way to represent a link for each entry.

This diff creates a multilink file for each `attributes/dir` directory, instead of one symbolic link for each file.

Reviewed By: jberdine

Differential Revision: D4067428

fbshipit-source-id: 911f8a9