[quandary] functions that transitively call sinks are sinks, not passthroughs
Summary:
Previously, we recorded direct sinks as sinks and transitive sinks as passthroughs. This makes it difficult to create an expanded interprocedural trace when recording an error because we can't distinguish between sinks (which we want to expand) and passthroughs (which we don't). This diff changes recording of sinks so that a sink is now the *last* function in a trace to call a sink. To find out what the original sink was, the summary for the transitive sink in the trace will now need to be (recursively) expanded until we bottom out in the original sink.

Will do the same for sources in a follow-up diff.

Reviewed By: cristianoc

Differential Revision: D4103759

fbshipit-source-id: 6f435f5