[clang] run assembly commands with the fcp clang
Summary:
The build system may expect the assembly commands to run. The only issue with
assembly commands is that we shouldn't attach the plugin to them.

This diff also moves the logic of what to capture to the `Capture.capture`
function to be able to reuse code from Capture. This makes sense because the
Capture module is the one with the knowledge of what to actually capture or
not.

Reviewed By: akotulski

Differential Revision: D4096019

fbshipit-source-id: 7fc99e1