[siof] convert domain to sink trace of pvar's
Summary:
The Quandary-style traces are too general for checkers like SIOF.
This diff adds a "suffix abstraction" of the trace for analyses that just care about sinks.
To show how to use it, we add it to SIOF.

Note: this diff converts the domain, but isn't actually doing the fancier reporting yet.
That will come in a future diff.

Reviewed By: jvillard

Differential Revision: D4124881

fbshipit-source-id: 5b9fd07