[GuardedBy] Model of trylock in ReentrantLock.java
Summary:
The way interfaces are dealt with led to a false positive,
where tryLock() works OK for a Lock but not for a ReentrantLock.
The solution is just to provide the model.

While I am at it I am adding some more standard tests for Lock and ReentrantLock, which were not present.

Reviewed By: sblackshear

Differential Revision: D4204551

fbshipit-source-id: 9b6de28