[C++] Don't ignore header files passed in --changed-files-index
Summary:
Whenever header file is in changed-files-index, it should be captured and analyzed on demand.
It was already being captured, but ondemand analysis wasn't triggered for code in header file. This diff does it.
Use hacky header->source mapping to go from header to source (cluster) and then analyze everything in that cluster (inlucing code coming from header)

Reviewed By: jberdine

Differential Revision: D4265495

fbshipit-source-id: 61606f4