[IR] add Static_local_var pvar kind and use it in clang
Summary:
Remember which globals are static locals.

It's useful to distinguish those from global variables in objc and in the SIOF
checker. Previously in ObjC we would accomplish that by looking at the name of
the variable, but that wouldn't work reliably in C++. Keep the old method around for
now as the way we deal with static locals in ObjC needs some fixing.

Reviewed By: akotulski

Differential Revision: D4198993

fbshipit-source-id: 357dd11