[thread-safety] use ownership info to avoid false alarms
Summary:
Maintain an "ownership" set of access paths that hold locally allocated memory that has not escaped.
This memory is owned by the current procedure, so modifying it outside of synchronization is safe.
If an owned access path does escape to another procedure, we remove it from the ownership set.

Reviewed By: peterogithub

Differential Revision: D4320034

fbshipit-source-id: 64f9169