[quandary] support tainted formals
Summary:
We currently can only model the return values of functions as sources.
In order to model inputs of endpoints as sources, we need the capability to treat the formals of certain functions as sources too.
This diff adds that capability by adding a function for getting the tainted sources to the source module, then using that info in the analysis.

Reviewed By: jeremydubreil

Differential Revision: D4314738

fbshipit-source-id: dd7d423