[traces] adding Source.Make functor for easier source creation
Summary:
There's a lot of boilerplate work to be done when adding a new kind of source.
This diff tries to reduce the boilerplate by making a functor do all the work.
The functor:
(1) adds a notion of "footprint kind" to the source
(2) packages the source with a call site

Reviewed By: jvillard

Differential Revision: D4349224

fbshipit-source-id: 5e1701a