[reporting] Fail loudly when we can't log an error
Summary:
When you try to log an error on a procedure P and a summary for P doesn't exist, the error gets quietly dropped on the floor.
But we should fail loudly instead, because this should only happen in the case of a user error.
Got burned by this today; I was trying to log an error on the *caller* of `Integer.parseInt`, but was accidentally logging it to `Integer.parseInt` itself instead.
Since no summary for that method exists, my error wasn't appearing.

Reviewed By: jvillard, jeremydubreil

Differential Revision: D4355546

fbshipit-source-id: db2a0e6