[infer][backend] Add the distinction between models and regular procedure as part of the attributes
Summary: Adding the information that a procedure has been modelled as part of the attributes, during the translation, instead of getting this information from where is the summary loaded from. This is more consistent with the use of the attributes in other parts of the analysis, but is also useful in the context of the lazy dynamic dispatch algorithm where the procedures, including the models, are cloned and reanalyzed with more specialized parameters. The information about whether a procedure is a model must persist when cloning the procedures.

Reviewed By: sblackshear

Differential Revision: D4356892

fbshipit-source-id: 40ff5ca