[log] no dummy init, no leaked channels
Summary:
Remove the need for a dummy initialization of log files.

The fact that we were not setting log files in some cases doesn't seem to be
relevant so I killed it. I observed no difference in output on simple clang and
javac examples. It will be easy to restore a better version of it in the next
diff if needed.

Also fix an fd leak: when opening new log files, previous ones were not being
flushed and closed (except at exit).

Reviewed By: jberdine

Differential Revision: D4365992

fbshipit-source-id: 940bc16