[cli] do not complain about "arguments must be .specs file" from non-toplevel invocations
Summary:
This error message is confusing when the user is not actually running
InferPrint, eg `infer foo`: `Load Error: file foo: arguments must be .specs
files`.

With this diff, we don't get any error for `infer foo`, which is not great
either and will need to be addressed (do we support all the python arguments in
OCaml now too and are able to turn on argument parsing errors in OCaml land?).

Reviewed By: jberdine

Differential Revision: D4397765

fbshipit-source-id: e7ca48f