[checkers] Add support for html output in checkers, starting with Eradicate
Summary:
Make the html output available to checkers when -g is used on the command-line.
A checker needs to call a function to start and finish the processing of each node,
and add prints during the processing.
This diff illustrates the case for Eradicate, by adding printing of the pre-state
and post-states.

Reviewed By: sblackshear

Differential Revision: D4421379

fbshipit-source-id: 67501ba