[thread-safety] don't warn on writes rooted in a parameter owned by the callee
Summary:
In code like

```
foo() {
  Object local = new Object();
  iWriteToAField(local);
}
```
, we don't want to warn because the object pointed to by `local` is owned by the caller, then ownership is transferred to the callee.

This diff supports this by introducing a notion of "conditional" and "unconditional" writes.
Conditional writes are writes that are rooted in a formal of the current procedure, and they are safe only if the actual bound to that formal is owned at the call site (as in the `foo` example above).
Unconditional writes are rooted in a local, and they are only safe if a lock is held in the caller.

Reviewed By: peterogithub

Differential Revision: D4429131

fbshipit-source-id: 2c6112b