[thread-safety] allow ownership to be transferred via return
Summary:
Previously, we would correctly be silent on code like `x = new T(); x.f = ...`, but would wrongly warn on code like `x = makeT(); x.f = ...`.
The reason is that we only allowed ownership through direct allocation.
This diff adds a boolean that specifies whether the return value is owned as part of the summary.
This allows us to correctly handle many common cases of (transitively) returning a freshly allocated object, but still won't work for understanding that ownership is maintained in examples like
 `x = new T(); y = id(x); y.f = ...`.

Reviewed By: jvillard

Differential Revision: D4456864

fbshipit-source-id: b5eec02