[tests] disregard traces in Java harness test
Summary:
I couldn't figure out why, but from within an infer release the traces we get
for this test are different than the expected ones. This is even consistent
across osx and linux.

In order to restore sanity, let's just hide this incomprehensible fact. Let's
come back to it if more tests exhibit this, maybe traces are not guaranteed to
be exactly the same across runs.

Reviewed By: sblackshear

Differential Revision: D4559405

fbshipit-source-id: dd88c59