[infer][java] avoid redundant calls to the function procedure_should_be_analyzed for Infer.SL
Summary:
This simplifies a bit the code to run the analysis on all the prcedures in the cluster. Before, the functions procedure_should_be_analyzed, which loads the attributes, and get_proc_desc were called twice for the analysis of every procedure.

The objective is to remove the calls to procedure_should_be_analyzed and hide it from the ondemand API since it is already called before the analysis of every procedure.

Reviewed By: sblackshear

Differential Revision: D4553397

fbshipit-source-id: 02cffaf