[access paths] kill type of field access and replace with utility functions for recovering type
Summary:
We waste a lot of space storing the types of field accesses and comparing them sets/maps with access paths.
Yet almost none of the code ever looks at these types (only a tiny piece of code in thread-safety).
If we know the base type, we have enough information to recover the type of the field.
Let's do that instead.

Reviewed By: jeremydubreil

Differential Revision: D4567996

fbshipit-source-id: e7fd2da