[test] make tests pass in clang-only or java-only mode
Summary:
This is useful to make sure `./build-infer.sh clang` and `./build-infer.sh
java` work correctly.

I had to move one unit test inside a new unit/clang/ directory used only if clang is enabled, and create a stub for it in unit/clang_stubs/ for when clang is disabled.

Renamed `OCAML_SOURCES` to the more consistent and descriptive `OCAML_CONFIG_SOURCES`.

Reviewed By: jeremydubreil

Differential Revision: D4571997

fbshipit-source-id: 9502114