[infer][backend] remove the concept of timestamp in summaries and use the summary status instead
Summary:
With the ondemand analysis framework, the concept of timestamp was only being use to check if a procedure has already been analyzed. There was already a concept of "active" procedure for the procedure that were already being analyzed. This revision removes the concept of timestamp and merge it with the concept of analysis status.

This can be simplified further once the analysis always goes through `Ondemand.analyze`.

Reviewed By: cristianoc

Differential Revision: D4610371

fbshipit-source-id: 0fc516b