[infer][ondemand] change the type of analyze_ondemand to return the analysis summary
Summary:
This is part of the plan to have every checker take a summary as input, and return the updated sumamry as output. Doing so, we can run all the registered checkers in sequence for every method

This diff change the type of `Ondemand.analyze_ondemand` to return the analysis summary.

Reviewed By: sblackshear

Differential Revision: D4626918

fbshipit-source-id: f8ad928