[clang] add clang models to `CLANG_DEPS`
Summary:
This is helpful to make sure tests are up to date wrt the models.

Also made the Java deps depend on the models.jar instead of the model sources
as that's what the tests will be using. In particular, updating the sources of
the models will not update the results of a test unless someone rebuilds
models.jar, so rerunning the tests when the models haven't been rebuilt is
useless.

Reviewed By: akotulski

Differential Revision: D4635129

fbshipit-source-id: 75b4ab6