[thread-safety] AccessDomain for better tracking of writes
Summary:
In order to be able to report races like

```
synchronized write() {
  this.f = ...
}

read() {
 return this.f;
}
```

, we need to track writes that happen inside of synchronization as well as writes that happen outside of synchronization.

This diff takes a step toward making that possible by defining an "AccessDomain" mapping a precondition for the safety of a write ( {Safe, SafeIf i, Unsafe} =~ {true, owned(i), false} ) to a set of writes that are safe if the precondition will hold.
We're not actually tracking safe writes yet, but this domain will make it easy to do so.

This also lets us kill the conditional writes/unconditional writes combo, which was a bit clumsy

Reviewed By: peterogithub

Differential Revision: D4620153

fbshipit-source-id: 2d9c5ef