[Eradicate] Improve detection of circularities in field initialization
Summary:
Eradicate detects circular field initializations (e.g. a field initialized with itself)
by checking in the typestate at the end of the constructor whether the origin
of the field is a field name in the current class.
This has the problem that the following initialization pattern is not recognized as correct:
  C(C x) { this.field = x.field }

To fix the issue, the origin information for field accesses x.f is extended
with the origin information of the inner object x.
Circularities are detected if the origin of x is "this".

Reviewed By: jberdine

Differential Revision: D4672472

fbshipit-source-id: 9277bdd