[backend] Add support for command-line option --per-procedure-parallelism
Summary:
Add a new command-line option `--per-procedure-parallelism`, to change the granularity of parallelism of the analysis from file to procedure.
This is intended for `--reactive` mode where e.g. a single file is changed and the analysis currently uses just one core.
When the option is used, the Makefile mechanism is replaced by using forking instead.
The parent process does as little allocation as possible, to avoid taxing the kernel.

Caveats:
- Not active in Java, (issues with camlzip).
- Not active in checkers, yet.

Example use:
```
infer --reactive --changed-files-index index.txt --per-procedure-parallelism -- analyze
```

Reviewed By: jberdine

Differential Revision: D4634884

fbshipit-source-id: e358c18