[cli] deprecate multiletter short options
Summary:
We want to stop supporting '-multiletter' options altogether in the future as
they are not standard.

For boolean switches, we used to have '-short' for true and '-nshort' for
false, but that second form is necessarily > 1 letters, which doesn't work in
this scheme. Replace these by '-s' for true and '-S' for false.

Reviewed By: jberdine

Differential Revision: D4674079

fbshipit-source-id: 95bacfe