Revert "[infer][java] add an option to disable the creation of the human readable list of issues found"
Summary:
This option is not needed anymore as it was introduced to counter an uncovered
perf issue with creating human readable reports. The perf issue has been
addressed.

Instead of this option, one can use `infer --report-hook /bin/true ...` to
disable reporting. However, right now the Buck integration doesn't honor it so
this would need to be fixed to be a true equivalent of `--disable-bug-list`.

Reviewed By: jberdine

Differential Revision: D4712877

fbshipit-source-id: a09304f