[clang][AST] Fix wrong type in translation of NoOp cast and MaterializeExpr
Summary:
This issue was spotted in the wild. There may be more of those, unfortunately it's hard to predict

More general problem is that types in infer frontend diverge from clang's types for DerivedToBase cast.
Then, infer uses types from clang anyway and that confuses backend. Getting it always right is very hard

Reviewed By: jvillard

Differential Revision: D4754081

fbshipit-source-id: 5fb7069