[infer][java] Simplify the ThreadSafety use of Ondemand
Summary:
We can simplify the code now that the procedure callback are always executed through Ondemand. The procedure callback is still registered for Ondemand analysis by the time we run the cluster callbacks. This allows to run allows to run `Summary.read_summary`, which may run the analysis on-demand, while collecting the summaries for reporting errors.

This allows further simplifications of the Ondemand API.

Reviewed By: sblackshear

Differential Revision: D4764251

fbshipit-source-id: d0bdda4