[build] fail early if running `make test` with only some analyzers, add `make config_tests`
Summary:
We shouldn't encourage contributors to run only a subset of the tests with
`make test`, but it's more helpful to complain clearly should they try to do
so.

Now `make test` will ensure that both the clang and the Java analyzers are
enabled, otherwise it will refuse to run and suggest to run `make config_tests`
instead.

It can still be useful to be able to run only the Java or Clang tests. The diff
also fixes the cases where that previously failed.

Fixes #634

Reviewed By: jberdine

Differential Revision: D4826515

fbshipit-source-id: 4b76029