[quandary] don't add callee-local state to the caller
Summary:
If we couldn't project the callee access path into the caller during summary application, we still added the corresponding trace to the caller state.
This was wasteful; it just bloats the caller with state it will never look at.
Fixed it by making `get_caller_ap_node` return `None` when the state won't be visible in the caller.

Reviewed By: jeremydubreil

Differential Revision: D4727937

fbshipit-source-id: 87665e9