[test] test the correct version of toplevel.cmo
Summary:
There are 2 possible places where to find toplevel.cmo:
- infer/_build/infer/toplevel.cmo, built by `make toplevel`
- infer/_build/test/infer/toplevel.cmo, built by `make test`

The test loading and running toplevel.cmo would give priority to the non-test
version of toplevel.cmo, which might be stale because `make test` doesn't
rebuild it.

This diff makes sure we use the test version of toplevel.cmo in `make test`.

Reviewed By: akotulski

Differential Revision: D4960550

fbshipit-source-id: 9c817ba