Minimal thread safety analysis for C++
Summary:
This is a minimal change to (poorly) recognize and model std::mutex
lock and unlock methods, and to surface all thread safety issues for
C++ based on the computed summaries with no filtering.

This ignores much of the Java analysis, including everything about the
Threads domain.  The S/N is comically low at this point.

Reviewed By: sblackshear

Differential Revision: D5120485

fbshipit-source-id: 0f08caa