[oops] do not report in analysis sub-processes
Summary:
Introduce a new option `--no-report` (conversely `--report`) to stop reporting
after the analysis. This is useful to call sub-`infer-analyze` processes with
as they shouldn't compute "result.json" or report bugs to stderr. This bug
would only manifest itself when per-procedure parallelism disabled, which
explains why it was noticed only on Java.

Reviewed By: jberdine

Differential Revision: D5182110

fbshipit-source-id: a892470