[infer][ondemand] rename the logging functions to outline the deprecated ones
Summary:
Now that we can run several inter-procedural analyses at the same time, we should no longer use the function `Reporting.log_error_deprecated` as it logs the errors in the specs table. This specs table is normally used for caching and will be deprecated in favor of having a cache summaries for the callees in the `Ondemand` module (to avoid deserialising a callee more than once within the same process).

This revision just renames the reporting functions.

Reviewed By: sblackshear

Differential Revision: D5205009

fbshipit-source-id: b066549