[infer][java] use a multi-target Buck project to test the Buck integration for Java
Summary:
:
What is relevant for the Buck integration is not the list of bugs that we find in a single target, which is essentially identical to testing `infer -- javac ...`, but to make sure that we still find the issues that are involving several Buck targets, and later other things like the caching mechanism.

This should also make the tests faster.

Reviewed By: jberdine, jvillard

Differential Revision: D5250205

fbshipit-source-id: 7f66b68