[checkers] use liveness analysis to create dead store checker
Summary:
Pretty basic: warn when we see an assignment instruction `x = ...` and `x` is not live in the post of the instruction.
Only enabled for Clang at the moment because linters already warn on this for Java. But we can enable it later if we want to (should be fully generic).

Reviewed By: jeremydubreil

Differential Revision: D5450439

fbshipit-source-id: 693514c