[diff] support for --gen-previous-build-command-script
Summary:
This allows the user to specify a different command to build the current version of the project vs to build the previous version of the project. Here's an example:

```
infer diff --gen-previous-build-command-script "echo clang -c hello.c" ... -- clang -c hello2.c
```

By default the two build commands are the same. If the script is to be used for both the current and the previous versions, then it's on the user to run it once first, eg:

```
infer diff --gen-previous-build-command-script "./myscript.sh previous" ... -- $(./myscript.sh current)
```

Reviewed By: martinoluca

Differential Revision: D5500989

fbshipit-source-id: 7374b44