[models] avoid race conditions when building and installing the models
Summary:
Saw these two types of errors before (but they're hard to reproduce locally) when building the models:

- `ERROR: Zip.Error("/mnt/btrfs/trunk-git-infer-739-1503054473/infer/bin/../lib/java/models.jar", "", "end of central directory not found, not a ZIP file")`. I think this means infer reads a partially-written models jar. We shouldn't try to load this in models mode.
- `install` would complain that the destination already exists. I think this can only happen if there's a race and the file gets created between when install first checks and when it tries to write to it.

This made me realise that the some of the models are computed in C and C++ mode
and we pick one computed spec arbitrarily. That sounds a bit dodgy but at least
now we do so in a non-racy way.

Reviewed By: jeremydubreil

Differential Revision: D5658389

fbshipit-source-id: 8077279