[quandary] apply summary for sinks
Summary:
A function can both be a sink and propagate source info, but we currently ignore the summary for any function that is also a sink.
This will cause us to under-report for (e.g.) `src1 = source(); src2 = strcpy(dest, src1); exec(src2)`.
This is both a potential buffer overflow and a potential shell injection, but we won't report the second issue.

Reviewed By: jberdine

Differential Revision: D5676167

fbshipit-source-id: 232ab2f