simpler logic for determining if an attribute is "fresher" than another
Summary: This is almost equivalent to the previous one except in the case where new and old are both undefined: before, we would just pick "old", but now we pick the biggest according to their source files. I think the previous behaviour was a bug because it was more non-deterministic.

Reviewed By: jberdine

Differential Revision: D5649481

fbshipit-source-id: aeb527d