fix infinitely-expanding types in the backend
Summary:
Simple instance of the problem: analyzing the following program times out.

```
#include <tuple>
void foo() {
  std::tuple<std::tuple<int>> x;
}
```

Replacing `std::tuple<std::tuple<int>>` by `std::tuple<int>` makes the analysis
terminate.

In the AST, both tuple<tuple<int>> and tuple<int> have the same template
specialization type: "Pack" (which means we're supposed to go look into the
arguments of the template to get their values). This is not information enough
and that's the plugin fault.

On the backend side, this means that two types have the same Typ.Name.t, namely
"std::tuple<_>", so they collide in the tenv. The definition of
tuple<tuple<int>> is the one making it into the tenv. One of the fields of the
corresponding CxxRecord is of type "tuple<int>", which we see as the same
"tuple<_>", which causes the loop.

update-submodule: facebook-clang-plugins

Reviewed By: mbouaziz

Differential Revision: D5775840

fbshipit-source-id: 0528604