[thread-safety] new threads domain
Summary:
Previously, we just tracked a boolean representing whether we were possibly on the main thread (true) or definitely not on the main thread (false).
In order to start supporting `Thread.start`, `Runnable.run`, etc., we'll need something more expressive.
This diff introduces a lattice:

```
   Any
  /   \
Main  Background
 \    /
 Unknown
```

as the new threads domain. The initial value is `Unknown`, and we introduce `Main` in situations where we would have introduced `true` before.
This (mostly) preserves behavior: the main difference is that before code like

```
if (*) {
  assertMainThread()
} else {
  x.f = ...
}
```

would have recorded that the access to `x.f` was on the main thread, whereas now we'll say that it's on an unknown thread.

Reviewed By: peterogithub

Differential Revision: D5860256

fbshipit-source-id: efee330