[cleanup] remove some needless language checks in checkers callbacks
Summary:
The only language types we have are Java/Clang/Python. The unit of analysis is a source file, and you can't write a source file that mixes two or more of these languages (to the best of my knowledge).
This diff simplifies using the assumption that all procedures in a file are written in the same language.

Reviewed By: jeremydubreil

Differential Revision: D5886942

fbshipit-source-id: 88c3759