[infer][checkers] simplify the lookup of the procedure description when executing callbacks
Summary:
The previous version of the code was trying to lookup from disk the procedure description of the procedure to analyze, which was in fact already loaded in memory.

This diff fixes one of the issues preventing the bi-abduction to run as a checker when using the lazy dynamic dispatch algorithm.

Reviewed By: sblackshear

Differential Revision: D6038210

fbshipit-source-id: 10a98ee