[cli] introduce `-a biabduction` and make `-a infer` an alias to it
Summary:
One day `-a infer` will alias `-a checkers` so for now create another, more
explicit analyzer name that can be used to migrate progressively. For instance,
after this commit what should continue to use `-a biabduction` can change to
that, so that when `-a infer` becomes an alias to `-a checkers` it can keep
working.

Reviewed By: jeremydubreil

Differential Revision: D6051146

fbshipit-source-id: 1ef4c34