[cli] introduce `--linters` and `--capture`
Summary:
Linters are now considered a "checker", like backend checkers. This makes, eg,
`--racerd-only` disable the linters, which is more intuitive.

We can now express `-a linters` and `--clang-frontend-action` in terms of these
two new options. For instance, `-a linters --clang-frontend-action lint` is the
same as `--linters-only --no-capture`.

This is another step in the direction of getting rid of `--analyzers`.

Reviewed By: dulmarod

Differential Revision: D6147387

fbshipit-source-id: 53622b2