[inferbo] Use a dispatcher for models
Summary:
:
As we want to model many C++ methods, using a lot of matchers with `if / else if` will be tiring.
This diff introduces a dispatcher which is a nicer way to write the same thing.
No new model for now, just a refactoring.

Ideally we'd need a parser generator for C++ names...

Reviewed By: jvillard

Differential Revision: D6209234

fbshipit-source-id: 49fae5e