[infer] always store the procedure description in the summary
Summary:
We need to use the procedure description of the callees for lazy dynamic dispatch and for the resolution of the lambda. We may also need this information in other analyses, e.g. for RacerD. This diff makes the procedure description of the callees as part of the summary.

The procedure description has been part of the summary for a while already without noticeable decrease in performance.

Reviewed By: mbouaziz

Differential Revision: D6322038

fbshipit-source-id: 84101cb