[infer][PR] [C/C++] Encode filename into the Procname of global initializers
Summary:
This resolves #796 . Effectively it adds file specific suffix to name of all global initializers (so initializersof two global variable of the same name will have unique Typ.Procname). which is the same rule as currently used by constructing Procname for the static functions.  However this change applies to initializers of all global variables and not just static (arguably it's a right thing. since GCC used to allow multiple global variables with the same name).

Consequences of this change that it becomes impossible to know name of generated initialization function of global ('extern') variables. However get_initializer_pname function is only referenced by the frontend (when creating initializer for the defined global  variables) and by the SIOF checker.
Closes https://github.com/facebook/infer/pull/801

Reviewed By: jvillard

Differential Revision: D6335034

Pulled By: dulmarod

fbshipit-source-id: 1a92c08