[lock-consistency] Only suppress truncated traces in Java
Summary:
In C++ some modeled functions have definitions, which leads to traces
that contain an access from the modeling, but continue on into the
implementation of the modeled function. Such traces appear the same as
those that are truncated due to limitations of the buck integration in
the Java analysis. Since all Java models are for functions without
definitions in the code base, this diff limits the truncated trace
suppression to the Java analysis.

Reviewed By: sblackshear

Differential Revision: D6373793

fbshipit-source-id: 1f01509