Fix unique_ptr model
Summary:
Our model of unique_ptr and shared_ptr relied on the fact that we could C-style cast a pointer to the internal pointer type used in the smart pointer.
This is wrong when the smart pointer is used with a custom deleter that declares its own pointer type whose is not constructible from just a single pointer.

Reviewed By: dulmarod

Differential Revision: D6496203

fbshipit-source-id: 1305137