[infer][scheduler] no need to filter the defined procedures when running the checkers
Summary: Filtering the defined procedure at this level is not necessary. This check already happens when running the analysis in Ondemand. This could also cause flakiness if the "definedness" here does not agree with the check done in Ondemand. The fact that the analysis of a procedure is triggered from the top-level iteration or on-demand when analyzing another procedure is not deterministic.

Reviewed By: sblackshear

Differential Revision: D6575057

fbshipit-source-id: ff0bc2d