[buck] delete results dir inside buck-out in infer children processes
Summary:
The infer results directories in buck-out/ are "cleaned up" to avoid polluting
the Buck cache with too much data or non-deterministic data. In particular, the
runstate is deleted, which confused subsequent infer processes trying to read
the pre-existing results directory.

Add a special case in infer to delete pre-existing results directories in
buck-out instead of trying to load their state.

Reviewed By: mbouaziz

Differential Revision: D6845128

fbshipit-source-id: 5c716aa