[java] don't warn on unclosed subclasses of non-resource types
Summary:
We already knew not to warn when non-resource `Closeable`'s like `ByteArrayOutputStream` weren't closed, but we still warned on their subtypes.
This diff fixes that problem by checking subclasses in the frontend.
This also removes the need for Java source code models of non-resource types, so I removed them.

Reviewed By: jeremydubreil

Differential Revision: D6843413

fbshipit-source-id: 60fe7fb