[infer][java] translate the annotations of the callees
Summary:
This allows Eradicate to lookup the annotations from the classpath and without requiring the code in the classpath to have been previously analyzed. The benefit is that source files can be analyzed independently of each other as long as the classpath is known.

The main goal is to run be able to run Eradicate as a linter without losing warnings.

We may have to add some more models of the standard libraries as no `Nullable` on a parameter does not necessarily mean that the method does not accept `null`.

Reviewed By: sblackshear

Differential Revision: D6921720

fbshipit-source-id: f525269