[clang] add id -> pvar bindings to C++ lambda capture
Summary: In Obj-C blocks, we explicitly insert reads of the captured vars. This does the same thing for C++. For example, `foo() { int x = 1; [x]() { return x; } }` would previously not contain a read of `x` in `foo`. Now, we'll create a temporary that reads from `x` and pass it to the closure value.

Reviewed By: dulmarod

Differential Revision: D6939997

fbshipit-source-id: f218afc