[inferbo] Prune by assert
Summary:
It prunes abstract memories on `assert` commands.

Problem: Since the assert command is sometimes translated to two
sequential `if` statments, it was not able to prune the memory
precisely at `assert` commands in Inferbo---the pruned memory at the
first branch was joined before the second branch.

Solution: To avoid losing the pruning information at the first branch,
now, it records which locations are pruned at the first branch and
applies the same pruning at the next branch if they have
semantically the same condition.

Reviewed By: mbouaziz

Differential Revision: D6895919

fbshipit-source-id: 15ac1cb