[ownership] prototype of ownership analysis for tracking reads/moves
Summary:
A simple intraprocedural analysis that tracks when a storage location is read or deleted.
For now, this works only with local variable storage locations; field and array accesses are ignored.

In order to test this, I added a new "use-after-lifetime" warning. It complains when a variable is read or deleted after it has already been deleted.

Reviewed By: jeremydubreil

Differential Revision: D6961314

fbshipit-source-id: 75e95a2