[liveness] don't report on dead stores of variables captured by reference in a lambda
Summary:
You can capture a variable by reference in a lambda, assign to it, and then invoke the lambda.
This looks like a dead store from the perspective of the current analysis.

This diff mitigates the problem by computing an additional analysis that tracks variables captured by ref at each program point.
It refuses to report a dead store on a variable that has already been captured by reference.

Later, we might want to incorporate the results of this analysis directly into the liveness analysis instead of just using it to gate reporting.

Reviewed By: jeremydubreil

Differential Revision: D7090291

fbshipit-source-id: 25eeffa