[sqlite] migrate tenv to sqlite
Summary:
Instead of storing the type environment in infer-out/captured/foo.c/foo.c.tenv,
store it in the `source_files` table of the SQLite db. This limits the number
of files we create on disk.

The "file local" type environemnts are specific to the clang integration. For
Java, there is a "global tenv" file. Instead of matching on string names, this
diff also makes the API of `Tenv` reflect this situation.

The global tenv is serialized as a separate file in "infer-out/.global.tenv"
instead of "infer-out/captured/global.tenv", because "infer-out/captured/" will
soon be removed as it now only contains the global tenv (except in debug mode,
where it will still be created).

In the DB, we either store the local tenv for the file, or "global" to indicate
that the global tenv should be consulted.

This diff also moves `Cfg.store` to `SourceFiles.add` because that function
deals with more than just `Cfg.t`.

Reviewed By: jeremydubreil

Differential Revision: D6937945

fbshipit-source-id: 001c10a