[infer][java] avoid conflicts between Eradicate warnings and Biabdduction warnings
Summary:
This is to fix the conflicts between Eradicate and the Biabduction when reporting the same kind of errors: when Eradicate is on, the Eradicate warnings will have priority over the null deference reported by the biabduction.

If this approach proved to be successful in prod, I will refactor the reporting mechanism in the analysis itself to simply not report the null dereference in this case at all. For the codebases that aren't yet fully consistently using `Nullable`, this combined approach looks like a good way to deploy Infer toward full null safety.

Reviewed By: mbouaziz

Differential Revision: D7102119

fbshipit-source-id: 35d3add