[quandary] don't crash if JSON source/sink is invalid procedure name
Summary:
At the moment, Java and Clang sources/sinks live in the same inferconfig entry.
If we try to parse a Java procedure that happens to be an invalid Clang qualified name (e.g., `MyClass.<init>`),
parsing will crash.

As a temporary fix, throw an exception and catch it instead.
In the future, we can avoid this by requiring that JSON source/sink specifications to indicate the language.

Reviewed By: mbouaziz

Differential Revision: D7291880

fbshipit-source-id: f8f4502