[reactive] kill the changed bit of procedure attributes
Summary:
This is unused, but the logic for keeping its value up to date was still alive
and kicking.

I don't know that making use of this flag would be easy: we could just use it
in `Ondemand.should_analyze`, but it would be unsound because a procedure might
need to be re-analysed becauese its dependencies have changed. Since there's
not code to deal with that currently I think it's best to remove it and
re-introduce it when we have some idea how to use it.

Reviewed By: sblackshear, jeremydubreil

Differential Revision: D7444179

fbshipit-source-id: 99a1ec5