[clang] do not run clang on unsupported commands
Summary:
Try to skip running even `clang -###` when we know there's nothing to capture
in the clang command. Do that by recording in the `ClangCommand.t` whether the
command is pre- or post-running `clang -###`, and apply slightly different
filters in each case. In particular, the filter has to let in more commands
than necessary pre-`clang -###` because the flags are not normalised at that
point.

Also regard `-x cuda` commands as unsupported.

Reviewed By: mbouaziz

Differential Revision: D7584934

fbshipit-source-id: c27bb63