[make] fix `make devsetup`
Summary:
Making the shell stricter revealed a bug in `make devsetup`:
`shell_config_file` can be used without being initialised. Also, the error path
when testing if `BUILD_MODE` was set to default was wrong and would execute all
except the first `echo` command unconditionally. I changed the test so that it
tests only if `BUILD_MODE` is set in the environment.

Reviewed By: martinoluca

Differential Revision: D7615333

fbshipit-source-id: 08c5eb0