[infer] refactor LintIssues for general use.
Summary:
One source of non-determinism is racing on procedure summaries when reporting.  In particular, the summary of a method may be computed and stored by one thread, but another may be trying to report on it (eg, in cluster checkers).

One solution (at least until everything is in sqlite) is to have separate files just for the reports, a la linters.  This diff improves the interface of LintIssues and generalises it ahead of using it in other analysers.

Reviewed By: jeremydubreil

Differential Revision: D7859973

fbshipit-source-id: 8672d3b