[frontend] Translating ToVoid cast
Summary: We were missing reads of `a` if it was used in void cast, i.e. `(void) a;` This caused dead store false positives: we were not using `exp` that was the result of translating `a`. This diff creates a call to built-in skip function with `exp` as its argument, which causes the analyses to see reads of `exp`.

Reviewed By: mbouaziz

Differential Revision: D8332092

fbshipit-source-id: f3b0e10