[starvation] report only on public methods
Summary:
We need to report on non-private methods (the opposite even leads to FPs sometimes on deadlocks). To do this, the domain needs to change so that the interpretation of an order pair `a,b` is no longer "lock `a` is taken in the *current method* and held until lock `b` is taken".  Instead the meaning is now "lock `a` is taken in some method *of the same class with the current method* and is held until `b` is taken".

These changes are quite drastic because the previous implementation optimised extensively around the previous use case.

Reviewed By: mbouaziz

Differential Revision: D8395351

fbshipit-source-id: a2bd22b