[cli] create missing directory
Summary:
The captured/ directory is not guaranteed to exist when the capture was run without debug mode, but the analysis assumes it will be. So, this crashes infer:

```
infer -- clang -c examples/hello.c
infer -g
```

There's no need for `DB.Results_dir.init` to create infer-out and infer-out/specs/ since this is now done in `ResultsDir`.

Reviewed By: jeremydubreil

Differential Revision: D8732739

fbshipit-source-id: aaac902