[siof] do not warn on simply taking reference of global
Summary:
The addresses of  global variables do not need initialisation to exist and be valid as they are part of the code or data segment of the program. This means that taking the address of a global is not in itself a danger for SIOF. However, dereferencing such an address would be. In order to avoid false positives but avoid being too unsound, only ignore them when the address is taken only to set another global. The general case would require a more complicated abstract domain.

Fixes #866

Reviewed By: ngorogiannis

Differential Revision: D8055627

fbshipit-source-id: 92307b2