[explore] change filtering options to filter in OCaml rather than SQLite
Summary:
Filtering on the SQLite side was done to be more efficient, but these are debug
options so it should be fine for them to be not very optimised.

Filtering on the OCaml side will allow us to re-use these filtering options for
other purposes, such as re-analysing certain procedures only.

Reviewed By: mbouaziz

Differential Revision: D8767691

fbshipit-source-id: e232660