[toplevel] install infertop in infer/bin/ to get models and stuff
Summary:
Some paths are hardcoded in infer as being relative to the current executable,
for instance the directory where to find the models. By copying infertop.bc to
infer/bin like we do for `infer` these relative paths lead to the expected
place, which means models can be loaded in the toplevel like they would be in a
normal infer execution. This is more useful for debugging than previously.

Reviewed By: jeremydubreil, mbouaziz

Differential Revision: D9197142

fbshipit-source-id: 48c4f82fb