[Perf][CI] Differential of costs, based on degree variation
Summary:
First version of differential for costs, based on polynomial's degree's variation. The rule is very simple:
For a given polynomial that is available before and after a diff, `if degree_before > degree_after`, then the issue becomes `fixed`. Instead, `if degree_before < degree_after`, then the issue becomes `introduced`.

Reviewed By: ezgicicek

Differential Revision: D9810150

fbshipit-source-id: d08285926