[kill -a][4/4] kill `Config.analyzer`, deprecate `--analyzer`
Summary:
Keep `--analyzer` around for now for integrations that depend on it.

Also deprecate the `--infer-blacklist-path-regex`,
`--checkers-blacklist-path-regex`, etc. in favour of
`--report-blacklist-path-regex` which more accurately represents what these do
as of now.

Rely on the current subcommand instead of the analyzer where needed, as most of
the code already does.

Reviewed By: jeremydubreil

Differential Revision: D9942809

fbshipit-source-id: 9380e6036