[starvation] allow propagation of order constraints to any class
Summary:
An order constraint (A,B) means we take lock A and before releasing it we perform B (whatever that is).
Previously if a method call crossed class boundaries, we removed the callee's order constraints before integrating the callee's summary to that of the caller.  The reasoning was that this may lead to reports blaming a caller for something they are very far from, plus a proliferation of reports with the same bad endpoint.

The first reason still applies, but this is a general problem.  It may be better to report and let developers deal with it.
The second reason is moot, since in differential mode most of these reports are hidden.

Reviewed By: jberdine

Differential Revision: D10173200

fbshipit-source-id: 9afbf292c