[proc-cfg][3/5] stop caching whole-file cfgs in exe_env
Summary:
Load proc descs from the "procedures" sqlite table instead of from
file-wide cfgs stored in the "source_files" table. This removes the need
for a cache of these file-wide CFGs, which was needed because loading
them is expensive and potentially needed in case we need to load the
proc descs of several procedures in the same file. Now we can just load
the proc descs one by one and not worry about caching.

Reviewed By: jberdine

Differential Revision: D10173355

fbshipit-source-id: 665636121