[log] print only innermost source file context of crash
Summary:
When the backend crashes we print which instruction/file/... we were analysing,
but because of recursion we can end up repeating that information all
the way to the toplevel call.

This makes sure we only print the innermost one, we don't care about the
calling context because the analysis is compositional.

Reviewed By: mbouaziz

Differential Revision: D10381141

fbshipit-source-id: 1c92bb861