[pulse] forget about addresses that are invalid on only one side of a join
Summary:
Getting this right will be long and complex so for now the easiest is to
underreport and only consider as invalid the addresses we know to be invalid on
both sides of a join. In fact the condition for an address to be invalid after
a join is more complex than this: it is invalid only if *all* the addresses in
its equivalence class as discovered by the join are invalid.

Reviewed By: skcho

Differential Revision: D12823925

fbshipit-source-id: 2ca109356