[clang] ensure we get the correct file path
Summary:
As explained in the added comment, clang started adding `-faddrsig` at the end
of every `-cc1` command, which trumps our heuristic for finding the file name
(thus we would write debug scripts to `-faddrsig.ast.sh`, do filename-based
filtering on `-faddrsig` instead of the source path, and more...).  We rely on
the file name being the last argument in `-cc1` commands because so far that's
always been the case, and we don't want to parse the clang command line and
have to know about all the clang options...

Thanks martinoluca for the trick of simply passing `-fno-addrsig`!

Reviewed By: martinoluca

Differential Revision: D12921987

fbshipit-source-id: 28bebe647