[preanalysis] do not delay killing variables taken by reference
Summary:
Before, the liveness pre-analysis would place extra instructions in the
CFG for either:
1. marking an `Ident.t` as dead, or
2. marking a `Pvar.t` as `= 0`

But we have no way of marking pvars dead without setting them to 0. This
is bad because setting pvars to 0 is not possible everywhere they are
dead. Indeed, we only do it when we haven't seen their address being
taken anyway. This prevents the following situation, recorded in our tests:

```
int address_taken() {
  int** x;
  int* y;
  int i = 7;
  y = &i;
  x = &y;
  // if we don't reason about taken addresses while adding nullify instructions,
  // we'll add
  // `nullify(y)` here and report a false NPE on the next line
  return **x;
}
```

So we want to mark pvars as dead without nullifying them. This diff
extends the `Remove_temps` SIL instruction to accept pvars as well, and
so renames it to `ExitScope`.

Reviewed By: da319

Differential Revision: D13102953

fbshipit-source-id: aa7f03a52