[pulse] take array indices into account
Summary:
This will be useful to make the analysis more precise. In particular, it
allows a disjunctive version of pulse to deal will deleting vector
elements in a loop: without this, deleting an array element in one
iteration will make the analysis think that the next array element is
invalid too since they are all the same. By keep track of the index, we
can detect when we are sure that two elements are the same and only
report in that case.

Reviewed By: ngorogiannis

Differential Revision: D13431374

fbshipit-source-id: dae82deeb