[pulse] add traces to the domain
Summary:
Record per-location traces. Actually, that doesn't quite make sense as a
location can be accessed in many ways, so associate a trace to each
*edge* in the memory graph. For instance, when doing `x->f = *y`, we
want to take the history of the `<val of y> --*--> ..` edge, add "assigned
at location blah" to it and store this extended history to the edge
`<val of x> --f--> ..`.

Use this machinery to print nicer traces in `infer explore` and better
error messages too (include the last assignment, like biabduction
messages).

Reviewed By: da319

Differential Revision: D13518668

fbshipit-source-id: 0a62fb55f