[eradicate] rename the warning raised when dereferencing nullable types
Summary:
The Eradicate backend is reporting nullable type errors, that are not always necessarily leading to null pointer exceptions.

For example, the analysis is designed to be consistent with the Java type system and report on the following code:

   String foo(boolean test) {
     Object object = test ? new Object() : null;
     if (test) {
        return object.toString();  // the analysis reports here
     }
   }

even though the code will not crash.

In order to make this aspect clear, this diff renames the warnings `Null Method Call` and `Null Field Access` into `Nullable Dereference`

Reviewed By: ngorogiannis

Differential Revision: D14001979

fbshipit-source-id: ff1285283