[inferbo] Reset LatestPrune at the assignment of return variable
Summary:
This diff changes a LatestPrune to use a return variable instead of another local variable, when the function returns a conditional value.  This is a preparation to propagate LatestPrune inter-procedurally in the following diffs.

context: If a function returns a conditional value, e.g. `return x == y`, the LatestPrune value includes a temporary local variable introduced by the SIL translation.  This diff is to avoid propagating the temporary local variables to its caller.

Reviewed By: mbouaziz

Differential Revision: D14321534

fbshipit-source-id: d157bfdd0