[sledge] Add Solver tests demonstrating incompleteness
Summary:
The SL solver is currently not always able to append segments which
have been split symbolically, that is, at an internal point expressed
using a variable, rather than merely a constant.

Also, existential instantiation, that is, the choice of witnesses
during proof search, is currently sensitive to the order of
subformulas. This can lead to fragile incompleteness.

Reviewed By: mbouaziz

Differential Revision: D14481991

fbshipit-source-id: 80fe2f0a8