[sledge] Compare stack component of edges as a inlined code location
Summary:
When comparing stacks as part of a control-flow edge, treat each as a
code location in a hypothetical expansion of the program where all
non-recursive functions have been completely inlined. In particular,
this means to compare stacks as if all Locals frames or Return frames
for recursive calls had been removed. Additionally, the from_call info
in Return frames is ignored.

Reviewed By: jvillard

Differential Revision: D14657601

fbshipit-source-id: b8a42d3fa