[ocamlformat] Move .ocamlformat config file to the project root
Summary:
Several tools that use ocamlformat look for the "project root" in
order to find the config for a given file to format. The project root
is inferred by looking in ancestor directories of the input file until
one of .git, .hg, or dune-project is found. Since the .ocamlformat
config file is currently two directories higher than dune-project,
this fails. This diff moves the config file.

Reviewed By: jvillard

Differential Revision: D14694260

fbshipit-source-id: 2fb51bf30