[SIL] change `variable_initialization()` builtin to a new auxiliary instruction
Summary:
Instead of emitting an ad-hoc builtin on variable declaration emit a new
metadata instruction. This allows us to remove the code matching on that
ad-hoc builtin that had to be inserted in several checkers.

Inferbo & pulse used that information meaningfully and had to undergo
some minor changes to cope with the new metada instruction.

Reviewed By: ezgicicek

Differential Revision: D14833100

fbshipit-source-id: 9b3009d22