[pulse] tighten up summaries
Summary:
Only throw values to the pre if they can be followed from "abducible"
variables: formals of the current method and globals.

Because figuring out if a `Pvar.t` is a formal of the current procedure
is actually a giant pain, hack something not too bad instead:
pre-register all formals at the start of the analysis of the
procedure. Then the only other variables we care about in the
precondition are globals, which we can detect easily.

This is mostly an optimisation (summaries won't include irrelevant
"abduced" facts about the procedure's local variables anymore), but it
also fixes a bug where we would sometimes overwrite things in the pre. I
think that's why the tests improved.

Reviewed By: ngorogiannis

Differential Revision: D14753493

fbshipit-source-id: 08e73637f