[report] avoid embarrassing "object `null` could be null" message
Summary:
A long-standing easter egg from infer error messages is the "object
`null` could be null and is dereferenced at line ...". I tried to fix
this but the part that generates the first "null" in the message and the
part that generates the second one are very far apart and it's hard to
see how to make the second part aware of the first in a clean way.

Instead, hack around it by detecting if the string representing the
value is literally `null` and in that case chop `could be null ` from
the error messages...

Reviewed By: jeremydubreil

Differential Revision: D14972324

fbshipit-source-id: ccc48ce6b