[tasks] move to a task generator interface
Summary:
Before moving to any kind of non-trivial scheduling, we need to change the Tasks interface.
In particular, it's too restrictive to expect that the tasks to be scheduled are provided as a list before starting execution.  For example, dynamic scheduling does not fit the bill here.  Also, the list expectation means all scheduling work has to be done up front.

The solution here is to move to a `Sequence`-like interface with one difference:
- The function returning the next task expects a task option argument.  That argument is the task that was just finished (if any) by the worker expecting new work.  This will be useful for things like task dependencies (for instance, a procedure has been analysed, and can be marked so).

Reviewed By: mbouaziz

Differential Revision: D15181613

fbshipit-source-id: 21f3ba825