[sledge] Fix bound not bounding recursion
Summary:
Sledge does not terminate on programs with recursion, because
functions get "infinitely inlined" and therefore recursion is not
treated as retreating edge.

This patch bounds the number of times the same function can "inlined"
to respect the bound (`-b` option). On each call we check the number of
occurances of the called function in the call stack. If that is higher
than the bound, we skip it.

Reviewed By: jvillard

Differential Revision: D15577134

fbshipit-source-id: 4cd3b62c6