[sledge] Fix crashing frontned
Summary:
When LLVM is built with assertions, it crash
`add_sym` if you try to get the global scope of a non global value.

This patch special cases add_sym, to just do nothing when `llv` is
an `UndefValue`.

Also enhances debuging printout of transalte to include the number of
functions and globals.

Reviewed By: jvillard

Differential Revision: D15669447

fbshipit-source-id: 4b5483810