[buck/java2] hashcons the global tenv during merging
Summary:
One "interesting" feature of the approach of merging the captured targets in Java, is that we union their type environments, as opposed to store partial tenvs together with each source file, which is the case for Clang.

This means
- the final global type environment is potentially huge because it contains all the types in all targets.
- all analysis workers start by loading that tenv in memory, meaning we consume `|size of tenv| x #cpus` memory, which can tip the balance towards OOMs

This diff attempts to economise on global tenv size. This is done by increasing sharing which is then preserved by marshalling.  It's done in a brute force way, with hashtables for each struct component, and is not fully effective due to the recursion amongst types and types names, as well types appearing inside other constructs such as procnames.

This is done when calling `Tenv.store` so that
- the computation can be parallelised somewhat (capture is parallel, merging is not)
- buck caching will benefit from smaller tenvs.

This saves about 24% of total memory devoted to the type environment.

Reviewed By: mbouaziz

Differential Revision: D15840054

fbshipit-source-id: 6f03be1a4