[pulse] more general graph visitor API
Summary:
Make it possible to re-use the graph visitor to compute all sorts of
things with a flexible API where you can pass a function that folds over
all addresses reachable from certain stack variables (specified with a
filter) and gets passed the access path that leads to each address.

This is used in later commits.

Reviewed By: mbouaziz

Differential Revision: D15824960

fbshipit-source-id: c424a71cb