[pulse] move to SIL
Summary:
It turns out HIL gets in the way of a precise heap analysis. For
instance, instead of:

```
n$0 = *&x.f
_ = delete(&x)
*&y = n$0
```

HIL tries hard to forget about intermediate variables and shows instead

```
_ = delete(&x)
*&y = *&x.f
```

Oops, that's a use-after-delete, whereas the original code was safe.

While it's easy to write SIL programs that are completely unsound for
HIL, they are not generated very often from the frontends. In fact, the
problem became apparent only when making the clang frontend translate
C++ temporaries destructors, which produces the situation above
routinely.

This diff makes the minimal amount of change to make Pulse build and
produce equivalent results (minus HIL bugs) starting from SIL instead of
HIL. The reporting sucks for now because we need to translate SIL
temporaries back into program access paths. This is done in the next
diff.

Reviewed By: mbouaziz

Differential Revision: D15824961

fbshipit-source-id: 8e4e2a3ed