[pulse] skip `folly::SocketAddress::~SocketAddress`
Summary:
The constructor of `folly::SocketAddress` conditionally deletes some
object and then makes that condition false. The destructor then does the
same. Pulse ignores conditionals so will see a double delete.

Just skip that function for now, but it should be easy for pulse to be
more correct here if it knew how to compare constant values.

Reviewed By: mbouaziz

Differential Revision: D16005395

fbshipit-source-id: 036f5091b