[sledge] Include alarms in debug trace
Summary:
Currently alarms are reported to stdout while the debug trace is
written to stderr. This makes synchronizing the two difficult. With
this diff, the alarm reports can also be included in the debug trace,
and analysis can be stopped when an alarm is encountered by tracing
the `Stop` module, e.g.:
```
sledge -trace Report+Stop.on_invalid_access
```

Reviewed By: kren1

Differential Revision: D16072611

fbshipit-source-id: 32c3639a2