[pulse] support modelling destructors
Summary:
We want to detect that variables and C++ temporaries go out of scope
even when their destructor happens to be modelled.

We lost a test to that because `std::function::~function` was poorly
modeled as deleting the lambda itself which would now cause a double
invalidation. This has to be modelled better now as something that
invalidates something *inside* the lambda, and also model `operator()`
as something that accesses that something, to recover that test. It's
not a vital test though, so Do It Later©.

Reviewed By: ngorogiannis

Differential Revision: D16121091

fbshipit-source-id: 6b777ca18