[sledge][summaries] Fix unsoundes due to missing frame
Summary:
When using summaries we first garbage collect the precondition and then
ask the solver to infer the frame of the precondition with respect to
grabage-collected footprint.

Currently if the solver fails to show the frame, we just give it an
empty frame. This is bad, because if grabage collection removed some
segments, they don't get added back on.

This patch throws an exception instead to be very explicit when the
solver cannot show the frame in this case.

Reviewed By: ngorogiannis

Differential Revision: D16339587

fbshipit-source-id: b88d0689c