[ondemand] refactor to avoid loading unneeded proc descs
Summary:
There was a little bit of code duplication around `analyze_proc` to deal with
the fact that we may be starting from either a proc name or a proc desc. Create
a new `callee` type that represents this more explicitly. This allows not
loading the `proc_desc` eagerly when we don't need it, although that doesn't
seem to impact perf measurably.

Reviewed By: ezgicicek

Differential Revision: D16442221

fbshipit-source-id: 8e8ebbd6b