[diff] delete infer-diff
Summary:
It's not being worked on and is not in a state where it works.

It would probably better to write this as a script of some kind or else
resurrect this subcommand in a form where it behaves more like a script,
ie fork/execs infer analyses instead of having them be function calls
(but then it might as well *be* a script as it would likely be more
flexible).

In any case...

youarealreadydead

Reviewed By: ezgicicek

Differential Revision: D16602417

fbshipit-source-id: d0d129539