Fail if incremental analysis is set without changed files
Summary:
It doesn't make sense to use incremental analysis without specifying changed files. This is a possible source of future bugs.

This commit causes infer to die if incremental analysis is used without changed files.

 ---

Previously:

I think this code is currently a bit brittle because the CI shadow builds sometimes use `--incremental-analysis` because they are called with the same command as the diff analysis.

I am worried that in the future the shadow builds could be broken by this, although everything looks like it works right now. This diff would prevent breaking smoke builds in future because the shadow builds don't set changed-files (afaik).

However, possibly this is not the right place to fix the problem. It might be better to change the CI, I'm not sure.

Reviewed By: mityal

Differential Revision: D16829192

fbshipit-source-id: 5ee1ce9d0