[inferbo] Fix bug in integer pruning by pointer
Summary:
This diff avoids that an integer value is pruned to the bottom by
comparing to a pointer.

For example, before this diff,

assume((int*)x == p);
assume((int*)x != p);

where x is an integer, x is pruned to the bottom in both of the assume
cases.  So, there were some, unintentional and false, unreachable
code.

Depends on D16960199

Reviewed By: ezgicicek

Differential Revision: D16964735

fbshipit-source-id: 90a3c8c80