Definie a prelim. AST for llair's semantics
Summary:
The AST is not complete on expressions, but it should have most of the
important features.

The representation is in some ways very different from the OCaml
implementation, because the OCaml code uses mutation to build the CFG as
an actual pointer graph in memory, and also because the expression
representation is optimised for the backend. For the former, it should
be easy to see that the AST here is isomorphic, representing the CFG
with finite maps from block labels. The correspondence is less clear in
the latter case, but the point here is not to model or verify
implementation optimisations, but to give a semantics to llair as a
language.

Reviewed By: jberdine

Differential Revision: D16807132

fbshipit-source-id: b0f64b3ec