[nullsafe] Refine semantics of @Nonnull: obey minimum surprise principle
Summary:
`nullsafe` currently allows the following:

```
public void Nonnull Object willBeOK() { return null; }
```

But disallows the following:
```
public void Object willBeAnIssue() { return null; }
```

This was a deliberate choice made back in 2014.
The motivation was to provide a way to tell the checker "I know it can not be null, trust me".

A huge problem with that approach is that it is extremely non-intuitive and surprising, and contradicts with pretty much everything when Nonnull or similar annotations are used in external world.
This is not the way how checkers should be supressed.

We do provide 2 options to express this intention, namely `assertNotNull` and `assumeNotNull` would do the thing.
This is a much better approach for additional reason: assertNotNull is
granular and applies only to the exact expression that is under
question. In contrast, suppressing the check on the whole function level
make any modifications of a function dangerous.

Reviewed By: artempyanykh

Differential Revision: D16984213

fbshipit-source-id: 0ba0f623b