[pulse] do not record addresses of logical variables
Summary:
Before this diff we would record when some values came from the "address
of" logical variables. This makes no sense and also was incorrectly
marking these addresses as "written to" when they appeared in the post
of a procedure, because their attributes weren't empty (they had the
"address of stack variable" attribute).

Reviewed By: ngorogiannis

Differential Revision: D17131210

fbshipit-source-id: 6cc3c465a