[sqlite] move all writes to one module
Summary:
Write contention is becoming a problem in parallel capture (eg when make runs with high parallelism) or when analysis writes CFGs to the DB in parallel (eg when analysing blocks in ObC).  This is believed to lead to BUSY errors in Sqlite.

This is step 1 of a process where all writes are cordoned-off in one module, and fixing the interface for that module.

Reviewed By: skcho

Differential Revision: D16985034

fbshipit-source-id: 3d7ce381b