Add method to clean the results directory
Summary:
I observed a bug in incremental analysis for thread safety analysis, where a thread safety violation was not being reported because the folder `racerd` was not being cleaned. This meant that the violation was determined to be a preexisting issue when it was actually an introduced issue.

This method can be used to fix this problem by cleaning the `racerd` folder. It also cleans the `captured` folder, I've done this following the original version of the method (see D16602417).

I'm not sure if the `captured` folder is used; it wasn't used in the tests I did. Thoughts about this?

Reviewed By: ngorogiannis

Differential Revision: D17261504

fbshipit-source-id: 8fea23e98