[nullsafe] Remove custom hash calculation for err_instance
Summary:
At some point, there was a custom equality function that deliberately ignored some fields in err_instance. It was deleted in D4232422, so having a custom hash function does not serve any purpose anymore.

Since 2016 there was no known problems with the change in D4232422.

If we decide that we need similar behavior that was before D4232422, it will be easier to reimplement the functionality again.

Reviewed By: jberdine

Differential Revision: D17313660

fbshipit-source-id: 5c6c29a0b