[nullsafe] Remove special treatment of @Nonnull for "Condition Redundant" check
Summary:
CONDITION_REDUNDANT_NONNULL was an attempt to reduce number of false
positives for condition redundant. (It is the most popular check as of
now).

The root case for most of false positives is that a lot of code is
simply not annotated (but should have been), so blaming developers for defense programming is
not actionable.

In attempt to solve the problem, a special issue type (for case when the
code is explicitly annotated with Nonnull) was introduced.

In follow up diffs we are going to introduce a generic way of doing the
same, not limited to this particular check only.

Namely, we will introduce notion of unknown nullability, so it will be
possible to distinguish not annotated yet (hence no warnings) and already
annotated (hence warnings) parts of code.

This piece of logic is incompatible with the aforementioned work, hence
we need to remove it.

Reviewed By: jvillard

Differential Revision: D17398768

fbshipit-source-id: 8bddf10e5