[nullsafe] Remove old traces of modelling the annotation for the field
Summary:
We historically had Model.Inference, which was an attempt to enhance
models with additional abilities to get the annotation.

This feature got removed in D9805110, including removing of the key
condition Models.Inference.field_is_marked.

This code also is not executed: `Config.eradicate` condition
was an old artefact of migrating Eradicate to callback infrastructure:
D1508451. We run eradicate only as a callback as of now, so this flag is
always true.

In follow up diffs we refactor AnnotationSignature module, and this
cleanup simplifies the refactoring.

Reviewed By: ngorogiannis

Differential Revision: D17419173

fbshipit-source-id: 1b30555de