[nullsafe] When inferring type based on the formal type, respect NullsafeType instead of reading annotations
Summary:
This continues work for eliminating Annot.Item.t from Nullsafe low-level
code.

The introduced function `from_nullsafe_type` is called when we infer
initial type of the equation based on the function or field formal signature.

Before that, we did it via reading the annotation directly, which
complicates the logic and making introducing Unknown nullability tricky.

## Clarifying the semantics of PropagatesNullable

This diff also clarifies (and changes) the behavior of PropagatesNullable params.

Previously, if the return value of a function that has PropagatesNullable params was
annotated as Nullable, nullsafe was effectively ignoring PropagatesNullable effect.

This is especially bad because one can add Nullable annotation based on the logic "if the function can return `null`, it should be annotated with Nullable`.

In the new design, there is no possibility for such a misuse: the code that
applies the rule "any param is PropagatesNullable hence the return
value is nullable even if not explicitly annotated" lives in NullsafeType.ml, so
this will be automatically taken into account.

Meaning that now we implicitly deduce Nullable annotation for the return value, and providing it explicitly as an alternative that does not change the effect.

In the future, we might consider annotating the return value with `Nullable` explicit.

Reviewed By: jvillard

Differential Revision: D17479157

fbshipit-source-id: 66c2c8777