[nullsafe] Look at NullsafeType in check_return_annotation instead of looking it up in Annot.Item.t
Summary:
This proceed the work of getting rid of Annot.Item.t.
This diff:
- Moves "check assignment rule" to recently supported NullsafeRules
- Implements their own "check overannotated" (defers consolidating this
check into NullsafeRule for the future diffs).

Note that we don't need PropagatesNullable logic anymore because it is
already ported to NullsafeType (return value will be marked as Nullable
in NullsafeType)!

implicit_nullable (a.k.a. Void types) will require a follow up diff to
model.

Reviewed By: artempyanykh

Differential Revision: D17499246

fbshipit-source-id: 14b473f29