[clang] Do not try to store block declarations as methods of a C++ struct.
Summary:
This was causing a crash, because when trying to create a procname from a block at that point we don't have the block return type, which is needed for the name. I don't understand why BlockDecl doesn't contain the type, but I looked again and it doesn't (also in clang). So in general we need to pass it from the context, but that's not possible in this case.
Also, one could argue that such a block is not a method from the struct, since it's just a block that is assigned to a field as initialization.

Reviewed By: skcho

Differential Revision: D17575197

fbshipit-source-id: 3974ead3f