[cost] Add eval mode for cost substitution
Summary:
This diff adds an eval mode for the substitutions of the cost results, in order to avoid precision
loss by joining two symbols.

The usual join of two different symbolic values, `s1` and `s2`, becomes top due to the limitation of
our domain.  On the other hand, in the new eval mode, it returns an upperbound `s1+s2`, because the
cost values only care about the upperbounds.

Reviewed By: ezgicicek

Differential Revision: D17573400

fbshipit-source-id: 2c84743d5