Upgrading opam dependencies with last versions of Sawja and Javalib
Summary:
Sawja and Javalib have recently released new versions that drop off
the camlp4 dependency. This is a minimal diff in order to update infer
opam depedencies.

Last (1.5.7) generates invokedynamic, but work on InvokeDynamic is still
in progress in Infer and not activated here yet. In this version, the
Java frontend  will still replace any InvokeDynamic by a dummy
InvokeStatic call (as introduced by Jeremy a long time ago).

Reviewed By: jvillard

Differential Revision: D17662979

fbshipit-source-id: f686ba442