[starvation] remove rarely-used logic for identifying locks
Summary:
Ideally the analyser should equate locks `this.x.f` and `a.x.f` in different methods if they can alias.
The heuristic removed here was rarely used and is in the way of a re-write of the analysis.
It was also badly implemented, as this should ideally be the comparison relation of `Lock`.

Reviewed By: mityal

Differential Revision: D17602827

fbshipit-source-id: 4f4576c1a