[starvation] change domain to use critical pairs `(set of locks held, event)`
Summary:
The old domain keeps two sets:
- `events` are things (including lock acquisitions) which eventually happen during the execution of a procedure.
- `order` are pairs of `(lock, event)` such that there is a trace through the procedure which at some point acquires `lock` and before releasing it performs `event`.

A deadlock would be reported if for two procedures, `(lock1,lock2)` is in `order` of procedure 1 and `(lock2,lock1)` is in `order` of procedure 2.  This condition/domain allowed for the false positive fixed in the tests, as well as was unwieldy, because it required translating between the two sets.

The new domain has only one set of "critical pairs" `(locks, event)` such that there is a trace where `event` occurs, and *right before it occurs* the locks held are exactly `locks` (no over/under approximation).  This allows keeping all information in one set, simplifies the procedure call handling and eliminates the known false positive.

Reviewed By: mityal

Differential Revision: D17686944

fbshipit-source-id: 3c68bb957