[infer] Eagerly try narrowing for the outermost loops
Summary:
This diff tries to narrowing the fixpoint of outermost loops, so that over-approximated widened values do not flow to the following code.

Problem: There are two phases for finding a fixpoint, widening and narrowing.  First, it finds a fixpoint with widening, in function level.  After that, it finds a fixpoint with narrowing. A problem is that sometimes an overly-approximated, imprecise, values by widening are flowed to the following loops.  They are hard to narrow in the narrowing phase because there is a cycle preventing it.

To mitigate the problem, it tries to do narrowing, in loop level, right after it found a fixpoint of a loop.  Thus, it narrows before the widened values are flowed to the following loops.  In order to guarantee the termination of the analysis, this eager narrowing is applied only to the outermost loops.

Reviewed By: ezgicicek

Differential Revision: D17740265

fbshipit-source-id: e2d454036