[required-props] Improve the heuristic to check build() for summaries
Summary: Before, we didn't track litho framework callees on client code which was wrong. Now, we replace this with the following:  If the callee is `build()` itself or doesn't contain a `build()` in its summary, then we want to track it in the domain. The former makes sense since we always want to track `build()` methods. The latter also makes sense since such a method could be a setter for a prop (as in the case of `prop1` in `buildPropLithoOK` which we were missing before due to the imprecise heuristic that prevented picking up callees in litho).

Reviewed By: ngorogiannis

Differential Revision: D17810704

fbshipit-source-id: 87d88e921