[starvation] change UI-thread trace element from string to a proper type
Summary:
Starvation analysis keeps a trace documenting why a method is seen as on the UI thread (many reasons possible, often confusing).  This was a call-stack plus string, for keeping the explanation of why the last callee is on the UI thread.  This is bad, because it takes too much memory/storage (each string is custom-made to the classes/method involved), and is effectively untyped.

Switch to a proper type for explaining this, so the cost is just a few pointers plus shared procnames/types, and then convert to string only when reporting.  This will also allow to push the UI trace inside each element of the starvation domain, so as to allow path sensitivity etc, without blowing up summary size.

Reviewed By: ezgicicek, artempyanykh

Differential Revision: D17810007

fbshipit-source-id: cdd743975