[sledge] Introduce the notion of types having the same semantics
Summary:
Typ.equivalent is currently defined the same as Typ.castable, but
conceptually they are different and castable needs to be
weakened. They are different since for example it is possible to cast
from an i64 to a f64, but those types denote different sets of values
in the semantics, and the bitcast is modeled using a conversion
function.

Reviewed By: bennostein

Differential Revision: D17725615

fbshipit-source-id: 973574f2a