[pulse] stop the analysis when precondition cannot be applied for reasons others than errors
Summary:
If a precondition cannot be applied, it means that this program path
somehow doesn't make sense for the caller and so should be pruned. Right
now we just treat this as skipping over the call instead.

This will become more important when specs start mentioning arithmetic
facts that must be satisfied at the call site. As it is we will only
stop if we discover aliasing in the pre not present at the call site or
vice versa.

Reviewed By: dulmarod

Differential Revision: D18115230

fbshipit-source-id: 4f1c7a583