[cost] Set the minimum basic cost of node
Summary:
The zero cost of node does not make sense especially when the abstract memory is non-bottom.  This
resulted in unreasonable zero cost results sometimes, e.g. when the checker could not find
appropriate control varaibles having interval values of iteration.  This diff fixes this, so sets
the minimum basic cost as 1, if the abstract memory at the node is non-bottom.

Reviewed By: ezgicicek

Differential Revision: D18199291

fbshipit-source-id: b215d10e5