[nullsafe] Respect external models when decide wether to report on passing wrong param or not
Summary:
This is a dirty hack to make nullsafe behave consistently for internal
and external models.

Medium term this code needs to be rewritten in a better way, so that we
pass all information in annotated signature (either in form of Inferred
nullability, or type origin, which tracks third party calls).

As a follow up, we will make reporting more clear.

Reviewed By: artempyanykh

Differential Revision: D18372660

fbshipit-source-id: 12c2449e1