[inferbo] Function call of Java enum values in class initializer
Summary:
This diff adds semantics of Java function calls of enum `values` inside class initializers.

* Java class initializer function initializes a specific field `$VALUES`, which points to the list
  of enum values.
* The `values` function of enum class returns the value of `$VALUES`.

The problem is when the `values` function is called inside the class initializer, for example:

```
  enum Color {
    RED,
    GREEN,
    BLUE;

    static {
      for (Color c : Color.values()) {}
    }
  }
```

This introduces a recursive dependency: the class initializer calls `Color.values` and the function
returns `Color.$VALUES` the value of which should be initialized in the class initializer.

To address the problem, this diff finds the value of `$VALUES` in its abstract memory when
`values` is called inside the class initializer.

Reviewed By: ezgicicek

Differential Revision: D18349281

fbshipit-source-id: 21766c20f