[inferbo] Add size alias when i=1
Summary: Sometimes there is a code like `for(int i = 1; i < x; i++){ l.add(); }`, where the first element in a list is addressed specifically. This case was not analyzed precisely, because the alias value is added only when `i` is initialized by 0 by heuristic. This diff extends the heuristic, so it adds a size alias between `i` and `l.size()` when `i` is initialized by 0 or 1.

Reviewed By: ezgicicek

Differential Revision: D18351867

fbshipit-source-id: e7d19a4ec