[infer] Narrow all loops at least once in the widening phase
Summary:
This diff tries more narrowing during analysis in order to get preciser results on nested loops.

In the widening phase, it does narrowing a loop right after its widening, for each loops. In general, this may make the widening phase non-terminating because it keeps the abstract state from monotonely increasing to the fixed point in a finite number of iterations. To avoid that situation, this diff applies the narrowing only when the first visit of the loop in the widening phase.

Reviewed By: ngorogiannis

Differential Revision: D18400631

fbshipit-source-id: cc76f7e85