[litho][required-props] Enable tracking all create and build methods intra-procedurally
Summary:
We were not tracking static `create ` and `build` methods intra-procedurally in the analysis and we just looked up their summaries. This resulted in not having static `create` methods in the trace. Let's fix that by pushing static check into `satisfy_heuristic`.

This results in a few FPs, but such cases are not very common in the codebase.

Reviewed By: skcho

Differential Revision: D18451955

fbshipit-source-id: d4ccf46f5