[nullsafe] Split TypeOrigin.Formal into MethodParam and This
Summary:
This diff is a part of work teaching Nullsafe to explain decisions it's
making.

1. `Formal` was bit cryptic name.
2. Splitting method param and this makes a lot of sense. It is almost an
implementation detail that hey happen to come from "param" in the
method's signature.
3. Apart from others, this diff fixes a minor bug - we used to treat
this as DeclaredNonnull, which (in future) means suppressing legit warnings
like condition redundant. This would be an issue if we were to start
showing "high confidence" condition redundant warnings.

Reviewed By: ngorogiannis

Differential Revision: D18451294

fbshipit-source-id: acc295e3f