[nullsafe][TypeOrigin refactor] Rename 0None -> OptimisticFallback
Summary:
This diff is a part of work teaching Nullsafe to explain decisions it's
making.

In this diff, we merely rename the value to clearly reflect its currect
usage.

In follow up diffs we are going to make usages of this instance more
restricted and concentrate them in several places in the code.

Reviewed By: artempyanykh

Differential Revision: D18451290

fbshipit-source-id: cf3773364