[pulse] report dereference of NULL and constants
Summary:
Note: Disabled by default.

Having some support for values, we can report when a null or constant
value is being dereferenced. The particularity here is that we don't
report when 0 is a possible value for the address, or even if we know
that the value of the address can only be 0 in that branch! Instead, we
allow ourselves to report only when we the address has been *set* to
NULL (or any constant).

This is in line with how pulse deals with other issues: only report when
1. we see an address become invalid, and
2. we see the same address be used later on

Reviewed By: skcho

Differential Revision: D17665468

fbshipit-source-id: f1ccf94cf