[pdesc] new pre-analysis to diverge after "noreturn" function calls
Summary:
A plugin update allows infer to know when a function doesn't return
according to its attributes. This propagates this info all the way to
the attributes of each function, and then use this information in a new
pre-analysis that cuts the links to successor nodes of each `Call`
instruction to a function that does not return.

NOTE: The "no_return" `CallFlag.t` was dead code, following diffs deal
with that (by removing it).

Reviewed By: dulmarod

Differential Revision: D18573922

fbshipit-source-id: 85ec64eca