[ocamlformat] Add link to .ocamlformat from c_stubs
Summary:
The infer/src/c_stubs directory contains a dune-project file. This
causes ocamlformat to conclude that the root of the project is
infer/src/c_stubs, and there is no .ocamlformat file there, so it does
not apply to the Fnv64Hash.ml file.

(There is currently a bug where it does apply anyway, but it is fixed
in master.)

Reviewed By: dulmarod

Differential Revision: D18684831

fbshipit-source-id: bcb53ce94