[nullsafe][bug fix] Introduce a param guiding optimistic treatment of third party params and fix a bug in the current behavior.
Summary:
We were wrongly not supressing third-party calls in
case third party repo is provided: there was a bug in the function is_modelled_externally.
because of this we accidentally stopped supressing third party and
started to advertise it in non-strict mode.

we do plan doing that, but little bit further ahead :)

So let's add a param guiding this.

Reviewed By: artempyanykh

Differential Revision: D18658271

fbshipit-source-id: 703f2675b