[cost] Fix the model of substring
Summary:
This diff fixes the model of substring.

Problem: The cost model of the substring function was to return `size of string - start index` as a
cost.  However, sometimes this was a negative number, because of state abstractions on paths, array
elements, call contexts, etc, which caused an exception inadvertently.

This diff changes the model to return just `size of string`, when it cannot say `size of string` is
bigger than `start index`.

Reviewed By: ezgicicek

Differential Revision: D18707954

fbshipit-source-id: 63f27e461