[sledge][NFC] Refactor to avoid an unused open warning
Summary:
OCaml 4.08 has a new warning (66) on unused `open!` statements. This
has a suboptimal interaction with `ppx_let`'s `let%map_open` which
leads to triggering the warning if any of a group of such let bindings
does not need the open.

In this case, the refactor is easy.

But, warning 66 is very dubious, so also just switch it off.

Reviewed By: jvillard

Differential Revision: D18708466

fbshipit-source-id: 77618ab6e