[nullsafe] Specify "InferredNonnull" type origin even if it was already Nonnull
Summary:
This will make our analysis more precise. E.g. in the future  we can classify
"condition redundand" warnings by origin, and warn only for
InferredNonnull, or autogenerate fixes based on the origin.

Error messaging does not currently take into account origin for nonnull
types, but this can be changed in the future

Reviewed By: ngorogiannis

Differential Revision: D18685304

fbshipit-source-id: 6a1f263f5