[uninit] do not spam spurious error
Summary:
This can happen for a number of reasons that are not errors (mostly the exit node being unreachable) so isn't actionable, and definitely not worth showing the user. It could be a debug message but I don't think that's even worth it. Other checkers don't warn in similar circumstances.

With OCaml 4.08 we started actually seeing these error messages (differences in flushing behaviour?) so their annoying nature was revealed.

Reviewed By: ngorogiannis

Differential Revision: D18808460

fbshipit-source-id: a47a1dcb4