[pulse] fix test for by-ref automatic initialisation
Summary:
Pointers are hard... The previous test had no chance of doing
initialisation of the pointer by reference and was in fact a false
negative (and still is, fix incoming). Renamed functions to stress the
false negative and added a test that is really (potentially) doing
pointer initialisation by reference.

Reviewed By: skcho

Differential Revision: D18888008

fbshipit-source-id: 1e72408c7