[litho] Do not remove non-build-called part at build calls
Summary:
In the previous code, it removed non-build-called method calls.  For example, it was like

```
{non-build-called: {prop1, prop3}}
{build-called: {}}

b.build();

{non-build-called: {}}
{build-called: {prop1, prop3}}
```

However, this behavior introduced a false positive when there is multiple builders that point to the
same abstract object and `build` is called one by one.

This diff changes the semantics to keep the method calls of non-build-called at `build` calls.

Reviewed By: ezgicicek

Differential Revision: D19144525

fbshipit-source-id: e2ace127f