[sledge] Close Equality.classes under one-step congruence
Summary:
Equality.classes was assuming a simpler representation, and was
incomplete as a result.

The 'representative' map is not kept in a normalized form, where
subterms are necessarily representatives. Therefore, applying the
representative map to subterms of terms in a class can reveal new
elements of the class. This mirrors how the `lookup` function in
`normalize` works.

Reviewed By: ngorogiannis

Differential Revision: D19221868

fbshipit-source-id: 4a2ed6d3f