[SelfInBlock] Reporting strongSelf Not Checked only in field access
Summary: After receiving feedback about this, I'm changing the reporting of strongSelf Not Checked to only in cases where it can cause a crash. Here I'm adding reporting for field access, and removing general reporting. In a next diff, I'll also add reporting for passing strongSelf to methods in not explicitly nullable positions.

Reviewed By: skcho

Differential Revision: D19329842

fbshipit-source-id: 35beb2aa3