[clang frontend] Translate BinaryOperators in ifs equally whether they are part of a ExprWithCleanups or not
Summary: I noticed when looking into a false positive of strongSelf Not Checked, that there were some inconsistencies in the translation of if statements with an and, with an extra redundant join only if using a method in the condition that returned an object. So I could repro the problem and investigate and found the place of the inconsistency in the translation. This diff fixes it without changing things too much.

Reviewed By: jvillard

Differential Revision: D19518368

fbshipit-source-id: 47a6a778c