[build] Add `make check` target to build artifacts for tooling/merlin
Summary:
Quoting the doc:
> `dune check` builds the minimal set of targets required for tooling
support. Essentially, this is .cmi, .cmt, .cmti, and .merlin files.

So, running `make check` can be a tad faster and more economical
way to typecheck the code as you develop (and make merlin understand
what's going on).

We already had `check` taget defined in `infer/src/Makefile` but it
was not exposed at the top level. Fixing it.

Reviewed By: jberdine

Differential Revision: D19619920

fbshipit-source-id: c143cecd8