[racerd] use concurrent context for reporting in C++/ObjC
Summary:
To emulate the `ThreadSafe` contract in C++/ObjC, reporting was gated behind a check that ensured a C++/ObjC class has a `std::mutex` member (plus other filters).  This is reasonable, but it has some drawbacks
- other locks may be used, and therefore must be added to the member check;
- locking mechanisms that use the object itself as a monitor cannot be modelled (`synchronized` in ObjC)

RacerD already has `ThreadsDomain` which models our guess on whether a method is expected to run in a concurrent context, and which in C++/ObjC boils down to whether the method non-transitively acquires a lock.  This should be a good enough indicator that the class should be checked regardless of whether the locks are member fields.  This diff gates the C++/ObjC check on that abstract property.

Reviewed By: dulmarod

Differential Revision: D19558355

fbshipit-source-id: 229d7ff82