[nullsafe][easy] Break CustomAnnotations.java into two subtests
Summary:
This test tests PropagatesNullable and TrueOnNull/FalseOnNull
annotations.

Both tests suites grew big so it is hard to observe them at glance and
make changes.

I could not figure out better name for TrueFalseOnNull.java, it is sort
of silly but I optimized for searchability, "FalseOnNull" will be
directly searched and "TrueOnNull" will be searched in IDEs that are
smart enough.

Reviewed By: skcho

Differential Revision: D19724512

fbshipit-source-id: 703961342