[starvation] use access expressions instead of access paths
Summary:
The goals are:
- Increase precision in C-languages by ditching access paths.
- Help with eventually sharing the abstract address module with RacerD.
- Reports are now language-mode specific (eg `->` in clang vs `.` in Java).

It's not exactly access expressions used here.  Instead the pattern `(base, access list)` is used where `access` is `HilExp.Access.t`.  This is done to ease the way `deriving` is used for creating two comparison functions, one that cares about the root variable and one that doesn't; and also because the main function that recurses over accesses (`normalise_access_list`) visits the accesses from innermost to outermost.

Also, kill some dead code.

Reviewed By: skcho

Differential Revision: D19741545

fbshipit-source-id: 013bf1a89