[nullsafe][refactoring] Simplify call to condition_redundant checks.
Summary:
This refactoring unblocks the changes in follow up diffs (plus fixes a
bug).

So what was happening?

Each comparison with null leads to CFG being splitted into two branches, one branch
is PRUNE(a == null) and another is PRUNE(a != null).

PRUNE(a != null) is where most of logic happens, it is the place where
we infer non-null nullability for a, and this is a natural place to
leave a check for redundancy.

Before this diff we effectively checked the same thing twice, and used
`true_branch` (only one of 2 instruction will have it set to true) as a symmetry breaker.

This diff removes the `true_branch` checks, but leaves only one call out
of two, hence breaking symmetry in a different way.

## Bug fix
The code around the removed check was (crazily) doing two things at
once: it processed results of (returning booleans!)
TrueOnNull-annotated functions AND
results of (returning Objects!) other functions, using the fact that all
of them are encoded as zero literals (sic!).

Not surprisingly that lead to a bug where we accidentally call the check
for non intended places (arguments of trueOnNull functions), which lead
to really weird FP.

This diff fixes it.

Reviewed By: dulmarod

Differential Revision: D19744604

fbshipit-source-id: fe4e65a8f