[nullsafe][refactor] Simplify logic for the method handling PRUNE instruction
Summary:
This diff proceeds clean up the mess in TypeState.ml

This refactoring unblocks the change in the logic for true on null and
allows to fix a bug, see follow up diffs.

The current code is trying to do two things at once: processing boolean
results (which normally need manipulations with the argument of a
function + additional logic for containsKey), and comparisons witn null
(which requires manipulation with return value in typestate plus check
for redundancy).

All this was done in sort of generic fashion, which, among other, lead
to weird situations for edge cases, e.g. in processing containsKey we used to override nullability
already inferred as a non-null twice with losing information about
original nullability; which made logs weird).

Reviewed By: ngorogiannis

Differential Revision: D19768795

fbshipit-source-id: c928e2cff