[nullsafe] Model Object.equals() and descentands as FalseOnNull
Summary:
Since we fixed a bug in implementation of FalseOnNull (see stack below),
we can finally ship this change.

Side note: this change is essential for the follow up diff (which adds extra check
for user-defined implementations of equal()), without it the follow
up change would introduce a lot of false positives.

Reviewed By: ngorogiannis

Differential Revision: D19771057

fbshipit-source-id: 7d7cf1ef7