[nullsafe] Treat parameter in overrides of `equals(Object param)` as nullable.
Summary:
We already warn about lack of nullable annotations in `equals()`, and even have a specialized error message for that.

But lack of an annotation is not as severe as direct dereference: the
latter is a plain bug which is also a time bomb: it will lead to an NPE not immediately.

This is widespread enough to be reported separately.

Reviewed By: dulmarod

Differential Revision: D19719598

fbshipit-source-id: a535d43ea