[racerd] improve handling of summaries of synchronized methods
Summary:
The domain has a notion of lock state (taken/not taken) and any access occurring will remember the current lock state.

Methods in Java can be `synchronized` meaning the lock is taken automatically at method start and release on return.

Currently, instead of starting analysis of a `synchronized` method with an initial lock state of "lock taken", the summary would be computed as if there is no lock, and then a caller would peek at whether the callee is `synchronized` and change the callee summary accordingly before taking it into account. Also, when an access happens (not a call) the analysis always consults the pdesc of the current method to check whether the method is `synchronized`.

Do things the right way: if method is `synchronized`, start with the lock taken.  When the method exits, the lock-state postcondition is computed by releasing once the lock.

This is a behaviour-preserving change.

Reviewed By: mityal

Differential Revision: D19742559

fbshipit-source-id: 1d0fce3f6