[ocaml] aliasing sexp_{list,option} doesn't work anymore
Summary:
That trick briefly worked when it was introduced but I think a change in
merlin broke it again as I see "sexp_list", "sexp_option" in the types
shown by merlin again. Disgusting!

Remove useless hack. Also referring to types such as `sexp_list` is
deprecated in core v13 for some reason.

Reviewed By: jberdine

Differential Revision: D19861586

fbshipit-source-id: 1c4c3af13