[Infer] Dedup reports of strongSelf Not checked
Summary: When we discovered that a strongSelf var was not checked for null, we then report in each occurrence which is spammy. Now we report only the first occurrence. To achieve that, we store a `reported` flag in the domain that gets set to true after we report once, and we only report if it's false.

Reviewed By: jvillard

Differential Revision: D19877218

fbshipit-source-id: c44109ae9