[nullsafe] Improve documentation for @Initializer annotation
Summary:
The previos one was too broad and did not indicate the main intended
usage, which is currently the nullsafe typechecker.
Also it was misleading: Initializer methods should NOT be called inside
constructors.
Finally, it recommended using Initializer in Builder pattern, which is
a questionable idea, so it better to avoid mentioning builders.

Reviewed By: artempyanykh

Differential Revision: D19942675

fbshipit-source-id: 0eb1ce796