[scheduler][restart] Quantify wasted work
Summary: Count the time used by the `RestartScheduler` for analysis (useful) and the time wasted because a a worker was not able to take a lock to have a metric to compare different versions of the scheduler. The wasted time is not actually count but it can be calculated by substracting useful time from the total time. This was implemented like these to avoid substractions that may make the floating point calculations more complicated.

Reviewed By: ngorogiannis

Differential Revision: D19969960

fbshipit-source-id: 68a1132ca