[annotations] Fix annotation comments that cause javadoc target to fail
Summary:
When trying to add annotations to code examples, Javadoc gets confused about `@` sign and there's no good way to fix it so it's both OK to read as a comment in the editor and as a Javadoc HTML, no matter what combination of <code>/<pre>/{code} or escaping you use.

Here we prioritise ability to read the code comment from the editor and therefore the comments are detached from annotations.

Facebook

Reviewed By: mityal

Differential Revision: D20093801

fbshipit-source-id: 25867c27a