[cost] Do not report `EXPENSIVE_EXECUTION_TIME` when a cost of a node is top
Summary:
> We don't report when the cost is Top as it corresponds to subsequent 'don't know's. Instead, we
> report Top cost only at the top level per function

The previous code just ignored top costed nodes, so it was able to report a non-top cost that was
from another node.  For example,

```
void foo() {
     linear-cost();
     top-cost();
}
```

It reported inconsistent reports: `EXPENSIVE_EXECUTION_TIME` with a linear cost and
`INFINITE_EXECUTION_TIME` at the same time.

This diff fixes it not to report `EXPENSIVE_EXECUTION_TIME` when there is a node with the top cost.

Reviewed By: ezgicicek

Differential Revision: D20139408

fbshipit-source-id: 9fedd4aec