[nullsafe][refactor][messaging] Refactor code + more consistent messaging
Summary:
In the previos diff we restructured error rendering utils for
TypeOrigin.MethodCall.

In this diff we do the same with TypeOrigin field: lets make the code
consistent.

We also clearly distinct third party from all other possible cases in
this branch.

This changes messaging and reported errors for strict modes (see test cases), and I believe this is a net improvement.

Reviewed By: artempyanykh

Differential Revision: D20139741

fbshipit-source-id: 84f502553