[buck] fix target quoting regexp
Summary:
The check for whether a buck target needs quoting is incorrect, since it uses `-` inside a character class to denote a dash, not a character range. The correct way to do this is to put the dash as the first character in the class (or last).

Facebook

iknowthis_regex

Reviewed By: jberdine

Differential Revision: D20159112

fbshipit-source-id: be6750ed8