[sledge] Distinguish star and or when computing variable strength
Summary:
Sh.var_strength determines, in part, if an existential variable has
only a single occurrence. The objective of this is to determine that a
variable is merely a placeholder and not used to express additional
constraints. For this, it suffices to check the weaker condition that
there is a single occurrence in each branch of a DNF expansion.

Reviewed By: ngorogiannis

Differential Revision: D19973779

fbshipit-source-id: 2c90c61f4