[java] fix primitive types
Summary:
A java byte is not a short and a java character is not a C character. Fix those types by mapping them to what the spec ordains.

An extra advantage is that this establishes a bijection between java types and their project in the `sil` type system, so pretty printing in Java mode can actually be made to work.

Reviewed By: mityal

Differential Revision: D20330525

fbshipit-source-id: 00ac2294b