[nullsafe] A dedicated test for testing field nullability memoization
Summary:
Some (all?) of this is already tested in other tests, but this feature
is important enough (and the implementation is scattered accross the
whole code), so I found it useful to have a small test that ensures the
very basic things are working as expected.

See `NestedFieldAccess.java` that tests far more advances things, but
here we focus only very basic things: conditions, local variable
assignments, and explicit assignments.

Reviewed By: jvillard

Differential Revision: D20339056

fbshipit-source-id: a6cfd0043