[impurity] Track unique accesses
Summary:
Impurity domain was tracking all changes to variables (with a list of traces that containing all write/invalid accesses). This results in having long traces with multiple access events for the same variable. For instance,

```
 void swap_impure(int[] array, int i, int j) {
    int tmp = array[i];
    array[i] = array[j]; \\ included in the trace
    array[j] = tmp;      \\ included in the trace
  }
```
here we recorded both array accesses.

This diff changes the domain to include accesses so that we only keep track of a single trace per access. Array accesses are only recorded once.

Note that we want to record all unique accesses, not just the first one, because impurity will be used for hoisting/cost where we will invalidate impure arguments and consider all the rest as not changing.

Reviewed By: jvillard

Differential Revision: D20385745

fbshipit-source-id: d3647dad3