[nullsafe] Encapsulate creation of ReportableViolation.t inside ReportableViolation
Summary:
`to_reportable_violation` is responsible for identifying if the
violation is reportable in this mode or not.

This logic is higly coupled with other functions in
`ReportableViolation`, such as `get_description`: You can not have
sensible description on the violation that is NOT reportable in a given
mode.

So from logical perpsective, creation of `ReportableViolation.t` should
belong to this module itself, not to the parent `Rule` module.

This change will make design clearer.

Reviewed By: artempyanykh

Differential Revision: D20511756

fbshipit-source-id: ef27b5057