[nullsafe] Don't ever issue violation when dereferencing params.
Summary:
# Problem

Consider
```
some_method(Object a) { a.deref(); }
```

What is nullability of `a` when we dereference it?
Logically, things like "LocallyCheckedNonnull" etc are not applicable
here.

This would be applicable if we called some_method() outside! But not
inside. Inside the function, it can freely treat params as non-null, as
long they are declared as non-nullable.

The best we can capture it is via StrictNonnull nullability.

Reviewed By: artempyanykh

Differential Revision: D20536586

fbshipit-source-id: 5c2ba7f0d