[impurity] Consider functions with empty pulse summary as impure
Summary:
As exemplified by added tests, pulse computes an empty summary (with 0 disjuncts) whenever it discovers a contradiction which might be caused by:
- discovering aliasing in memory
- widening limited number of times in loops and concluding that loop exit conditions are never taken

However, AFAIU, it is not possible to have a function with 0 disjunct apart from such anomalities. Even a function which does nothing like `void foo(){}` has 1 disjuncts:
```
Pulse: 1 pre/post(s)
#0: PRE:
      { roots={ };
        mem  ={ };
        attrs={ };}
    POST:
      { roots={ };
        mem  ={ };
        attrs={ };}
    SKIPPED_CALLS: { }
```

The aim of this diff is to consider functions with 0 disjuncts as **impure** because most often such cases are impure, rather than actually pure.

Reviewed By: skcho

Differential Revision: D20619504

fbshipit-source-id: 3a8502c90