[python] migrate `infer explore --html` to OCaml, missing GitHub integration
Summary:
Re-implement the generation of an HTML report (with bug traces) in
OCaml.

Kills the --only-show as a side-effect, it is of dubious use since there
is already infer-out/report.txt to get the report list as text. A
follow-up diff adds numbers to the list in infer-out/report.txt for easy
cross-referencing with `infer explore --select 123`.

Reviewed By: skcho

Differential Revision: D20672769

fbshipit-source-id: 39b3a299d