[inferbo] Limit depth of abstract location
Summary:
This diff limits the depth of abstract location by a constant.

problem: Inferbo generated too many of abstract locations, especially when struct types had many pointer fields and Inferbo was not able to analyze the objects precisely. Since the number of generated abstract locations were exponential to the number of fields, it resulted in OOM in the end.

(reported by zyh1121 in https://github.com/facebook/infer/issues/1246)

Reviewed By: jvillard

Differential Revision: D20818471

fbshipit-source-id: f8af27e5c