Storing class source file locations in Struct (Tenv) records
Summary:
Java bytecode format does not record the declarations location
for classes and fields. We set up a first infrastructure to recover this
information. Currently we only track location for classes and only gives
the first line of the corresponding source file. We will enhance this location
with source file (baby) parsing in a next diff.

Reviewed By: mityal

Differential Revision: D20868187

fbshipit-source-id: d355475e9