[pulse] Distinguish error state at top level
Summary:
As soon as pulse detects an error, it completely stops the analysis and loses the state where the error occurred. This makes it difficult to debug and understand the state the program failed. Moreover, other analyses that might build on pulse (e.g. impurity), cannot access the error state.

This diff aims to restore and display the state at the time of the error in `PulseExecutionState` along with the diagnostic by extending it as follows:
```
type exec_state =
  |  represents the state at the program point that caused an error *)
```

As a result, since we don't immediately stop the analysis as soon as we find an error, we detect both errors in conditional branches simultaneously (see test result changes for examples).

NOTE: We need to extend  `PulseOperations.access_result` to keep track of the failed state as follows:
```
type 'a access_result  = ('a, Diagnostic.t * t [denoting the exit state] ) result

```

Reviewed By: jvillard

Differential Revision: D20918920

fbshipit-source-id: 432ac68d6