[java] kill ZipLib (aka specs-in-a-jar)
Summary:
Simplifies the models story. The model jar is still needed to look up
fields in modelled classes (we save the .class of each model!). This is
sad as we should be able to get these from the models' tenv, which
should be more compact, but that's a bigger change.

Reviewed By: skcho

Differential Revision: D20891117

fbshipit-source-id: fcb5104ae