[infer-out][4/9] migrate infer-out/tmp/
Summary:
First real step to keep an inventory of all the entries in infer-out/ in
a single place, and associate meta-data to each entry. In particular, we
want to avoid adding things in infer-out/ without a clear idea of
whether they should be cleaned up before going into a cache, or before
an incremental analysis.

Migrate infer-out/tmp/ first just as an example and an excuse to write
the scaffolding code needed for all the other entries.

Reviewed By: skcho

Differential Revision: D20894300

fbshipit-source-id: f796fca55