[sledge sem] Organise the tranlation corectness proof
Summary:
Separate into separate files the theorems that are just about the
translation (mostly about the structure of the variable->expression
mapping that the translation builds) from theorems about the translation
and the semantics.

Also move the stuff about dominator_ordered into the SSA Theory, since
it only makes sense for SSA programs, but doesn't have anything to do
with the translation.

Reviewed By: jberdine

Differential Revision: D20673124

fbshipit-source-id: 9d8b08164