[nullsafe] Don't explicitly recommend annotating a nested class as @Nullsafe
Summary:
In the previous diffs, nullsafe behavior was changed to the following:
nested class mode is inherited from the outer class mode.

Though it is possible to e.g. make nested class Nullsafe and outer not,
or make nested class STRICT and outer just LOCAL, this is an edge case
and we don't want to recommend annotating nested classes by default. The
right way is to make outer class Nullsafe instead.

In this decision, we take into account user experience and codebase
readability.

Reviewed By: ngorogiannis

Differential Revision: D21255806

fbshipit-source-id: 0200cb555