[nullsafe] Skip analysis of autogen methods
Summary:
We have a common entry point where we skip analysis in nullsafe.

This logic is copied from `Reporting.log_issue_from_summary`.
I believe this should not exist in Reporting: it is not the right place
to decide whether to suppress issues: we should not try to report it in
first place.

Because of that we falsely report "needs improvement" meta-issue while
we don't issue any (they were suppressed but participated in needs
improvement count calculations).

Now this change will make meta-issue to be synced with what the user
actually sees.

Down the line we should have a more reliable fix for that.
So far I reviewed suppressing code and looks like we should not suppress
anything else (unless explicitly SuppressLint-ed, which is fine).

Reviewed By: artempyanykh

Differential Revision: D21328634

fbshipit-source-id: 120ce06d1