[nullsafe] More consistent version of anonymous classes in JavaClassName
Summary:
There is a case where the class name is "A$1$B$2".
In this case, we would correctly say it is an anonymous class, and
return A$1$B as user defined, which is bad: now we can get the outer
class which will be A$1 that will be anonymous again.

This was leading to tricky bugs.

Now, get_user_defined_class name will return something that is indeeed
contains only user defined names.

There is one tricky pathological case left: when the outermost class is
anonymous, which can in theory occur. This might lead to other tricky
bugs, so the follow up will be to rewrite API of JavaClassName.t so that
this case is made clear for the client.

But lets unbreak nullsafe fore now first.

Reviewed By: ngorogiannis

Differential Revision: D21449686

fbshipit-source-id: b0ba4702e