[sledge] Remove Exp.size_of and Term.size_of
Summary:
Having `val size_of : Typ.t -> t` in the signature of `Term` and `val
size_of : t -> t` in the signature of `Exp` gives the impression that
`Term` and `Exp` know something about `Typ`. But they don't, those
functions are only trivial convenience wrappers, and only have a few
uses, so just inline them to clarify that it is `Typ` that knows about
the sizes of types.

Reviewed By: jvillard

Differential Revision: D21441535

fbshipit-source-id: 09b135a8c