[litho] Add sematics of assume null
Summary:
This diff adds semantics of assume null, heuristics.  When `assume(x == null)`, it removes the
methods called on the builder `*x` from the abstract state.

```
x -> {p}
p -> {method1 called}

assume(x == null)

x -> {p}
```

This heuristics is unsound: Even though `x` (a pointer to builder object) points-to an builder
object, which cannot imply that the object `p` does not exist in the concrete semantics.  The
unsoundness may appear when there is an alias (see the FP test added).

Reviewed By: ezgicicek

Differential Revision: D21502923

fbshipit-source-id: 2e392bd89