[pulse] Modeling std::vector::end()
Summary:
It is undefined behavior to dereference end iterator.

To catch end iterator dereferencing issues we change iterator model: instead of having `internal pointer` storing the current index, we model it as a pointer to a current index. This allows us to model `end()` iterator as having an invalid pointer and there is no need to create an invalidated element in the vector itself.

Reviewed By: ezgicicek

Differential Revision: D21178441

fbshipit-source-id: fd6a94b0b