[pulse] Add a flag to model methods for memory ownership transfer
Summary:
Just like `CFBridgingRelease` we want to be able to model functions that are specific to a given codebase that make a transfer of memory ownership so that developers don't need to worry about releasing that memory anymore, and hence, we don't want to report leaks on that memory.
Things get a little more complicated, because some of the functions we want to model are in a specific namespace, so with this flag we take both cases into account, when we are dealing with namespaces or not.

Reviewed By: jvillard

Differential Revision: D21404409

fbshipit-source-id: c36bd7afc