[uninit] Use type information from locals if argument is a pointer in function signature
Summary: In an intra-procedural analysis, we assume that parameters passed by reference to a function will be initialized inside that function. To do that we use the type information of a formal parameter to initialize the fields of the struct. This was causing false positives if the formal parameter in function signature had type `void*`. To solve this, we used type information from local variables instead. However, we also get false positives for any kind of pointer if we use cast. We fix this by using type information of local variables as in `void*` case.

Reviewed By: jvillard

Differential Revision: D21522979

fbshipit-source-id: 4222ff134