[IR] Fix the function hashable_name for ObjC methods
Summary:
This function had been computing the name for ObjC methods wrong, with only the class name. This was causing wrong error messages in Pulse.

The main issue was that `Procname.to_simplified_string` was writing `Classname::methodname` for ObjC methods, which is not the convention. This confused the `hashable_name` funtion. So changing the method name to `Classname.methodname` which is more standard, and this also fixes `hashable_name`.

Reviewed By: ngorogiannis, jvillard

Differential Revision: D21570880

fbshipit-source-id: 13ed62cf8