[errlog] reporting asserts checker matches issue-type
Summary:
Add an extra argument everywhere we report about the identity of the
checker doing the reporting. This isn't type safe in any way, i.e. a
checker can masquerade as another. But, hopefully it's enough to ensure
checker writers (and diff reviewers) have a chance to reflect on what
issue type they are reporting.

Reviewed By: ngorogiannis

Differential Revision: D21638823

fbshipit-source-id: b4a4b0c0a