[UI] use human-readable issue type in console reporting
Summary:
Problem: issue types can be reported by several checkers. The current
solution is to change the name of the issue, eg `BIABD_USE_AFTER_FREE`.
This doesn't look great for the user.

We already store a "human readable" name for each issue. These need not
be unique. Use this instead in textual output. In order to keep the link
between the text output and the true "issue type", eg to pass to
`--disable-issue-type`, also print the `unique_id` part of the issue in
the summary:

```
examples/hello.c:12: error: Null Dereference
  pointer `s` last assigned on line 11 could be null and is dereferenced at line 12, column 3.
  10. void test() {
  11.   int* s = NULL;
  12.   *s = 42;
        ^
  13. }

Found 1 issue
          Issue Type(ISSUED_TYPE_ID): #
  Null Dereference(NULL_DEREFERENCE): 1
```

We could also print the issue id in each report but that looks worse.
Other tools use numbers for issue ids, but these are not descriptive at
all, eg `--disable-issue-type 86` is not very telling.

Reviewed By: skcho

Differential Revision: D21663957

fbshipit-source-id: 506b0fda9