[pulse] Special case for std::function:operator=( nullptr )
Summary: Assigning `nullptr` to `std::function` was causing `NULLPTR_DEREFERENCE` as our model was expecting to get an object in the right hand side of the assignment (`std::function::operator=`) and was dereferencing that object. Assigning `nullptr` to `std::function` removes callable object from it. We model this special case by creating a fresh value.

Reviewed By: skcho

Differential Revision: D21685318

fbshipit-source-id: 2d4af1933