[issues] make sure the immutable parts of issues are always the same
Summary:
We allow some fields of issues to be defined dynamically, more precisely
when loading AL files. We don't want this to be abused and in particular
we don't want to miss an issue being declared once for a checker and
another time for another as this would be hard to debug.

Also, only register unknown issue types as coming from AL if they
haven't been registered already.

Reviewed By: skcho

Differential Revision: D21685879

fbshipit-source-id: 9e9438a75