[cost] suppress check is_on_ui_thread for ObjC method
Summary: Cost analysis has an additional mode that checks whether the function occurs on a UI (main) thread. If so, it warns the user. This check is only supported for Java and C++ but not for ObjC, so the diff suppresses this check for ObjC, and set ```is_on_ui_thread``` to ```false``` by default.

Reviewed By: ezgicicek

Differential Revision: D21952470

fbshipit-source-id: 838dd5639