[sledge] Change: Return domain and range with Var.Subst constructors
Summary:
Change the `Var.Subst` `freshen` and `restrict` constructors to return
the domain and range of the substitution explicitly. Clients generally
need to compute them immediately, and they are at least partially
constructed during the initial substitution construction anyhow. This
may be an incidental minor optimization.

This allows removing the `apply_set` operation, as it's use can be
handled directly from the domain and range sets.

This also allows `Sh.rename` to be split into a function that assumes
that the substitution is restricted to the vocabulary of the formula,
and a wrapper that does this restriction and calls through. This
allows `Sh.freshen_xs` to be simplified slightly, and avoids some
redundant restriction, domain, and range computations.

Reviewed By: jvillard

Differential Revision: D21974017

fbshipit-source-id: aa8b3db24