[cost] extend BasicCost to BasicCostWithReason
Summary:
Extend BasicCost to BasicCostWithReason which contains a record of the form
```{cost: BasicCost.t; proc_name_list: Procname.t list}```

This is done so that we can keep track of top cost function.
So the idea is that 80% of functions with Top cost are caused by calling top-costed callees, i.e. callee's Top cost is simply propagated to its transitive callers, so the aim is to investigate such root callees along with the number of their transitive callers.

Therefore, we create an extension that match `cost` to the root cause function.

This diff only handles the extension. Details about how we update the root cause function is in the next diff.

Reviewed By: skcho

Differential Revision: D22158717

fbshipit-source-id: 6498d904f