[pulse] Add model for NSObject.init
Summary:
This model is very important in the analysis of ObjC classes because the pattern

```
- (instancetype)init {
  if (self = [super init]) {
    ...
  }
  return self;
}

```

is very common, so we need to know that if the super class is `NSObject`, the implementation of `init` is returning `self`, otherwise it's a skip function and we don't get the correct spec for the function.  We fix some memory leak FP with this model, see test.

Reviewed By: ezgicicek

Differential Revision: D22259281

fbshipit-source-id: 3ee48c827