[nullsafe] Fix handling of assignment expressions in prune
Summary:
We already had a heuristic to deal with assignment expressions, but it
relied on the very previous CFG node to have a non-empty list of instrs.

In some cases, however, this previous node is a Join_node with no instrs,
so we need to take one more step back to find what we're looking for.

I've also added a bit more logging around this functionality, so it's
easier to debug/tune in future.

Reviewed By: ngorogiannis

Differential Revision: D22282930

fbshipit-source-id: 024eec145