[log] fix gc stats merging
Summary:
Messed up the aggregation of GC stats in the previous commit.

It's cleaner to have GC stats (and analysis time) outside of
BackendStats as the rules for computing them is different than the rest,
eg notice how "analysis time" needed to be corrected at the end of the
run, and similarly for GC stats. Thus, refactor this part.

Also output different aggregations of GC stats: +/max/average.

Reviewed By: skcho

Differential Revision: D22332496

fbshipit-source-id: eefd9dd72