[pulse] change recency maps to be backed by lists
Summary:
This one is observed to be more memory efficient. Intuitively, maps need
to be re-allocated more often than lists for balancing. In pulse, we'll
often only ever add new values, in increasing order (when they are fresh
variables created as we symbolically execute the program), which pushes
maps into their worst-case allocation pattern. At least I suspect that's
what happens. With lists, this case is handled much better as lists are
not re-allocated when adding elements.

This is somewhat confirmed by benchmarking and observing GC stats.

Reviewed By: skcho

Differential Revision: D22140908

fbshipit-source-id: 29815112f