[sledge] Refactor: Separate out conversion from Llair to Fol
Summary:
The eventual aim is for the conversion of Llair to Fol to be external
to Fol. Fol should not need to depend on Llair, nor vice versa. This
is not yet possible, but a step forward is to move the conversion
functions into separate modules outside the core Fol modules.

Reviewed By: ngorogiannis

Differential Revision: D22170522

fbshipit-source-id: 4860b4c07