[clang] Remove protocols from the list of superclasses of ObjC classes
Summary:
The frontend was hackily adding protocols as superclasses in the tenv, with the implicit encoding that the first element in the list was the actual superclass. This was clearly very fragile.
Protocols are not used in the backend at the moment, so for now we will remove them from the list of superclasses to have more consistency in the tenv.

Reviewed By: ngorogiannis

Differential Revision: D22525078

fbshipit-source-id: 2aef1fab1