[pulse] model operator== and operator!= for iterators
Summary:
Pulse has models for iterators that make them use a fake field to
remember the element of the collection they point to. But, not all
methods are modelled, and some of them look at the real field, eg
`operator==`. Since we don't update the real field in the model, this
causes imprecision.

The imprecision was visible in pudge.

Reviewed By: skcho

Differential Revision: D22576003

fbshipit-source-id: 2af6be646