[cost] Revert subclass lookup heuristics
Summary:
In Cost/Inferbo checkers, it tried to find a subclass with heuristics when a method of interface or
abstract class is called.  While it makes preciser analysis results in general, sometimes introduced
tricky FPs since the behavior of the heuristics depends on targets.  This diff revert the heuristics
to suppress the FPs.

Reviewed By: ngorogiannis

Differential Revision: D22924485

fbshipit-source-id: 9f151231f