[nullsafe][refactor] Make AssignmentViolation depend on AnnotatedNullability and InferredNullability instead of Nullability
Summary:
This change will simplify the further refactoring I am doing in follow
up diffs.

Logically, AssignmentRule is about passing actual value to the declared
one, and the proper abstactions for this is exactly AnnotatedNullability
and InferredNullability.

This will also make the client less error prone (no way to call if you
don't have a declaration and actual value to compare).

In the next diff, we will do the same in DereferenceRule.

Reviewed By: artempyanykh

Differential Revision: D22923779

fbshipit-source-id: 5f8f0931c