[cost] Add a FP test due to imprecise loop invariant
Summary:
This diff adds a false positive test that is introduced imprecise loop
invariant detection.

In the example, `i` and a temp variable `$irvar = get_size(arr)` are
all included in the control variables, so the complexity becomes
quadratic.  The problem is that `$irvar` is not addressed as a loop
invariant:

* `is.read` is analyzed as modifying a global
* all return variables, including `$irvar`, of unmodeled functions are invalidated

Reviewed By: ezgicicek

Differential Revision: D23051414

fbshipit-source-id: 011fd086b