[specs] store analysis and report artefacts separately
Summary:
Constructing the report is done by reading all the summaries, and using certain parts thereof.  However, the payloads, which typically account for the greatest size, are not used (with the exception of costs).

This diff splits the storage of summaries into analysis and report summaries, and only reads and deserialises the latter for the report phase. This makes a big difference for runs with a large number of procedures.

Reviewed By: jvillard

Differential Revision: D23105072

fbshipit-source-id: 359067a0f