[sledge] Change: Revise Sh handling of empty and pure approximation
Summary:
Make the relationship between Sh.is_empty and Sh.pure_approx stronger
and more precise. In particular:

> If [is_empty q], then [pure_approx q] is equivalent to
> [pure (pure_approx q)].

This enables replacing Solver.excise_pure with a simpler pure_entails
function. In particular, the heavy reliance on normalization of pure
formulas to true or false literals is eliminated, and only pure
entailment is needed.

Reviewed By: jvillard

Differential Revision: D22571146

fbshipit-source-id: 2fca64a61