[cost] write ObjC's NSCollection model separately and refactor
Summary:
When implementing iterator, we find out that because some semantics of inferbo is Java-specific, we cannot simply use Java's `Collection` model for `NSCollection`.
So this diff writing `NSCollection` model separately.

This diff also extracts the common parts of `NSCollection` and `Collection` into `AbstractCollection` to combine the duplicate the parts.

Reviewed By: ezgicicek

Differential Revision: D22975159

fbshipit-source-id: daed3f99f