[cost] fix enumerator when not used in for-each loop in ObjC
Summary:
In the previous diffs, we implement enumerator in order to estimate the cost of for-each loop in ObjC, but when we have FP case when enumerator is used not in for-each loop. For example, the following code has top cost before the fix.

```
void nsarray_enumerator_linear_FP(NSArray* array) {
  id obj;
  NSInteger sum = 0;

  NSEnumerator* enumerator = [array objectEnumerator];

  while (obj = [enumerator nextObject]) {
    sum += (NSInteger)obj;
  }
}
```

Reviewed By: skcho

Differential Revision: D23294895

fbshipit-source-id: 50c7b359f