[sledge] Add: Normalization of constant, equal and negated subterms/formulas
Summary:
Strengthen normalization performed by Term and Formula constructors to
eliminate literal 0 subterms and true or false subformulas, as well as
cases where subterms or subformulas are either equal or opposite.

This strengthens the ability of Context.implies to prove formulas that
involve embedding or projecting between terms and formulas, as the
added normalization sometimes reduces if-then-else formulas to
literals that are then directly provable.

Reviewed By: ngorogiannis

Differential Revision: D23459512

fbshipit-source-id: 6d4d90399