[sledge] Improve: Normalize polarity of conditional formulas
Summary:
Normalize conditional formulas to ensure that their "condition"
formula is not "negative". This avoids redundant formulas such as `(x
= 0 ? p : q)` and `(x ≠ 0 ? q : p)`. The choice of which formulas are
"negative" for this purpose is mostly arbitrary, with the only real
constraint being that negating a negative formula should produce a
positive one.

Note that conditional formulas themselves are considered to be
"positive" since negating them produces another conditional formula
with the same condition formula.

Reviewed By: ngorogiannis

Differential Revision: D23487502

fbshipit-source-id: 63606d89c