[scheduler][restart] use filenames instead of procnames for dependencies
Summary:
Limit communication bandwidth and serialisation burden by sending procedure filename strings (which are bounded at ~100 bytes) instead of serialising procnames through the socket to the scheduler (which are unbounded and have been seen to reach ~30kB in the worst case for templated procedures).

Context:
Under the restart scheduler, a worker working on a procedure X that discovers a race on a dependency Y it needs fails the computation of X and sends to the scheduler the procname Y.  The next time X is about to be rescheduled, the scheduler checks whether Y is still being analysed, by checking if the lock for Y still exists.  This check uses the procedure filename already, so we can send that instead.

Reviewed By: jvillard

Differential Revision: D23554995

fbshipit-source-id: 9828e71a2