[sqlite] enforce attribute.is_defined iff procdesc is not NULL
Summary:
The `attr_kind` column has now exactly two possible values: defined and undefined (since D22187238 (https://github.com/facebook/infer/commit/61ae2d1e1bb9a6eb6aec24ee5b0a30ee80702278)). This is also what should be stored in the field `is_defined` in the procedure attributes. Finally, the cfg can be `NULL` or not.  This diff makes all three of these agree, in order to allow for the removal of the `attr_kind` column up the stack, since it will be equivalent to `cfg is NOT NULL`.

The changes in the tests indicate improved correctness: previously, specialising a modelled callee would result in a dummy entry in the procedures table, because all clang procedures were given a procdesc, so the specialisation would specialise the empty procdesc.  Now, the model is not shadowed by the specialised dummy procdesc, and is specialised itself (and that's why it also ends up in the capture DB, which is slightly counter-intuitive).

Reviewed By: jvillard

Differential Revision: D23536931

fbshipit-source-id: 983216cb6