[biabduction][models] delete objcpp copy
Summary:
ObjC++ models are a copy (symlink) into the ObjC ones, albeit with different compile time flags.  However, the resulting procnames are identical, and we are left with only one copy of models anyway.  This means that only one version (chosen arbitrarily by the build order) is used for analysis.

This diff deletes the ObjC++ version.

Reviewed By: jvillard

Differential Revision: D23704266

fbshipit-source-id: 1dc94251f