[frontend] Rearrange translation of lambda's body
Summary: Variables captured by reference without initialization are missing dereference in their type inside lambda's body. This causes the translation to miss one dereference. To fix this, we want to add missing reference to the type. However, we first need to make sure that lambdas body is translated after the translation of captured variables.

Reviewed By: jvillard

Differential Revision: D23564099

fbshipit-source-id: 6a2ae053d