[inferbo] Introduce closure value
Summary:
This diff extends inferbo's domain to include closure values.  The goal of this extension is to
follow missing semantics where closures are handled as values, for example, a closure is assigned to
an object field, then it is got later to call.

Due to the bottom-up nature of the analyzer, sometimes we don't know which values are written in a
field, which is the same for the other non-closure values.

Reviewed By: ezgicicek

Differential Revision: D23932186

fbshipit-source-id: 4a575d0de