[cost] Refactoring: Avoid using functors that are used only once
Summary:
This diff avoids some usages of functors when they are applied only once in `polynomials.ml`.  While
functor is in general helpful to abstract data and make code cleaner, it makes us hard to understand
code, eg `merlin-locate` does not give an actual definition point of terms when it is from a module
parameter.

Reviewed By: ngorogiannis

Differential Revision: D23991164

fbshipit-source-id: c61289e84