[cost] Extend the polynomial domain (preparation)
Summary:
This diff is a preparation to extend the polynomial domain.  What I will do in the following diff is
to extend the polynomial domain to include a symbol representing "cost of function pointer call".
Then, the symbol will be instantiated to another polynomial cost at call sites.

The polynomial domain is defined as a sum of

* non negative insteger
* map from `bound` to polynomail domain

```
type poly = {const: NonNegativeInt.t; terms: poly M.t}
```

The problem is all `bound`s are adderessed as integer values, rather than polynomial, thus it
does not fit to include the symbol for the "cost of function pointer call".

In this diff, it introduces a `Key` module for the map `M`, the type of `Key.t` is the same to the previous key of `M`. The actual extension of the key type will be in the following diff.

Reviewed By: ezgicicek

Differential Revision: D23991401

fbshipit-source-id: cec64347f