[scheduler] document option
Summary:
I wanted to change the default to "callgraph" but that created issues in
our tests, introducing flaky behaviours and even a failure due to trying
to run the pre-analysis multiple times (not 100% sure it was related).

Instead, document the various options and put the option in the analysis
manual so users can choose.

Reviewed By: martintrojer

Differential Revision: D24193751

fbshipit-source-id: 4b7c33a79