[preanalysis] Add a new preanalysis for ObjC's copy and mutableCopy
Summary:
In ObjC, `NSObject.copy` returns the object returned by `copyWithZone:` on the given class. This method must be implemented if the class complies with `NSCopying` protocol. Since we don't have access to `NSObject`'s code, to follow calls into `copyWithZone:`, we replace such `copy` calls with calls to `copyWithZone:` when a) such a method exists in the class and b) the class conforms to `NSCopying` protocol.

This is done in the preanalysis because
- we need to know if there is a `copyWithZone:` method in the class.
- so that other analyses also benefit (as opposed to doing this in cost and inferbo models).

Note that `NSObject` doesn't itself conform to `NSCopying` but all its  subclasses must confrom to the protocol and support the same behavior as above.

https://developer.apple.com/documentation/objectivec/nsobject/1418807-copy

 Similarly for `mutableCopy` -> `mutableCopyWithZone:` for classes implementing `NSMutableCopying` protocol.

Reviewed By: skcho

Differential Revision: D24218102

fbshipit-source-id: 42900760e