[sledge] Represent arithmetic terms using polynomials
Summary:
Change the representation of Fol terms to use polynomials for
arithmetic. This is a generalization and simplification of those used
in Ses. In particular, the treatment of division is stronger as it
captures associativity, commutativity, and unit laws, plus being the
inverse of multiplication.

Also, the interface is staged and factored so that the implementation
of polynomials and arithmetic is separate from the rest of terms.

Reviewed By: jvillard

Differential Revision: D24306108

fbshipit-source-id: 78589a8ec