[cost] Enable passing enumerator as paramter in ObjC
Summary:
This diff revises memory model of enumerator in ObjC to enable passing it as a parameter.

The cost checker was not able to analyze a function precisely when it gets an enumerator as a
parameter because the offset of an enumerator was available only when the analyzer knew the correct
relation between the enumerator and an array.

This diff simplifies the enumerator to have a similar value with `array->elements`, so its offset can
be taken without the relation between enumerator and array to get them.

Reviewed By: ezgicicek

Differential Revision: D24446574

fbshipit-source-id: 27cdc051e