[sledge] Switch IArray from Core_kernel.Array to NS.Array
Summary:
Change implementation of IArray from a wrapper of
Core_kernel.Array.Permissioned to NS.Array, and remove magic.  Also
add operations to Array and Iter in order to ensure that IArray is an
extremely thin wrapper of Array: only defining conversions to/from
arrays as well as adding hashing support.

Reviewed By: jvillard

Differential Revision: D24306095

fbshipit-source-id: 97b9187be