[sledge] Replace uses of Base.With_return with Iterators
Summary:
Expressing the sort of short-circuit evaluation in the changed code is
conceptually more direct using iterators.

Also, when using With_return, getting usable backtraces relies on the
compiler recognizing that the `raise` in the implementation of
`Base.Exn.raise_without_backtrace` should be a `reraise`. Using
iterators avoids this potential fragility.

Reviewed By: jvillard

Differential Revision: D24306094

fbshipit-source-id: b1abe04fb