[nullsafe][Annotation graph] Glueing together: preparing to build annotation graph
Summary:
Now we have all pieces of information needed to build the annotation
graph (to be finished as a follow up).

This diff utilizes the same approach that is done when we calculate node
promotions:
1/ AssignmentRule and DereferenceRule know that some issues are related
to provisional annotations and expose this information.
2/ ClassLevelAnalysis collects all the issues, extract violations from
them, and outputs them.

Next step will be changing the output to the meta-issue printing the
annotation graph in json.

Reviewed By: artempyanykh

Differential Revision: D24621410

fbshipit-source-id: c240da3aa