[nullsafe][annotation graph] Support provisionally nullable in fields
Summary:
This change is on par with the logic that we already have for methods
and method params.

In this logic, we explicitly distinct class under analysis & the
external class when fetching AnnotatedNullability (we needed this to support
various Nullsafe modes as well).

Here we use the same approach for fields.

NOTE: this is not intended to work with nested classes just yet.

Reviewed By: artempyanykh

Differential Revision: D24650934

fbshipit-source-id: f555bcc3f