[java][reporting] @SuppressLint now blocks _only_ specified issue types
Summary:
The problem in Reporting.ml:log_issue_from_summary is that it merely
checks the presence of `SuppressLint` annotation on method's body to
decide whether to log or not the issue. This means that regardless of
issue types specified in `SuppressLint`, all issues on such method will
get blocked.

Here we fix that.

Reviewed By: ngorogiannis, mityal

Differential Revision: D24726604

fbshipit-source-id: c9cae3833