[sledge] Drop fresh variables that do not appear in the solution
Summary:
In the process of computing `Context.solve`, fresh variables can be
generated. Not all of these end up in the final solution
substitution. Currently all of the freshly generated variables are
returned to the client, which leads to extraneous existentials. This
diff trims the returned fresh variables to only those that appear in
the final solution.

Reviewed By: ngorogiannis

Differential Revision: D24746241

fbshipit-source-id: 59a2f221b