[sledge] Do not translate LLVM vector types
Summary:
While vector types can be translated, vector operations are not
currently handled. Translating type can lead to later failures (such
as bogus Convert expressions) that are not clearly identifiable as
originating from vector types. So fail eagerly when a vector type is
seen. Note that the only vectors that the frontend sees are due to
incompleteness of the LLVM vectorizer pass.

Reviewed By: jvillard

Differential Revision: D24746224

fbshipit-source-id: 30c0715eb