[sledge] Base implementation of localization on atoms instead of vars
Summary:
The computation of provable reachability through the heap currently
uses a set of variables whose values are either determined by the
desired roots or by the heap constraints. This requires globals to be
treated as variables. In preparation for distinguishing globals from
variables, this diff changes the reachability computation to use a set
of atomic terms instead of variables.

Reviewed By: jvillard

Differential Revision: D24886573

fbshipit-source-id: c0e6763b6