[clang] fix bad interaction between ConditionalOperator and initializers
Summary:
This is several inter-connected changes together to keep the tests
happy.

The ConditionalOperator `b?t:e` is translated by first creating a
placeholder variable to temporarily store the result of the evaluation
in each branch, then the real thing we want to assign to reads that
variable. But, there are situations where that changes the semantics of
the expression, namely when the value created is a struct on the stack
(eg, a C++ temporary). This is because in SIL we cannot assign the
*address* of a program variable, only its contents, so by the time we're
out of the conditional operator we cannot set the struct value
correctly anymore: we can only set its content, which we did, but that
results in a "shifted" struct value that is one dereference away from
where it should be.

So a batch of changes concern `conditionalOperator_trans`:
- instead of systematically creating a temporary for the conditional,
  use the `trans_state.var_exp_typ` provided from above if available
  when translating `ConditionalOperator`
- don't even set anything if that variable was already initialized by
  merely translating the branch expression, eg when it's a constructor
- fix long-standing TODO to propagate these initialization facts
  accurately for ConditionalOperator (used by `init_expr_trans` to also
  figure out if it should insert a store to the variable being
  initialised or not)

The rest of the changes adapt some relevant other constructs to deal
with conditionalOperator properly now that it can set the current
variable itself, instead of storing stuff inside a temp variable. This
change was a problem because some constructs, eg a variable declaration,
will insert nodes that set up the variable before calling its
initialization, and now the initialization happens *before* that setup,
in the translation of the inner conditional operator, which naturally
creates nodes above the current one.

- add a generic helper to force a sequential order between two
  translation results, forcing node creation if necessary
- use that in `init_expr_trans` and `cxxNewExpr_trans`
- adjust many places where `var_exp_typ` was incorrectly not reset when translating sub-expressions

The sequentiality business creates more nodes when used, and the
conditionalOperator business uses fewer temporary variables, so the
frontend results change quite a bit.

Note that biabduction tests were invaluable in debugging this. There
could be other constructs to adjust similarly to cxxNewExpr that were
not covered by the tests though.

Added tests in pulse that exercises the previous bug.

Reviewed By: da319

Differential Revision: D24796282

fbshipit-source-id: 0790c8d17