[pulse] do not stop at the first error in function calls
Summary:
We deliberately stopped as soon as an error was detected when applying a
function call. This is not good as other pre/posts of the function may
apply cleanly, which would allow us to cover more behaviours of the
code.

Went on a bit of a refactoring tangeant while fixing this, to clarify
the `Ok None`/`Ok Some _`/`Error _` datatype returned by PulseInterproc.

Now we report errors as soon as we find them during function calls but
continue accumulating specs afterwards.

Reviewed By: da319

Differential Revision: D24888768

fbshipit-source-id: d5f2c29d7