[nullsafe] Record correct param index in json
Summary:
In all other places, we index params from 0, but accidentally recorded
the wrong number in json. It was because of the confusion between index
and user-visible param position that we show for the user message.

This diff fixes it: now we use 0-based indices internally (but of course still
report 1-based ones in the error message).

Reviewed By: artempyanykh

Differential Revision: D24916878

fbshipit-source-id: 45532c5ff