[sledge] Fix vocabulary handling of symbolic execution of multi-spec insts
Summary:
Instructions that have multiple specs sometimes have inconsistent
footprints for one or more of their specifications. In such cases, the
handling of the existential and universal vocabularies was sometimes
incorrect before this diff. In particular, the ghosts of a spec need
not appear in the pure approximation of the footprint, which the
previous code incorrectly assumed. Also, the unit of disjunction is
`false` with an empty vocabulary, not the precondition's vocabulary as
the previous code incorrectly used.

Reviewed By: ngorogiannis

Differential Revision: D24989067

fbshipit-source-id: eca3bff55