[sledge] Move handling of realpath to the frontend and cli
Summary:
The use of realpath on paths obtained from debug info and the current
working directory is application-usage-specific behavior that does not
belong in the backend library. This diff moves these uses to the
frontend and cli, respectively. Also, the use of realpath in the
frontend is memoized along the same lines as the other frontend
translation functions.

This was also the last use of `core` in the `sledge` library, so the
dependency is moved to `sledge_cli` and `sledge_report`.

Reviewed By: ngorogiannis

Differential Revision: D24989070

fbshipit-source-id: c21b275f5