[clang] force node creation in switch statements
Summary:
The translation of `switch` cases needs to insert nodes around the
translation of each `case` sub-statement, so we need to force node
creation in these sub-statements so the nodes around it can be connected
to the translation of the sub-statements.

Also added more logging I found useful when debugging that.

Reviewed By: da319

Differential Revision: D24991455

fbshipit-source-id: d3a622142