[impurity] Collect all accesses
Summary:
Previously, impurity analysis only collected one access for a single modification but not all other modifying accesses. This diff

- changes the impurity domain to collect all modifying accesses
- tracks and prints all the accesses seen to reach the modification, improving readability&debugging

Recording all accesses are needed in the next diff to determine if a method modifies any immutable fields. To determine that, we need to know all modifications, not just a single one.

Reviewed By: skcho

Differential Revision: D25186516

fbshipit-source-id: 43ceb3cd8