[sledge] Fix scope on return in case actual return clashes with formals
Summary:
It can happen that the actual return variable, which is the name of
the variable in the caller's scope that receives the returned value,
clashes with the formals of the callee. The scope of the return
formula was wrong in this case, as the actual return was outscoped via
existential quantification with the rest of the formals after the
return values was passed.

Reviewed By: jvillard

Differential Revision: D25196727

fbshipit-source-id: 94cf25418