[cost] Record zero operation cost for procedures that simply throw
Summary:
Developers complain when a function that used to only throw an exception has complexity increase in the updated revision. Let's suppress such issues by giving those functions 0 cost which is already suppressed by differential reporting.

One common case to the above throw pattern is Java methods that throw an unsupported implementation exception for a functionality that has not been implemented yet. When the developer adds the supported implementation, we don't want to warn them with complexity increase since they are adding new functionality.

This is a design choice/heuristic to prevent noisy results for now.

Reviewed By: skcho

Differential Revision: D25495151

fbshipit-source-id: 94a82b062