[sledge] Freshen nested existentials also wrt ancestor universals
Summary:
Previously Sh had a representation invariant which ensured that
existentials of a subformula did not shadow universals of its
superformulas. The current implementation of Sh.freshen_nested_xs
mistakenly still assumes this invariant, but it no longer holds. This
diff fixes freshen_nested_xs to freshen nested existentials with
respect to not only ancestor universals in addition to existentials.

Reviewed By: jvillard

Differential Revision: D25756551

fbshipit-source-id: 54c48b067