[sledge] Improve Sh.simplify using stronger Context.elim
Summary:
The implementation of quantifier elimination in Sh.simplify currently
relies on a subtle implementation detail of Context.apply_subst to
remove some identity mappings. Now that Context.elim has been
strengthened and generalized, it can be used to to give a much clearer
implementation, that is also more robust to representation changes in
Context.

Reviewed By: jvillard

Differential Revision: D25756554

fbshipit-source-id: 6be0de2f3