[sledge] Fix bug in Sh.simplify
Summary:
When a stem has an occurrence of an existential, say x, and x is
substituted out by different witnesses, say a and b, in two disjuncts,
then the connection to other occurrences of a and b are lost.

This diff fixes this problem by propagating the solved variables that
survive normalization down to subformulas, and not removing those
variables from subformulas.

Reviewed By: jvillard

Differential Revision: D25756583

fbshipit-source-id: dabda743f