[sledge] Compile debug mode with optimization enabled
Summary:
The ~20% hit in compilation time is dominated by the time saved
running the tests. This also makes the memory allocation numbers more
robust to inconsequential changes. Note that `make check` is still
fast.

Note that normally dune passes `-g` to ocamlopt itself, but not if the
build profile includes an `ocamlopt_flags` entry. So add it explicitly
to ensure that compilation includes debug symbols.

(Also remove `-error-style short` that was missed in 9c4f263.)

Reviewed By: jvillard

Differential Revision: D25756570

fbshipit-source-id: 94d18bac2