[sledge] Strengthen Context.solve_for_vars on uninterpreted functions
Summary:
Context.solve_for_vars returns a solution substitution consisting of
oriented equalities that are implied by the given context. It is
logically valid to express these equations using terms that are
normalized with respect to the solution substitution itself. This diff
normalizes uninterpreted terms with the solution substitution when
extending it. This preserves the logical strength of the solutions,
and strengthens other purely syntactic operations on the substitution,
such as Context.Subst.is_valid_eq.

Reviewed By: jvillard

Differential Revision: D25756582

fbshipit-source-id: cd997c46b