[frontend] Handle struct copy in C
Summary:
When there was an assignment of C struct, `x = y;`, it was translated to the statements of load and store.

```
n$0 = *y
*x = n$0
```

However, this is incorrect in Sil, because a struct is not a value that can be assigned to registers.  This diff fixes the translation as assignments of each field values :

```
n$0 = *y.field1
*x.field1 = n$0
n$0 = *y.field2
*x.field2 = n$0
...
```

It copies field values of C structs on:

* assign statement
* return statement
* declarations.

It supports nested structs.

Reviewed By: jvillard

Differential Revision: D25952894

fbshipit-source-id: 355f8db9c