[pulse] Loosen fieldname comparison in state
Summary:
When C and C++ code handle a common struct typed value, the struct
type is handled as a `CStruct` in the C code, but as a `CppClass` in
the C++ code.  On the other hand, `Fieldname.t` contains a string of
field and **the struct type**.  As a result, even if a same field is
accessed in C and C++ code, the accessed fieldnames are different.

```
void callee_in_c(struct s* x) {
  x->a = 3;
}

void caller_in_cpp() {
  struct s x;
  x.a = 5;
  callee_in_c(&x);
  // HERE
}
```

For example, in the above code, `caller_in_cpp` sets the field `a` as
5, then calls `callee_in_c`, which sets the field `a` as 3.  However,
at `HERE`, the value of `x` in Pulse is `{a -> 5, a -> 3}`, because the two
fieldnames are addressed as different ones.

To avoid the issue, this diff loosens the fieldname comparison in
Pulse.

Reviewed By: jvillard

Differential Revision: D26000812

fbshipit-source-id: 77142ebda