Add a notion of retries to database serialization
Even with O_SHARE_DELETE, the "atomic-rename" mechanism used to make
sure that the summaries are written atomatically does not work under
Windows, which offers fewer atomaticity guarantees regarding
filesystem operations. So, we add a parameter 'retry' to
with_intermediate_temp_file_out, and we re-attempt rename operations
that fail. (Because the file was likely temporarily locked.)
We randomize a bit the delay before retrying, in order to minimize
the change of further races.