Use mode O_SHARE_DELETE when opening files
This has an effect under Windows only, and allows the file to be
deleted (or renamed) by a process even though it is opened in read
mode by another process.  This is used to handle the race conditions
that arise on database summaries updates.

We also catch the Unix_error, and re-raise the exception that
would have corresponded to the original call.