[frontend] Fix incorrect order of statements
Summary:
This diff fixes incorrect order of statements on assignments.

In the translation of `LHS=RHS;`, if `RHS` is a complicated expression that introduced new nodes, eg a conditional expression, some load statements for `LHS` came after its usage.  To avoid the issue, this diff forces it to introduce new nodes for `LHS`.

Reviewed By: jvillard

Differential Revision: D26099782

fbshipit-source-id: 27417cd99