[sledge] Fix Arith.trms to include proper subterms of uninterpreted terms
Summary:
For arithmetic terms that are a single monomial, Arith.trms is
currently the singleton iterator containing the term itself. In
particular, Arith.trms is not the *sub*terms in this case. In this
case the principal constructor is multiplication. This diff fixes
these cases and defines Arith.trms to contain the factors of a
singleton monomial.

Reviewed By: jvillard

Differential Revision: D25883719

fbshipit-source-id: a88611bba