Restore support for recovering from Ctrl-C
When using fork/exec to implement sub-processes, signals such as SIGINT
are not handled as when using fork. So we let the father receives the
SIGINT, and then have it forcefully kill the children at the end of
its execution if they are still alive somehow.

As a consequence, remove the existing epilogue, which does not force
children to die.