[cost] Remove expensive modeling for Provider.get
Summary:
We are getting lots of FPs due to modeling `Provider.get` as expensive. This is coming from Dependency Injection and Infer cannot statically determine the type of the provider and determine whether that provider is expensive (requires a global analysis and instrumentation).

Instead, we are downgrading this method to the default constant cost.

Reviewed By: skcho

Differential Revision: D26223978

fbshipit-source-id: 79f81c997