[frontend] Reset id generator before generating ObjC getter/setter
Summary:
This diff resets the id generator before generating ObjC getter/setter, so parsed results are the
same without regard to the generation order.  Note that the order may change when we change the type
of Procname.t since their hash values are used for the hash set of procnames.

Reviewed By: ngorogiannis

Differential Revision: D26277348

fbshipit-source-id: a66d77845