[clang] add a new flag --skip-non-capture-clang-commands to makes clangdriver skip compilation on commands that doesn't capture data
Summary:
ClangWrapper.ml was skipping clang commands that didn't capture
by default. It was using the 'skip_analysis_in_path_skips_compilation'
flag to NOT skip commands. This is a confusing use of that flag.

Default should be to run clang (in case it does something useful),
and a new flag to disable this.

Reviewed By: ngorogiannis

Differential Revision: D26459100

fbshipit-source-id: 7f2e9a269