[sledge] Optimize by inlining functors
Summary:
Inline functor applications, principally those involved in the
representation of (arithmetic) terms. This enables direct calls to
functions in the result of functor applications, instead of indirect
via an offset from the start of the module block resulting from the
functor appication. This also then enables further inlining
opportunities.

Reviewed By: jvillard

Differential Revision: D26338016

fbshipit-source-id: 27b770fa3