[sledge] Change encoding of program var ids to preserve original order
Summary:
Negating the ids of program variables leads to inverting the order on
them. This is logically fine, the order is still a valid total order.
But it can lead to choosing younger variables as equality class
representatives over older variables, and thereby lead to more churn
as adding an equality is more likely to cause a change of
representative, and hence additional normalizing rewrites.

Reviewed By: jvillard

Differential Revision: D26451304

fbshipit-source-id: eb20d1901