[liveness] ignore most dead stores when the variable has been passed by reference
Summary:
See added tests. Passing a variable by reference to a function `foo` can
cause the variable to be added to the global state so any stores after
that might be live as long as there is another function call after the
store (since the global state shouldn't outlive the scope of the
function). Currently we don't check that the latter is true; to report
these we would need to extend the abstract domain to remember which
stores have been made without a subsequent call.

Also change Blacklisted -> Dangerous everywhere since the corresponding
option is called "liveness_dangerous_classes".

Reviewed By: skcho

Differential Revision: D26606151

fbshipit-source-id: e869e5df1