[pulse][nullptr][objcpp] Warn on sending a message to nil when return type is non-POD
Summary: In ObjC, when a method is called on nil, there is no NPE, the method is actually not called and the return value is 0/false/nil. There is an exception in the case where the return type is non-POD. In that case it's UB and we want to report an error.

Reviewed By: jvillard

Differential Revision: D26815687

fbshipit-source-id: 8126414ab