[pulse] refactor errors
Summary:
This will enable further improvements: basically we want to be able to
abort the symbolic execution of a single disjunct whenever an error is
detected. Right now there is only one kind of error, which is now
explicitly called `ReportableError`.

The next diff refactors Pulse.ISL to add its own error type so that we
are able to get rid of the isl_status field (ISLOk/ISLError) inside
abductive states. ISLError states are really `Error _` states but
previously it would have been too much of an API change to expose that.
Now it's all going to be part of `AccessResult.t`.

A further change will add another error type for when a value is found
to be 0 after the fact by the arithmetic.

Reviewed By: ezgicicek

Differential Revision: D26821178

fbshipit-source-id: 2923db8e7