[pulse] protect against Z exceptions
Summary:
There could still be divisions by zero, eg in the "mod" case: consider
"x mod (1/2)" (doesn't matter what x is). Then we'd check "1/2 =? 0" and
since it's false conclude that it's safe to take the modulo... oops!

To make things safer, harden `Z` to not throw anymore.

Also add a layer of defense in depth by wrapping the functions that do
Z/Q operations in another layer of exception catching because we really
don't want to crash the entire analysis due to that.

Reviewed By: martintrojer

Differential Revision: D27262569

fbshipit-source-id: e22187ca0