[CLI] rename biabduction-specific option
Summary:
Back in the days©, biabduction was the only analysis in infer, so a lot
of things implicitly assume we are talking about biabduction. Update
option names that are biabduction-specific to start with
`--biabduction-`.

Also update a few other options in the same spirit, eg for quandary, and
amend some doc messages just because they were not informative enough.

Also do some alphabetisation and moved a *bunch* of options that were in
the wrong section of Config.mli.

Reviewed By: skcho

Differential Revision: D27268094

fbshipit-source-id: a1c784188