[sledge][llvm] Handle nullptr in Llvm.global_initializer
Summary:
LLVMGetInitializer returns nullptr in case there is no
initializer. There is not much that can be done with nullptr in OCaml,
not even test if it is null. Also, there does not seem to be a C or
OCaml API to test if there is an initializer. So this diff changes
Llvm.global_initializer to return an option.

Reviewed By: jvillard

Differential Revision: D27188302

fbshipit-source-id: 3474ec840