[sledge] Add convenience links to executables
Summary:
The names of the compiled sledge executables are long, which is
inconvenient. Also, all build modes produce an executable named
`sledge`, so it is not possible to add them to PATH and still be able
to run the desired executable. This diff adds a directory of symbolic
links with different names that refer to the executables built in the
different build modes. The intent is for people to add .../sledge/bin
to their PATH.

Reviewed By: jvillard

Differential Revision: D27315805

fbshipit-source-id: 7e84e43a7