[opam] Move the opam files to an opam directory at repo root
Summary:
This diff renames and moves the opam package definition files to
`<package>.opam` in an `opam` directory at the root of the
repository. This enables opam pinning the different packages contained
in the repo. It is necessary for these files to be either at the root
of the repo or in a directory named `opam` at the root of the repo, or
else opam cannot find them.

Reviewed By: skcho

Differential Revision: D27326495

fbshipit-source-id: 4c95c6955