[pulse] cap the size of literals in formulas
Summary:
On some pathological examples of crypto primitives like libsodium, later
diffs make pulse grind to a halt due to an explosion in the size of
literals. This is at least partly due to the fact the arithmetic doesn't
operate modulo 2^64.

Due to the fact the arithmetic is confused in any case when we reach
such large numbers, cap them, currently at 2^128. This removes pathological
cases for now, even now on libsodium Pulse is ~5 times faster than before!

Take this opportunity to put the modified Q/Z modules in the own files.

Reviewed By: jberdine

Differential Revision: D27463933

fbshipit-source-id: 342d941e2