[ConfigImpact] Use cost analysis to determine cheapness
Summary:
Reporting all ungated (un configed?) function calls causes many FPs. Instead, we rely on complexity analysis to determine whether a function is cheap/expensive: if the callee's complexity is not symbolic (e.g. constant), we consider it as cheap and don't keep track of it.

Note that we don't take the instantiated/modeled cost into account yet. So, if we have  `foo(int n)` with complexity `O(n)`, and call it as `foo(3)`, we would still keep track of it. Similarly, if `foo` is a modeled function with constant time complexity, we would have no summary for it hence would keep track of it.

These will be improved later.

Reviewed By: skcho

Differential Revision: D27430485

fbshipit-source-id: d5f66320d