[sledge][llvm] Omit unnecessary GC root registrations
Summary:
The current code does not follow the simple interface to the OCaml GC,
where GC roots are registered conservatively, only initializing
allocations are performed, etc. This is intentional, as stated in the
opening file comments. On the other hand, the current code does
register GC roots in many situations where it is not strictly
necessary. This diff omits many of them.

Upstream Differential Revision: https://reviews.llvm.org/D99475

Reviewed By: ngorogiannis

Differential Revision: D27564873

fbshipit-source-id: f9e5b183f