[ConfigImpact] Take cost insantiation and models into account
Summary:
In D27430485 (https://github.com/facebook/infer/commit/a6ab4d38cfdce34307c333d25af6501aef13e2f2), we used the static cost of the callee to determine whether it was cheap/expensive. This diff improves on that by taking the whole instantiated cost of the function call (not just the callee's cost).

Also, if the callee is an unmodeled call, we consider it to be expensive as before.

Note: cost instantiation was used by hoisting. I refactored bunch of code there to reuse as much as code possible.

Reviewed By: skcho

Differential Revision: D27649302

fbshipit-source-id: 07d11f3dd