[pulse] Separate issue type for nil messaging of non-pod return type
Summary: Added a new issue type for sending a message to nil when its return type is non-POD. To distinguish these issues from other nullptr dereference issues, we extend the `MustBeValid` attribute to contain the reason of why an address must be valid. For now a reason can only have `SelfOfNonPODReturnMethod` as it's value, but in the future we will use it for other nullability issue types, such as nil insertion into collections.

Reviewed By: jvillard

Differential Revision: D27762333

fbshipit-source-id: 689e5a431