[ConfigImpact] Filter known expensive callees when cost is constant
Summary:
This diff filters known expensive callees when cost is constant.

previous:
```
foo() {
  known_expensive_call();
}
```

current:
```
foo() {
  known_expensive_call();
  goo();
}

// cost is constant
goo() {
  known_expensive_call();
  unknown_call();
}
```

When callee's cost is constant and its summary includes known expensive callees, the checker addressed it as a non-constant-cost callee, i.e., it copies all ungated callees from the callee's summary. However, sometimes this full-copying introduces unexpected issues. For example, suppose a callee `goo` is added and `goo`'s cost is constant as above. Since it includes `known_expensive_call`, all ungated callees of its summary is copied to the caller `foo`'s summary:

* `foo`'s ungated callees (before): {`known_expensive_call`}
* `foo`'s ungated callees (after): {`known_expensive_call`, `unknown_call`}

As a result, it would report about `unknown_call` is added. However, this is not what we intended: In the example, `unknown_call` is reported because it is called in the same function with `known_expensive_call`, not because it is expensive.

To fix that issue, this diff filters known expensive callees from `goo`'s summary in that case.

Reviewed By: ezgicicek

Differential Revision: D27852552

fbshipit-source-id: d207eef1c