[ConfigImpact] Add semantics for temporary boolean variable
Summary:
This diff adds semantics for temporary boolean variables to keep config values.

* It extended value domain to have `TempBool` that is basically a pair of `ConfigChecks.t`; one is a
  set of config values checked when the temporary variable is true, and the other is that when the
  temporary variable is false.
* It assigns the `TempBool` value when `temp=1` or `temp=0`.
* It uses the `TempBool` value when pruning condition expression.

For example, when there is an `if` statement of

```
return (config && b);
```

it is translated in SIL,

```
if (config) {
  if (b) {
    temp = 1; // (1)
  } else {
    temp = 0; // (2)
  }
} else {
  temp = 0; // (3)
}

return temp;
```

then we can say

* When `temp` is true, i.e. at (1), it is gated by `config`
* When `temp` is false, i.e. at (2) and (3), we are not sure about the the gatedness; at (2) it is gated by `config` but at (3) it is gated by `!config`.

So, we record such information as a `TempBool.t` value.

Next, when we use the return value at its caller,

```
if (ret) {
  // then branch
} else {
  // else branch
}
```

We can say "then branch" part is gated by `config`, but we are not sure if "else branch" part is gated, by using the `TempBool.t` value of  `ret`.

Reviewed By: ezgicicek

Differential Revision: D28056490

fbshipit-source-id: e90d8afd3