[pulse] discard "pruned" atoms that refer to variables outside the pre
Summary:
See added test: pulse sometimes insisted that an issue was latent even
though the condition that made it latent could not be influenced (hence
could the issue could never become manifest) by callers because it was
unrelated to the pre, i.e. it came from a mutation inside the function.
In these cases, we want to report the issue straight away instead of
keeping it latent.

Reviewed By: skcho

Differential Revision: D28002725

fbshipit-source-id: ce9e6f190