Fix taskbar nesting for ondemand analyses (#1431)
Summary:
On big analyses, we routinely observe analyses nesting levels of the
order of ~5k (i.e. 5342> on the interactive taskar). This call depth
is meaningless given the size of the codebase we analyze.

Instead, I believe this is due to the fact that Ondemand does not
properly snapshot/restore the nesting level when starting an analysis,
and more importantly when abandoning parts of it. This commit
fixes that oversight.

Pull Request resolved: https://github.com/facebook/infer/pull/1431

Reviewed By: ngorogiannis

Differential Revision: D28118297

Pulled By: jvillard

fbshipit-source-id: 0faf28f84