[pulse] deduplicate histories and traces for memleaks
Summary:
Most/all of the time we expect the history of the value to faithfully
trace how it got allocated. That history was then added as a prefix of
the trace leading to the same place, leading to duplicate information in
the report trace.

We may need to do the same for other bug types.

Reviewed By: ezgicicek

Differential Revision: D28536891

fbshipit-source-id: a83a2d038