[ConfigImpact] Add FP test due to anonymous class
Summary: An anonymous class name includes an index number, for example `AnonymousClass$2` represents that it is the 2nd anonymous class implemented in the `AnonymousClass` class.  Problem is that when we insert a new anonymous class, all index of anonymous class names below increase by one, which introduces incorrect comparison on reportdiff.

Reviewed By: ezgicicek

Differential Revision: D28568753

fbshipit-source-id: 2a6c576eb