[pulse] more config options to model dynamic memory management
Summary:
Add an option for realloc and fiddle with the other options' help for
consistency.

Moved the memory leak test to memory_leak.c and added more.

Moved the place where we take the options into account closer to their
corresponding models to defend a bit against modifying one without
modifying the other.

Reviewed By: da319

Differential Revision: D28543340

fbshipit-source-id: 75894d06d