[erl-frontend] Create (empty) Procdesc for Erlang functions.
Summary:
Each Erlang function now has a Procdesc in `results.db`. The
ProcAttributes record if a function is exported or not by using the
access Public or Private, respectively.

This adds also `ErlangTypeName`. We use a fixed set of "type names" for
the different types of values in Erlang (i.e., for Erlang's "dynamic types").

Reviewed By: jvillard

Differential Revision: D28385954

fbshipit-source-id: f8278505a