[pulse][1/5] add a test that we report on the first null access
Summary:
Spoiler alert: we don't. The next diffs fix that.

When there are several invalid accesses to report at a function call
instruction, we want to report the first one to occur within the
function. This is to avoid confusing reports where pulse reports, eg, a
null dereference for a pointer at a point where it's already been
dereferenced before in the same function.

Reviewed By: skcho

Differential Revision: D28674730

fbshipit-source-id: acb029e4b