[pulse][4/5] add a path context to record timestamps
Summary:
Add a new `PathContext.t` component to the abstract state. For now it
tracks only the current "timestamp" of symbolic execution inside the
procedure, i.e. which step of symbolic execution we are in (bumped by 1
each time we've executed one instruction). In the future this will also
hold, eg, which conditionals we've been through on the path (for
reporting traces with that information).

Most of the diff is about propagating the path context through many of
the APIs.

We use timestamps only in `MustBeValid` attributes to report the first
incorrect access in a function call for now.

Reviewed By: skcho

Differential Revision: D28674726

fbshipit-source-id: 2cd825e73