[F] Fix strong updates when updating multiple locations.
If multiple destination locations are updated, they must be weakly
updated even if each one represents a single value. For example, if
a dereference of a pointer is updated and the pointer may point to
more targets, each target must be weakly updated.

* infer/src/bufferoverrun/absLoc.ml
(can_strong_update): If a single location is updated, the update is weak
if it represents multiple values.

* infer/src/bufferoverrun/bufferOverrunDomain.ml
(weak_update): Perform weak update even if the updated location
represents a single value because in this case there are multiple
locations updated.