[Top] Use top as a default for value of a location.
Use top as a default value of a location instead of bot. Top as
a default value is safer than bot.

To be able to write something to a location representing the content
of an array, write bot to it when creating such location in malloc
if the identical location isn't already present in the abstract
state (otherwise, we are in a loop, the location represents multiple
allocsites and has to retain original values).

Context:

In bufferoverrun, there is a single location representing all elements of
an array. Since this location represents multiple values, it is always
weakly updated. Before this commit, the initial value of this location was
bottom (that is, the first update of this location in a procedure was
"strong").

TODO: Note that there are three kinds of locations representing multiple
values - a location representing all elements of an array and a location
containing a starfield. Currently, both are treated uniformly and the latter
ones are thus never updated with anything different from top. It would probably
make sense to have bottom as a default value for locations containign a
starfield.

Note that given the fact that we have a single location representing all
array elements, we are not able to capture the common pattern when an
element of an array is updated and read and there is not any update of
any other element of the array in between. For example:

   Arr (I) := 1;
   ... -- no update of Arr
   pragma Assert (Arr (I) = 2);

To be able to do that, we could use a recency abstraction. To represent
elements of an array, two locations and one value would be used.
First location representing values of all elements of the array, the second
location representing the value of the latest update of the array, and
the identification of the lastly updated element.