[F] Don't prune locations that can be only weakly updated.
Locations which can be only weakly updated cannot be constrained using
a prune operation. Explicitly disable pruning of these locations.
This is needed in particular because MemReach.weak_update uses bottom
as a default value for such locations so they would be thus effectively
pruned otherwise (note: this is no longer true after the commit
c57d6c8 Use top as a default for value of a location).

This avoids false positive test always false message in the following
example:

         if Value < Arr (1) then
            return True;
         elsif Value < Arr (2) then -- don't want test always false msg
            return True;
         end if;