[F] Fix a value used to represent a pointer to a function.
Use an abstract value representing any non-null pointer to
represent value of a function name constant instead of Val.Itv.top
that represents any numerical value.

TODO: Use an abstract value exactly representing the function by
setting Val.func_ptrs to be a singleton set containing the function.
This is not done because it would be a more intrusive change requiring
modification of the type FunctPtr.t.