[F] Fix detection whether location represents multiple values.
The detection whether a location represents multiple values assumed
that the value of a location is stored in the abstract state at the
time of the detection. As a result, function Mem.is_rep_multi_loc
could return false for a location representing multiple values
resulting in FPs.

Example:
The expression A < B was evaluated to false because X was an alias of
both A and B and it wasn't correctly detected that it can represent
multiple values. X was Table->_data[*].Name_Len.