[Arrays] Produce top bound if symbol in a bound is imported as bot.
When importing bounds, if there was a symbol which was imported as bot,
the whole bound was imported as bot. In parctice, this can happen when
an uninitialized array is passed to a procedure where it is updated
or constrained. The interval representing the value of such array to
be imported then contains a symbol representing the value of the array
before the call. This symbol is imported as bot. Importing the whole
interval as bot leads to false positives and this commit fixes this
problem by importing the interval as top.

Example:
itv_callee = [min(-2147483648, Vec->_data[*].lb(s)),
              max(2147483647, Vec->_data[*].ub(s))]
given Vec->_data[*] = bot v_callee was imported as bot to the caller

Note that before implementing more precise representation for intervals,
we had v_callee = top which was imported as top