[ondemand] only look at capture DB or model summaries for CFGs
Summary:
Ondemand should not analyse a pre-analysed CFG so we should never load a summary to get its CFG in Ondemand.

The only exception is if the function is a model, where its (pre-analysed) CFG only lives in a summary.

This diff eliminates the general look for a summary and only uses it if the function is a model, otherwise it looks in the capture DB.

Reviewed By: da319

Differential Revision: D29932885

fbshipit-source-id: 781ddf51d