[pulse] Show Nil Block Call warnings only for objc blocks
Summary: Nil Block Call is ObjectiveC specific issue type, however, we also report it for cpp code when we call function pointer that is nullptr. We want to keep reporting for nullptr call, however, we want to change the issue type from Nil Block Call to NullPtr Dereference. We use call flags to determine if we call objc block.

Reviewed By: skcho

Differential Revision: D30603571

fbshipit-source-id: 8084d05fb