[pulse] remove unknown calls memory leak FPs
Summary:
When calling an unknown function (one for which we don't have a summary,
eg because we don't have the code for it), Pulse will reassign the
values of pointer parameters to represent arbitrary mutation by the
function. But, if these pointers had been allocated (or something
reachable from these pointers was) this can create a memory leak.

Deallocating everything reachable from such pointers removes such leaks,
but only intra-procedurally. The next diff makes this work
inter-procedurally too.

Reviewed By: ezgicicek

Differential Revision: D30730582

fbshipit-source-id: c63e6b7bb