[erl-frontend] Report badmap errors from map library functions
Summary: Previously the translation ensured that map functions are only called on maps. But library functions such as `maps:is_key(K, M)` can be called on any value `M`. If we pass something that is not a map, a `badmap` error should be reported. This diff checks and reports `badmap` in such cases, fixing false positives/negatives from the previous diff.

Reviewed By: jvillard

Differential Revision: D30877541

fbshipit-source-id: f829d6ef5