[istd] add RecencyMap.mapi
Summary:
I thought I needed this but I don't but it seems useful regardless.
Don't try to preserve physical equality on the "old" map anymore since
we don't preserve it on the "new" one anyway. This also allows us to
remove elements of the `old` map that have been overwritten in `new`.

Also change the signature of `fold` to be more standard and pass the key
to `~f` too.

Reviewed By: rgrig

Differential Revision: D30992299

fbshipit-source-id: c8de083e6