[clang] skip no-op calls to empty implicit destructors
Summary:
As explained in the code comment, these cause unknown calls to
destructors.

Here's an example of the issue this gets rid of:

```
// note: POD object, clang doesn't generate a body for Y::~Y()
// but does include it in the AST nonetheless
struct Y {
  int a;
};

struct X {
  Y _y;
  X() {}
  ~X() {} // because ~X() is here we inject a call to ~Y(this._y) at the
          // end of ~X() (really at the end of __infer_inner_destructor_~X()
};

void build_X() {
  X x;
} // ~X(x) called here -> unknown call to ~Y()
```

After the diff, we notice that `~Y()` has no effect when generating
`__infer_inner_destructor_~X()` and so do not inject a call to `~Y()`
anymore.

Reviewed By: ezgicicek

Differential Revision: D31054602

fbshipit-source-id: bc6c1680a