[sledge] Add checks that term simplification does not create new solvables
Summary:
Construction functions perform semantics-preserving simplifications.
This means that `Trm.t` and `Arith.t` can not be considered as a term
algebras generated by the constructor functions. That is, for a
constructor `cons : t -> t -> t`, `cons a b` need not have `a` as a
proper subterm, need not have principal constructor `Cons`, etc. It is
guaranteed, however, that the constructor functions do not create new
solvable/uninterpreted proper subterms. That is, each subterm in
`solvable_trms (cons a b)` is in either `transitive_solvables a` or
`transitive_solvables b`. This property is required by the first-order
solver.

Reviewed By: peterogithub

Differential Revision: D31129960

fbshipit-source-id: bac690bf6