[ConfigImpact] Join unchecked callees from methods with the same hash
Summary:
Before this diff, when there are multiple methods with the same hash, we took one of them that had the biggest unchanged callees set and used it in differential, representing the others.  However, we found this introduces non-deterministic FPs, e.g.

* when there are multiple methods with the same size of unchanged callees, it will take one of them non-deterministically,
* when code changes resulted in changing the method with the biggest unchanged callees, it will compare two totally different methods.

To avoid the non-determinism, this diff joins all unchecked callees of methods with the same hash, instead of taking one of them arbitrarily.

Reviewed By: ezgicicek

Differential Revision: D31171336

fbshipit-source-id: 831285b69