Readme.md
Hey,

The docker options need to be given first in-order to activate the docker command. If option "-v" is specified after "infer" (i.e. container name), the option "-v" is passed as an entry point to the container.

Please see CONTRIBUTING.md for how to set up your development environment and run tests