[clang] Add flags to copy ctor
Summary:
This diff adds a flag to function calls that marks whether a call is for a copy constructor. This will be useful when we try to determine unnecessary copies in the upcoming stack.

Initially, I tried to add a new procname kind for CPPCopyConstructor (like we do for `ObjC_Cpp.CPPConstructor` but that is not readily possible since we don't have a way to determine whether a `CXXConstructorDecl` is a copy constructor in our clang AST. If needed, we might be able to export the necessary flag(?). For now, calls to copy constructor `CXXConstructExpr`  however already had this flag exported, so I went with the low effort.

Reviewed By: skcho

Differential Revision: D31336637

fbshipit-source-id: 5bf0d8d0a