[Java frontend] Fixing Java bytecode local variable table before Sawja transformation
Summary:
The .class format contains a debug table to retrieve the
source name of local variables. The table generated by the Kotlin
compiler does not follow the same convention than the javac
compiler. This patch is a quick fix. it is not fully correct because
we don't check outgoing edge while going forward in the opcode
array, but we expect the bacward iteration to be always safe (and short).

Official JVM spec:
https://docs.oracle.com/javase/specs/jvms/se17/html/jvms-4.html#jvms-4.7.13

Reviewed By: ngorogiannis

Differential Revision: D31431271

fbshipit-source-id: 0cc608039