[sledge] Add direct support for atomic instructions
Summary:
Before this patch, the LLVM lower_atomic pass is used to transform
atomic instructions such as atomic compare and exchange to sequential
instructions. This patch removes use of this pass, and adds atomic_rmw
and atomic_cmpxchg instructions to LLAIR. Additionally, Fence can be
treated as a nop given the current interleaving semantics.

Differential Revision: D31412461

fbshipit-source-id: 6f3c98d58