[erl-pulse] Bugfix in handling dynamic types
Summary:
Pulse models for built-in Erlang functions sometimes have the pattern
"if type of x is T, then access these fields of x". The "if" part was
sometimes producing a subtly unsatisfiable state: attributes say that x
is an invalid reference, but path condition requires that x has type T.
This led to a crash.

The fix here works by not producing "subtly unsatisfiable" states, but
by explicitly pruning the exploration of such execution paths.

Reviewed By: mmarescotti

Differential Revision: D31543370

fbshipit-source-id: 88cd4299e