[sledge] Fix treatment of quantifier scopes in Context.interN
Summary:
When Sh.simplify intersects first-order contexts, it is possible to
violate the scopes of the existential quantifiers on disjunct
subformulas. This diff fixes this by treating Context.interN as a
combined projection/restriction and intersection operation.

It would be possible to add a separate projection operation that
trimmed a context to only unbound variables. However such an operation
would not be cheap, and it is more efficient to combine it directly
with the intersection operation.

Differential Revision: D31912679

fbshipit-source-id: 9c3ad372b