[sledge] Eliminate Sized term form
Summary:
The Sized term form acts as a mechanism to reify the size of a
flexibly-sized sequence / byte-array. This size information is
important to the canonization of Extract and Concat terms in
particular. Semantically, at the first-order level, the Sized term
`⟨n,a⟩` is equivalent to just `a`, since the underlying dependent sum
type and its pack and unpack operations are not treated directly by
the solver. This apparent equivalence has the result of allowing
redundant representations of sequence values, either with or without
sizes. In some uses (such as canonizing Extract or Concat), the sizes
are necessary, and in other uses (such as an indeterminate of a
polynomial), the sizes problematically block term simplification.

This diff resolves this issue by eliminating the distinct Sized term
form altogether. It turns out that is suffices to fuse the
size-tagging with the Extact and Concat constructors, so that they
operate only on explicitly sized terms, and then the sequences can
appear without sizes everywhere else.

This approach works, prehaps surprisingly, since it happens that
symbolic execution and the solver do not need the flexibility to
construct Extract or Concat sequence terms and only later tag the base
sequences with their sizes: the sizes happen to be known whenever
Extract or Concat terms need to be constructed. As a result, Sized
terms are not strictly necessary to communicate the size information.

Differential Revision: D31912683

fbshipit-source-id: 7fdd41b2e