[dune] avoid using inline_tests to be able to specify link flags
Summary:
We want the ability to specify link flags for all our binaries, but
`inline_tests` currently does not support customising link flags:

  https://github.com/ocaml/dune/issues/1530

So, instead, manually build the inline tests runner and use it as part
of `dune build runtest`. This mimicks the previous behaviour. In
particular, if the actual test output is different than the expected
result recorder in the test file, we
1. show the diff to the user, and
2. amend the source file with the new output

The only caveat is that we always do (2) whereas normally we'd do it
only when dune triggers promotion. However, since `make runtest` runs
`dune build runtest --auto-promote` the difference is not visible
unless users start messing with dune directly. (it's possible to account
for this behaviour if needed)

Reviewed By: martintrojer

Differential Revision: D32066074

fbshipit-source-id: 5f11f577d