[paramNotNull] Adding a new checker for blocks that are executed without being checked for null first
Summary:
This adds a new simple checker to handle the null dereference case when an Objective-C block is passed as a parameter to a function and it's not checked for null before being executed in the function.
This is a common pattern in Objective-C code, that we have been supporting with biabduction, and now we want to move it out into its own checker in order to deprecate biabduction.

This adds the first version of the checker and some tests, with handling annotations following in a next diff.

I chose to keep the name of the checker ParameterNotNullChecked in case we want to handle other parameters in a similar way, and the issue type to be specific about blocks.

Reviewed By: skcho

Differential Revision: D31792373

fbshipit-source-id: fbd061b2d