[pulse] detect contradictions on non-integer integer values
Summary:
A known weakness of the arithmetic engine in Pulse is that it treats
values as rationals instead of integers. This mitigates this weakness by
introducing an `is_int()` predicate that gets added on each load and
store of/to a known-integer value. The only rule currently implemented
for it is that `is_int(e)` is `true` if `e` is a known integer literal,
and `false` (contradiction) if `e` is a known non-integer literal.

Reviewed By: da319

Differential Revision: D31932970

fbshipit-source-id: c92cae436