[pulse] a simplex-based algorithm for arithmetic inequalities
Summary:
This follows "Solving Linear Arithmetic Constraints" by Harald Rueß and
Natarajan Shankar (SRI Technical Report, 2004).

The main idea is to represent an inequality `x ≥ 0` as `x = u` with `u`
belonging to a special class of "restricted" (or "slack") variables,
which are always non-negative, and then deal with linear equalities on
restricted variables (the *tableau*) instead of linear inequalities. The
dark magic happens when the tableau needs to be put in a certain form to
yield contradictions.

Implementation wise, the main trick is to have restricted variables be
easily distinguishable by reserving variables represented by negative
integers to be restricted. It was already the case that the normal
(unrestricted) variables were represented using positive integers so
that part doesn't have to change. This interacts nicely with how linear
equalities were implemented provided we choose a judicious order over
variables so that we favour canonicalising variables to restricted ones
one the one hand, but favour solving for unrestricted variables in
linear equalities.

Reviewed By: skcho

Differential Revision: D31966204

fbshipit-source-id: 019e28788