[tests] Move test about ObjC inheritance translated properly from biabduction to Pulse
Summary:
Moving a test from biabduction to pulse. This is just checking that the frontend translates inheritance properly, also Pulse is able to figure out that testFields returns 29.
Moved it to a specialized null_deref folder within pulse, we have so many different error types in pulse now, that a flat namespace will become not enough soon.

Reviewed By: jvillard

Differential Revision: D32206351

fbshipit-source-id: bfb041ad9