Detect whether DBWriter can actually be used
... on systems where it is assumed available.

We have had customers under Linux with DBWriter related errors. Some
filesystems (e.g. ClearCase, or Vagrant synced folders) do not allow
creating a Unix domain socket on them. This patch tries to detect
whether sockets are supported, and falls back to --no-dbwriter
behavior when they are not.

Also catch potential exceptions when stopping the server, in order
to improve reporting to the user.