[tests] [18/n] Moved to own folder and improved retain cycle test
Summary: This is a test about that we treat properly the fields that come from protocols, and take into account their attributes weak and strong. However, there was only a weak case, the test for the strong case was incorrect. So fixed it so that we get a cycle in the strong case and not in the weak case.

Reviewed By: da319

Differential Revision: D32353664

fbshipit-source-id: 1a0927f47