[ConfigImpact] Suppresses changes from methods of gated classes
Summary:
This diff suppresses ungated changes from the methods of the classes that are allocated with gates safely.  For example, if a class is allocated always with gates, it suppresses the changes of which.

```
implementation GatedClass
// some code changes
end

// for all [GatedClass alloc]
if (GK) {
  c = [GatedClass alloc];
}
```

Similar to how we collect config fields, we collect gated classes as follows:

1. gated classes are analyzed intra-procedurally
2. collect all gated classes from all summaries
3. using the collected gated classes, suppress issues when printing the config
impact json result.

Reviewed By: ngorogiannis

Differential Revision: D32358397

fbshipit-source-id: 0699fd68a