[pulse][unnecessary copy] Check whether the source is modified after copy
Summary:
One of the weaknesses of the existing checker was that we were only keeping track of modifications to the copy but not source. However, if we want to suggest that the copy is unnecessary and one can simply add an &, we also need to ensure that the sourec is not modified.

To handle such case,s we extend the map to include an additional map that keeps track of  "source address" -> "copy var".
- when we call a copy ctor, we ensure to record the source of the copy
- when source goes out of scope, we also check modifications to the source, and mark the corresponding copy as modified if so.

Reviewed By: skcho

Differential Revision: D32355105

fbshipit-source-id: 3c26da0e4