[sledge] Fix corner-case variable clash bug in existential freshening
Summary:
Previously the implementation of existential freshening effectively
assumed that the existentials of the body formula were preserved by
freshening the subformulas. However, it is possible for each disjunct
in a disjunction to be freshened to the same formula, leading to the
disjunction being hoisted to the stem, and its existentials being
combined with those of the stem. This case was handled incorrectly
prior to this diff.

Associated changes are:

- Add an omitted inconsistency check during simplification.

- When freshening variables that compare equal, prefer to use those
  that are in the set to be freshened over those that are being
  freshened with respect to.

- Improve replay debugging support.

Differential Revision: D32511365

fbshipit-source-id: 53e46e41a