[tests] Add test about handling methods from categories to Pulse
Summary: This test shows the correct handling of category methods in Infer. We basically treat them as class methods in the frontend.  Also removed the similar test from biabduction, including the frontend test that was not really necessary, as there isn't anything special about this example in the dot files.

Reviewed By: skcho

Differential Revision: D32559657

fbshipit-source-id: e45039f9a