[paramNotNull] Do not consider captured variables to be formals for this checker
Summary:
For Objective-C blocks we add the captured variables to the formals of the block in the frontend. In this checker, however, we need to handle only the real formals, not the captured, to avoid false positives. So here we compute the real formals and use them to compute when a formal has been executed without being checked for null.
Many of the reports that we would remove with this diff are valid to report, but it requires a more complex analysis to get this right that we will do in following diffs. This diff fixes the current inaccuracy and keeps this checker in feature parity to the biabduction checker that it will replace.

Reviewed By: skcho

Differential Revision: D32616186

fbshipit-source-id: d7d3175538