[pulse][objc][nil] add models for some NSString methods
Summary:
Add models for:
- `initWithFormat`
- `stringWithFormat`
- `stringWithString`
- `stringByAppendingString`

They all raise an exception when their `string` (or `format`) argument
is `nil`. The new models check for this and issue a `Null
Dereference` if it is the case. They also enfore their results to be
non-`nil`

This commit also fixes the model for `stringWithUTF8String` to follow
the same rule as above

Reviewed By: dulmarod

Differential Revision: D31930767

fbshipit-source-id: 9e327557e2