[erl-frontend] Remove redundant built-in functions for Erlang
Summary: Certain constructs in Erlang are both expressible with a syntax and library functions (e.g., `M#{K => V}` is equivalent to `maps:update(K, V, M)`. We added buult-in functions to support the syntax, but since then library functions are also supported, so built-in functions became redundant: we can translate the syntax to library function calls.

Reviewed By: skcho

Differential Revision: D32934439

fbshipit-source-id: 51f714e83b