[infer] Move annotations to formals in SIL
Summary: This diff moves annotations of function parameters to `formals` list, so that checkers do not care about mismatched numbers of formals and annotations in its semantics. Especially, it can make logics of manipulating parameters more straightforward, e.g. when adding new parameters in the closure specialization, we had to update the annotation list together.

Reviewed By: dulmarod

Differential Revision: D32993775

fbshipit-source-id: 307482d92e