[pulse] richer branching histories
Summary:
Fix an issue where Pulse could believe there was an invalidation event
in the history of a value when there wasn't, because contextual events
(histories of conditions in conditionals) contained invalidations.

This gives a richer structure to branching histories instead of just
"branching". This could be used later to give more information in traces
too, eg what events in the trace correspond to what, or skip some events
of contextual traces to avoid cluttering the main trace, or ...

Reviewed By: fantazio

Differential Revision: D32960402

fbshipit-source-id: de30a8ea9d