[pulse] continue execution on recoverable errors
Summary:
When reporting any error, Pulse would stop the symbolic execution on
that path. This can hide further errors in some cases, in particular if
the first error is not "fatal". Distinguish between errors that must
abort the program and those we can recover from to cover more code.

This turned out more ambitious than I thought, with a new error monad to
replace the too-simplistic Ok/Error one. Now we distinguish and
accumulate non-fatal errors on a path. At the end of each `exec_instr`
we make sure to report these accumulated errors and start fresh at the
next instr, as before.

Reviewed By: skcho

Differential Revision: D32316980

fbshipit-source-id: 1250b3313f