Exception handling in absint
Summary:
The current absint framework provides a Exception CFG but it
handles exception edges as normal edges. We separate the two notions
when computing the predecessor abstraction of a node and we carefully
turn an exception state into a normal state when reaching a handler.
We activate this framework on Pulse and test it on simple Java resource
open/close patterns relying on try..finally.
We don't handle `catch` very precisely for the moment since we do not record
the thrown exception object.

Reviewed By: skcho

Differential Revision: D33017691

fbshipit-source-id: 1c4759e6d8