[F] Fix strong updates when handling subprogram calls - update FBs baselines.
Improvement, but reveals a bug (not reproducible with Ada frontend)

The problem is that an allocsite representing the content of two different arrays
is strongly updated:

		Arr1 = createArray (length1); – creates an array of length length1
		Arr2 = createArray (length2); – creates an array of length length2

There is one allocsite representing both Arr1 and Arr2 and this allocsite is strongly updated
so Arr1’Length is wrongly assumed to be length2 after the second call.

The problem doesn’t arise for code generated by Ada frontend because there, the allocsite
imported from the callee is further copied in the caller creating new allocsite.

Fix:
- whenever importing an allocsite from callee to caller, add an identification of a caller
  and node in the call node in the caller
- Or just set force_strong_update to false in BufferOverrunAnalysis.instantiate_mem_reachable