[I] Improve type information in the presence of starfield.
Before this commit, non-None type was returned by typ_of_param_path in the
presence of the star field only when the field was a special field
constructed in BufferOverrunField (e.g., a field for C string's length).
Make it possible to use type environment to get the type in the presence
of a starfield the same way as it would be a normal field.