[pulse][retain-cycle] add builtin __objc_get_ref_count
Summary:
The new builtin `__objc_get_ref_count` allows for a better dealloc.
Previously, dealloc used to unconditionally dealloc an object's field.
In addition, we do not handle recursive functions well and in the case
of retain cycles, we would have a dealloc trying to call itself
(potentially indirectly) and therefore havoc its fields' content. This
lead to broken cycles because of wrongful deallocation.

Now, dealloc can decrement the `__infer_model_reference_count` field of
the object before doing all the regular deallocation work if this field
is equal to `0`. This copies objective-c's `release` implicit method
which would call the `dealloc` on the final `release` call
The `__infer_model_reference_count` field is set before any call to
dealloc to be equal to the number of strong references to the object

Reviewed By: dulmarod

Differential Revision: D33257380

fbshipit-source-id: f5ab549d17