Fix cache name
runner.os is actually a very uninformative name: either 'Linux' or 'macOS'.
When ubuntu-latest or macOS-latest changes, the cached binaries become
incompatible with the new OS. But this is not detected by the cache
key we use. Thus, use 'os-version-action' to get this information.

Also, use 'matrix.os' instead of 'runner.os', which is slightly more
precise.  This changes nothing in practice, since Infer is tested on
only one version of each OS for the moment.