Make CapturedVar contain Pvar
Summary:
This is in preparation to add lambda support to Erlang lineage analysis.

Consider the code `f() -> X=1, F=fun() -> X end, F().`. There are two
occurrences of the variable X. Currently, the Erlang frontend translates
both to a Pvar that has the var name "X" and the procname "f". In
addition, it remembers on the anonymous procedure that it captures a
variable "X". Since it remembers that using a `CapturedVar.t`, and since
`CapturedVar.t` used to have "X" but not the procname "f", it was
impossible to recover the name "f" from a procedure's attributes.

Reviewed By: fantazio

Differential Revision: D34080582

fbshipit-source-id: 6de6a6b18b