[erlang] Adjust names of anonymous functions
Summary:
Functions usually have the name module:fun/arity. For anonymous
functions we generate a 'fun' based on the enclosing function
(enclosing_module:enclosing_fun), thus ending up with a ':' in the 'fun'
part, which ca be confused with the separator between 'module' and
'fun'. This diff makes sure that ':' does not occur in the name
generated for an anonymous function.

Reviewed By: mmarescotti

Differential Revision: D34108198

fbshipit-source-id: 31cd174456