WIP Fix cache name in CI (#1594)
Summary:
runner.os is actually a very uninformative name: either 'Linux' or 'macOS'.
When ubuntu-latest or macOS-latest changes, the cached binaries become
incompatible with the new OS. But this is not detected by the cache
key we use. Thus, use 'os-version-action' to get this information.

Also, use 'matrix.os' instead of 'runner.os', which is slightly more
precise.  This changes nothing in practice, since Infer is tested on
only one version of each OS for the moment.

Pull Request resolved: https://github.com/facebook/infer/pull/1594

Reviewed By: skcho

Differential Revision: D34173740

Pulled By: jvillard

fbshipit-source-id: 999f07e31d