Fix dotnet resource leak assign (#1601)
Summary:
Hi,

In this PR, we fixed an issue in assign instruction, in which the alias of resources released expression was not traced to the original expression with resources acquired. Meanwhile, we add a couple of unit tests to validate this change.

Please take a look and let me know if you have any suggestions or questions.

Thanks!

Pull Request resolved: https://github.com/facebook/infer/pull/1601

Reviewed By: skcho

Differential Revision: D34270417

Pulled By: jvillard

fbshipit-source-id: 6a6f778134