[DBWriter] Setup socket before fork
Summary:
This diff setups socket before fork.

problem:

In the previous implementation, the socket setup was done by the server process after the fork. So, the client process had to do polling if the socket is ready to use.

* During the polling, the client checks the socket file exists.
* If the socket file exists, the client assumes that it is ready to communicate and tries handshake with opening the socket.

However, the problem is that the existence of the socket file does not imply the socket is ready to listen.

```
  let setup_socket () =
    if socket_exists () then L.die InternalError "Sqlite write daemon: socket already exists@." ;
    let socket = Unix.socket ~domain:socket_domain ~kind:Unix.SOCK_STREAM ~protocol:0 () in
    in_results_dir ~f:(fun () -> Unix.bind socket ~addr:socket_addr) ;
    Unix.listen socket ~backlog:Config.jobs ;
    socket
```

The socket file is introduced by `Unix.bind` and the socket is ready to open a connection by `Unix.listen`. Therefore, when after `Unix.bind` and before `Unix.listen`, if the client checks the socket file and tries to open a connection, we will see this connection error:

```
Uncaught exception:

  (Unix.Unix_error "Connection refused" connect
   "((addr (ADDR_UNIX sqlite_write_socket)))")
```

Reviewed By: jvillard

Differential Revision: D34309439

fbshipit-source-id: 6468be24db