[pulse] Show all issues in tests
Summary:
In Pulse we suppress nullptr dereference issue if its access trace does not contain invalidation as it usually leads to a confusing error trace.

In tests it might be useful to show the suppressed issues and be able to notice any changes in them. For example, at the moment if we have a false negative we do not know if Pulse does not find the issue at all or is it just being suppressed.

A new flag `pulse-report-issues-for-tests` is added for reporting suppressed issues in the tests and an additional trace element "** SUPPRESSED **" is added to note suppressed issues.

Reviewed By: jvillard

Differential Revision: D34390750

fbshipit-source-id: e95e1606b7