[ConfigImpact] Wrap Field module with a new module LatentConfig
Summary:
This diff introduces a new domain `LatentConfig` that does wrapping of previous `Field` domain.

context: The checker has a mechanism to understand code gated by config fields:

* During visiting/analyzing procedures, it records the code gated by fields.
* After analyzing all procedures, if it found a field may have a config value in any execution path,
  the gated code by the field is suppressed to report.

In this stack of diffs, I would like to apply a similar method, but this time on function
parameters.  Thus, if a config value is given as a function parameter, we suppress the gated code by
the parameter in the function body.

* In this diff, it introduces a new wrapper type LatentConfig.t, which now includes only Field.t.
* In the next diff, it will actually extend the domain/semantics for handling function parameters.

Reviewed By: ngorogiannis

Differential Revision: D34451328

fbshipit-source-id: 9376217efa