[pulse] simplify history contexts in some cases
Summary:
For a sequence of instructions under a conditional, we would generate
redundant InContext elements, eg:

```
if (blah) {
  x := 0
  y := x
  z := y
}
```

gave a history for z with the same context repeated many times:
```
InContext {context=blah; main=Sequence (assigned, InContext {context=blah; main=Sequence (assigned, ...)})}
```

In order to avoid that, make the construction of sequences work a little
harder to detect and avoid this pattern.

This doesn't affect traces because when the trace is created these
duplicate contexts would be detected and dropped already. It was making
debug info hard to read though, and allocating more than necessary (not
sure whether it actually affects perf).

Reviewed By: skcho

Differential Revision: D34447934

fbshipit-source-id: 33c1cc1ba2