[pulse] record potential false negative
Summary:
```
/* This report is currently suppressed because there is no evidence
that p can be NULL, but since it is compared to NULL *in the same
function* it may be worth reporting this. */
void FNsuppressed_no_invalidation_compare_to_NULL_bad() {
  int* p = unknown_int_pointer();
  int x;
  int* q = &x;
  if (p == NULL) {
    q = p;
  }
  *q = 42;
}
```

We don't report issues like this because there's no invalidation
associated with the pointer. This is a good thing whenever the
comparison to NULL is done in a callee as it doesn't mean the programmer
thinks that `p` itself can be NULL, but when it's in the same function
then there's at least inconsistent intent from the programmer (unless
this all comes from the macro !).

Reviewed By: da319

Differential Revision: D34450553

fbshipit-source-id: e25d6abede