[pulse] Handle copies created with multiple states
Summary:
If we have an unnecessary copy created with multiple disjuncts we didn't catch it:
```
  auto copy = get_cond_arr_ref(arr1, arr2, cond); // call to copy ctor, on two disjuncts
```

where `get_cond_arr_ref` returns two disjuncts d1 and d2 where for d1, arr1 points to v1 and for d2, arr2 points to v2. Existing analysis only kept track of a single disjunct d1 for storing the copied heap but it considered all states d1 and d2 at ExitScope and mistakenly thought that copy was modified because it was comparing d2 with d1.

To deal with this, we must keep track of both d1 and d2 for the copy in the non-disj domain. This diff handles that by changing the NonDisjDomain so that it keeps track of a map from copied variable and optional source address to the corresponding heap (e.g. either of d1 or d2) and moves the old SourceMap into attributes. Now, we need to figure out which heap to compare at ExitScope for both 1) source and 2) copy. For 1) we rely on the new attribute `CopiedVar` for the source value in the disjunct. For 2), we rely on a new attribute `SourceOriginOfCopy` for a given copy address that keeps track of the corresponding source address that can be used to lookup up the heap from the NonDisjDomain.

2) means that Pulse attributes depend on the values but sine these attributes don't appear in the summaries, it is ok.

Reviewed By: jvillard

Differential Revision: D34341797

fbshipit-source-id: db9d35031f