Improve evaluation of callee paths repr. ptr derefs in caller.
When evaluating a callee path representing a dereference of a pointer,
create a symbolic value representing the dereference when there
is no or unknown value for it in the abstract state of the caller.

This makes it possible to import, e.g., the fact that "**A_Callee=2" for
a pointer "*A_Caller" having an unknown value in the caller (A_Callee
begin a formal parameter of type pointer to pointer, A_Caller being an actual
parameter of type pointer to pointer). Note that (in-)out parameters are
passed by reference so an (in-)out parameter of a pointer type is indeed
of a type pointer to pointer. Also note that with closures, global variables
are passed as in an in-out mode.

However, the fix doesn't make it possible to import the fact that "*A_Callee=2"
for a pointer "A_Caller" having an unknown value in the caller (A_Callee
and A_Caller being of a pointer type). The reason is that A_Caller is
passed by value, the actual parameter is thus direcly the unknown value
and there is no way to get a location representing A_Caller which is
necessary to create a symbolic value of *A_Caller.