[stats] Display top n procedures with longest analysis duration
Summary:
When analyzing targets that take unexpectedly long time, it is hard to pinpoint which procedures are to blame. This diff aims to keep track of top N procedures that take the longest.

N is set by a config argument `top-longest-proc-duration-size`

We display the top results in a sorted order for better readability.

allow-large-files

Reviewed By: ngorogiannis

Differential Revision: D34581904

fbshipit-source-id: 9056c07528