[pulse] decompile values into source expressions to include in error messages
Summary:
Add a new component to the abductive domain: a decompiler that keeps
track of the most recent way an abstract value was obtained from the
source code. Use it in error messages to describe these values.

The separate decompiler is needed because this information is not easy
to get from the rest of the state, in particular because 1) several
access paths lead to the same value, and 2) we don't even track loads in
value histories.

Reviewed By: skcho

Differential Revision: D34549303

fbshipit-source-id: fb07a2c246