docs: add social banner in support of Ukraine (#1605)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

![image](https://user-images.githubusercontent.com/12485205/156470783-99067198-086d-4c72-9652-34cdf43bfb29.png)

Pull Request resolved: https://github.com/facebook/infer/pull/1605

Differential Revision: D34597398

Pulled By: dmitryvinn-fb

fbshipit-source-id: 38988085e7