[topl] Make it easier to spot bad properties
Summary:
It often happens to me that I get the static pattern of a transition
wrong. Then it never matches, and it is difficult to figure out that the
wrong static pattern is the reason of the false negative. With this
commit, if you use `--topl-trace`, you get a user warning that lists all
those transitions whose static pattern never matches. (The dynamic part,
which looks at values, may still not match. But, at least you know you
didn't make some silly mistake such as misspelling a method name.)

Differential Revision: D34481028

fbshipit-source-id: 77e540517e