[ConfigImpact] Ignore source code in test paths
Summary:
This diff adds an option `--config-impact-test-paths`, so that the checker can ignore code in the
paths.

context: When detecting a gated class, the checker inspects all allocation sites of the class are
gated.  However, sometimes test files under test paths are using the class without a gate, which
introduced false positives before.

Reviewed By: ezgicicek

Differential Revision: D34786928

fbshipit-source-id: 10c0c8462ab13f6edfa81bbbc62251518a1fca19