[logging][bugfix] Flush formatters on exit
Summary:
In `Logging`, the `close_logs` function is registered as epilogue and
goes through the `!logging_formatters` list to flush and close. However,
that list was left empty by the call to `reset_formatters`. (In fact,
even if it were not left empty, the list contained `formatters` and
`formatters ref` that were inconsistent: one was used for writing
(`formatters ref`) and the other for flushing (`formatters`).)

There are also some simplifications of the code afforded by using
`formatter_of_out_functions`.

Reviewed By: ezgicicek

Differential Revision: D34641885

fbshipit-source-id: ada82003c2a127a27eb248967b4961ca58d9de24