[erl-frontend] Parse and store function specs with overloads
Summary:
Previously we only supported function specs that had a single list of arguments and a single return value. Now we support multiple of these, which in [Erlang's terminology is overloading](https://www.erlang.org/doc/reference_manual/typespec.html#specifications-for-functions), for example:
```
-spec f
    (atom()) -> integer();
    (list()) -> binary().
```
This reads as "f either takes an atom and returns an integer OR takes a list and returns a binary". Hence, I named them disjuncts in our internal representation.

Reviewed By: mmarescotti

Differential Revision: D34926599

fbshipit-source-id: 660a1f4b027b87c9505515099535bde18d11e79b