[lineage] Add location on edges
Summary:
For call instructions, neither the source nor the target of an edge has
a location corresponding to the call instruction. (This is in contrast
with, say, assignment instructions, where the location of the target is
the same as the location of the assignment instruction.) So, we add
locations on edges.

Differential Revision: D34862164

fbshipit-source-id: 896e5cd7bca879391459d5ddad23ca8582290c0b