[erl-frontend] Add support for user-defined types in type assumptions
Summary: Each user defined type is translated to a function with one argument. When called, it returns true if and only if the argument has the specified type. For example, `-type my_type() :: atom().` results in a function `my_type(X)` which returns true if and only if `X` is an atom. When we see a user defined type, we call this function. We don't inline this check because it might be called from different modules (remote types).

Reviewed By: rgrig

Differential Revision: D35113662

fbshipit-source-id: 7300a7bd90f5f58f15adce587b709f4ed3a8082d