[pulse] configurable taint kinds and policies
Summary:
This is going to change a bit in the next diff but this is still a
valuable step: introduce user-configurable "policies": which sources
flowing to which sinks are errors.

Because kinds are user-defined they cannot be a(n efficient to
represent at runtime) variant, they need to be a string. Pre-emptively,
and possibly needlessly, avoid bloating the state with too many
instances of equal strings by hashconsing them. It's not hard to do any
way.

Examples to come in a next diff.

Reviewed By: skcho

Differential Revision: D35317599

fbshipit-source-id: 429b6572f36ac7dd331673f16788e64f43d97489