[sledge] Improve reporting of failure to parse trace specs
Summary:
If a parsing function passed to Core.Command.Arg_type.create raises an
exception, indicating command line argument parsing failure, the
exception is converted to an sexp, then to a string, and then reported
to the user. This diff adds a Trace.Parse_failure exception, makes the
error messages more informative, and registers a function that will be
used to convert these exceptions to sexps when reporting command line
parse failures.

Reviewed By: bennostein

Differential Revision: D35500539

fbshipit-source-id: 708ffe91a72d8a25c7a2bd2be1ebf6aa297acc77