[pulse] import quandary/Files.java test
Summary:
We have FNs when arguments are passed to functions with a variable
number of arguments, indirectly because we don't do anything for arrays
yet.

Also, unlike Quandary the default taint doesn't interact with all other
taint types so change `inferSecretSource` to also be of a kind that
interacts badly with Files to get reports in the test file.

Reviewed By: skcho

Differential Revision: D35394366

fbshipit-source-id: 09ffba7ea85f8c9bab707af02a13bafa30f986c6