[Pulse][Resources] Fixing Map models wrt reachability
Summary:
The resource analysis relies on an abstract memory graph traversal in order to
approximate the unreachable adresses at the end of each methods.
The current `java.util.Collection` models is a recency abstraction that forget some elements
of the current collection. These elements must however considered as always reachable.
We mark them with a special new attribute and make sure to consider all reachable adresses
from them to be reachable too.

Reviewed By: skcho

Differential Revision: D35494804

fbshipit-source-id: 0b6842c000cb1aa7bd70db0e258ea4b49d10a366