[pulse] intra-procedural-only tainted formals, import quandary/TaintedFormals.java test
Summary:
The semantics of "tainted formals" was wrong, as this test revealed:
when tainting formals of the current procedure before analysing it, we
shouldn't propagate the taint to the procedure's summary to avoid double
reporting. We're just making (pessimistic) assumptions about the taint
of values that will be passed to the procedure, but if we're in a caller
we should just check their actual taint.

Reviewed By: fantazio

Differential Revision: D35434368

fbshipit-source-id: 6c8c19312afc8a1247949237bc2a154801057d4c