[pulse] propagate taint across unknown calls, import quandary/UnknownCode.java test
Summary:
Add a new attribute to track taint dependencies introduced by calls to
unknown functions so that we remember them inter-procedurally.
Previously, we would only propaget whatever taint was *already there*
when we do the unknown call, but this does not work if the unknown call
is inside a callee and the taint is only at the call site of said callee.

When checking that a value is not tainted (because we're reached a
sink), we now also check that values *transitively* reachable via this
new attribute are not tainted. We also record MustNotBeTainted for each
such value as otherwise the fact that we checked them for taint may also
get lost at function boundaries (eg the value that reached the sink does
not live long enough to be in the summary but values that it depended on
for taint are in the summary and so should be recorded as
MustNotBeTainted).

This introduces a risk of duplicate reports when seeing a sink, as many
dependent values may be tainted. To avoid that, remember which policy
violations have been reported to report at most once per policy, which
seems sensible.

Next diff cleans up these PropagateTaintFrom attributes so as not to
remember values that do not appear in the summary.

Reviewed By: ezgicicek

Differential Revision: D35496293

fbshipit-source-id: ceb3317898a9bf1129f32f3905cfc260e74b0e0f