[frontend][pulse] Non-static const member functions should have *this as const ptr
Summary:
Non-static const member functions cannot modify or call non-const functions on object parameter `*this`. Hence, the type of this should be a pointer to const. I am not sure why clang doesn't have the correct type so I fix them manually.

For instance, from [the documentation](https://en.cppreference.com/w/cpp/language/member_functions):

```
struct Array
{
    std::vector<int> data;
    Array(int sz) : data(sz) {}

    // const member function
    int operator[](int idx) const
    {                     // the this pointer has type const Array*
        return data[idx]; // transformed to (*this).data[idx];
    }

    // non-const member function
    int& operator[](int idx)
    {                     // the this pointer has type Array*
        return data[idx]; // transformed to (*this).data[idx]
    }
};

```

This change has implications on Pulse results: when a formal is not a pointer to const, the actual is havoc'ed. Now, by making more things as pointer to cost, we will be havoc'ing less. There is a test change but I didn't understand exactly how it became Latent.

This diff will actually be useful on the next diff where we add `WrittenTo` attribute to all unknown functions that have non-const-ptr formals.

Reviewed By: skcho

Differential Revision: D35462604

fbshipit-source-id: 33a83a2b541c1af7ebc19f7169bbbb6f80a2b83d