[pulse] Add WrittenTo attribute to unknown functions with non-const-ptr type
Summary:
We were getting lots of false-positives because analysis didn't consider unknown functions to be potentially modifying the copy arguments.

This diff partially handles this issue by adding `WrittenTo` attribute to actuals (+reachables from it) that have a corresponding non-const-ptr formal type. If they have const ptr type, they cannot be modified anyways.

There are some false-negatives since we cannot distinguish between library calls that should not modify *this (unless the argument is already const& resolved by dispatcher).

There is also another identified FP due to infer not recognizing the type of iterators as pointers. This will be handled in the next diff.

Reviewed By: skcho

Differential Revision: D35394458

fbshipit-source-id: f630456b0c1096ec5670676ee741f4bdbf7d636b