[pulse][block] capture as much as possible
Summary:
When specializing at pulse-time, we used to only go one level deep into
the captured vars: specialize over the blocks in the captured vars and
capture what they need. However, in some cases we could need to go
deeper than one level deep (as in the new `test_localization_bad`)
because the captured block may capture another block itself and so on...
The only way we had to specialize over all those blocks would have been
to retrigger the specialization until everything has been accounted for.

Now, the analysis-time preparation work for the specialization looks as
deep as possible in the captured variables. From the information found
it creates a map similar to what the specialization creates which
associates a pvar to the corresponding `passed_block` information. This
`extra_formals_to_blocks` map is what enables to account for all
captured variables at any depth without having to re-specialize. Because
the specialization knows how to manipulate this kind of data, it will be
able to know and account for all the blocks in this map

Reviewed By: skcho

Differential Revision: D35281769

fbshipit-source-id: 14d61ca29c4b3d1125af276b69728560b841dcdb