[erl-frontend] Translate spec-only functions
Summary:
For functions that only have a spec, but no body, create a procedure that returns an unknown value, but pruned based on the type. This is mostly useful for OTP functions where we don't want to analyze the body, just have some assumptions on their return value.

For example `lists:sort/1` has the spec
```
-spec sort(L) -> L when L :: list().
```
which results in a procedure creating a fresh value and pruning that it is either Cons or Nil:
```
R := fresh_value();
T := (R is Cons) or (R is Nil);
prune(T is true);
return R;
```

Reviewed By: rgrig

Differential Revision: D35849056

fbshipit-source-id: fe9334644b1abd8a370906be6c67c3d459c9b1af