Improve the equality model to make it more precise on non-integers
Summary:
The recent equality model worked well for integers, but fell short for
comparisons involving other types, especially comparing integers
vs. for instance atoms. One identified issue was that, by returning
a disjunct list of cases split according to the types of the compared
values, the model was quickly hitting the maximal disjunction number,
leading to a lot of precision loss.

We improve the model by building instead a disjunction in one single
abstract state, by the use of logical operators. This disjunction
encodes the dispatching of specific comparison operators according to
the types of the compared values.

For now, this disjunction is still built manually and only considers
each value being an integer or not (leading to 4 cases: integer value
equality, always false type-incompatible equality (x2), and unknown
result for unsupported types equality). A future work should involve
building it programatically with respect to all the types that we want
to support. We tried to structure the code as clearly as possible to
make this improvement easier.

We introduce a new structure for defining the comparators on known
types, based on objects to pave the way for an easy extension on other
comparison operators and/or other types. A comparator, corresponding
to a high-level comparison operator (equal, lower, ...), is an object
that defines several methods to compare values that have a known
type (for instance, the method `#integer` will be used for comparing
integers). It also defines an `#unsupported` method for comparisons on
unsupported types (currently, everything but integers) and an
`#incompatible` method for comparison between a supported type and an
unsupported one (which for equality builds the false constant).

Extending the model to other comparisons will involve defining new
comparators (and the corresponding one-line functions that use them
the the model and the pruning version). Extending the model to another
supported type will involve extending the global disjunction and add
a method to the comparators to support this type.

We improve a bit the coverage of tests on int equality, and add some
tests involving atoms to showcase unsupported and incompatible
comparisons.

Differential Revision: D35613159

fbshipit-source-id: 84e59eab1b2a85e2dee9ebb7b23bc99682543b00