[pulse][decompiler] Improve error messages for ObjectiveC getters
Summary: We access a property in Objective-C through a getter, e.g. for a property `x_ptr`, `obj.x_ptr` is translated to a function call `x_ptr(obj)`. Since `x_ptr` is a function, the error message only mentions the function name `x_ptr` (with an Interface in which the property is defined). Instead, we want to improve the error message to include the object which property we are accessing.

Reviewed By: dulmarod

Differential Revision: D35896348

fbshipit-source-id: 6dfd9a979f302aaca1ff2018486b510b42989a41