[topl] When hitting limits, try harder to keep satisfiable
Summary:
Previously, the expensive check for satisfiability was used only when
extracting summaries. This led to a huge number of unsatisfiable states
being explored, leading to hitting the limit on disjuncts, which in turn
led to dropping many disjuncts. Now, the expensive check for
satisfiability kicks in whenever Topl is about to start dropping
disjuncts.

Also, the expensive simplification now takes into account the dynamic
type, which prunes more infeasible paths.

Differential Revision: D34686762

fbshipit-source-id: 7896279eaa615b452e735e32c866f18fa0b66ddc