[topl] Drop states that refer to garbage
Summary:
When the register of the Topl monitor refer to values that became
unreachable in the program state, it is usually impossible to reach an
error later on. So, as a heuristic, we drop tracking of these monitor
states. This is done only when we hit the limit of how many Topl states
can be tracked in parallel.

Differential Revision: D34819839

fbshipit-source-id: 65ddf98ec03f18d3f675311f9f35d0fb1d2737df