Make erlang qualified dynamic calls explicit and support all dynamic calls in lineage
Summary:
Instead of being marked as "unsupported", qualified erlang dynamic
calls are now translated to a special infer procedure that is similar
to erlang:apply but takes the original erlang parameters one-by-one
instead of a list.

This makes lineage results clearer for this kind of call.

We also provide an unqualified version of this procedure, which
Lineage uses to translate unqualified dynamic calls.

In the end, on all dynamic calls, lineage now shows results similar to
what would haev been obtained would the code being written with
erlang:apply. The main difference is that parameters are given
one-by-one, which makes the final result more precise than putting
them all in an erlang list (and is the reason why we need two
different special names, since otherwise there would be no way to
differentiate unqualified vs. qualified calls).

Reviewed By: rgrig

Differential Revision: D35818126

fbshipit-source-id: 9446640df7e802b04bb05a7e9e6228e888d92ff2