[objc][block] store orig pdesc copy for specialized version
Summary:
We used to store an empty pdesc for block-specialized function at first
and then never update the stored pdesc to the specialized one because we
don't need to store the specialized one. However, when retrieving the
pdesc from the DB, we don;t want to get an empty one either (this would
cause issues with the upcoming alias-specialization). Instead, we can
store a direct copy of the original pdesc with the specialized attribute
allowing the retreived information from the DB to be useful on its own
and not requiring further investigation into the original pdesc

Reviewed By: dulmarod

Differential Revision: D36137434

fbshipit-source-id: 6c2e172b0f5223a2391a2827116400eb52ceb1d6