[pulse][alias] preparation work for alias specialization
Summary:
When calling a function with the same pointer for 2 different arguments,
we currently raise an Aliasing contradiction and fail to call the
function. This is because the callee's pre expect the arguments to be
distinct by default. We do not want to change that default behavior.
However, we can create a specialized version of the callee with a pre
that expects arguments to be aliases when they are given the same value
in the caller. This work identifies such calls that need alias
specialization

Reviewed By: dulmarod

Differential Revision: D35814019

fbshipit-source-id: e717ca0359da0c69be7ca8cad4ee30f458e8a146