[pulse][taint] Report all policy violations, not just first
Summary:
When checking potential taint flows against policies, do not stop after
encountering a violation.
Instead, continue and gather all violations (without duplicating) and their
associated errors.
Also add a simple test demonstrating this behaviour.

This change is in preparation for reporting all data flows to a sink -
for these, we do not want to stop until all flows are found.
Therefore, this commit brings flows from known/specified sources/sinks
in line with this expected change.
In particular, it will allow the expected change to build on the
existing code more easily.

Reviewed By: skcho

Differential Revision: D36775289

fbshipit-source-id: df49d13d41d4a0f655b884c2bb2f2fca8024b80c