[pulse] Add an allocation attribute for Objective-C
Summary:
Add a constructor to the `PulseAttribute.allocator` type, intended to
represent an Objective-C `alloc` call.
Ensure that the Pulse model assigns this attribute on encountering such
an `alloc`, but doesn't report memory leaks for it.
In particular, as Objective-C is memory managed we do not want to report memory leaks - so whenever this new attribute is found in the Pulse state, it is considered already-freed by the memory leak analysis.
Therefore, this attribute is just a marker indicating data flow of allocated objects - the intention is to use it to examine which allocated objects flow into taint sinks.

Reviewed By: dulmarod

Differential Revision: D36810250

fbshipit-source-id: e8234e8ba00556ede84601440d905a41b4ad0f06