[racerd] fix FPs from [[noreturn]] function calls
Summary:
In C++ calls to `[[noreturn]]` functions result in a successor edge to the exit node from the call. This means that the post of the call will be joined to all other posts at the exit node.
This causes FPs in RacerD because if a lock is taken before the call, the summary will record that the containing function takes but does not release the lock at exit.

This diff adds information to the abstract state to interprocedurally track the inevitability of calls to `[[noreturn]]` functions. When such a call occurs, the state is replaced by (an approximation of) the identity element in the abstract domain, so that the join at the exit node leaves other posts unaffected.

Reviewed By: da319

Differential Revision: D36479165

fbshipit-source-id: 81411e5ea7dc47dd613e85e6579f24ab5ca8d5af