[pulse][taint] Report some data flows from taint sources
Summary:
Instead of reporting a "sensitive data flow" advisory alongside taint errors, report one for every path which is
discovered from a taint source.

This requires changing the `destination` field of the `FlowFromTaintSource` diagnostic to a procedure name.
This is all the information we have on the destination of a generic flow from a taint source for now.

Also add a very basic test (`SensitiveDataFlow.m`) which creates a tainted object and passes it through a few dummy
functions, before passing it to a final function intended to emulate a sink (`might_be_a_sink`).
In the `basics.m` test, also update the `callNonSinkOnSource` test to be
named `Bad` rather than `Ok`: it correctly reports a `SensitiveDataFlow` from a source to a non-sink.

The implementation does not yet handle sanitizers or taint propagation (`PropagateTaintFrom`, e.g. for arrays), which will be tackled in future commits.
More importantly, it does not yet satisfactorily handle interprocedural flows.
In particular, in the test we would like flows from the taint source to be reported all the way to `might_consume` -
however, they are only reported to `mutates_then_might_consume`.
The deficiency here can be most clearly seen by comparison to regular taint errors - an interprocedural flow from a
taint source to a taint sink is found by determining:
   - which tainted values flow into the call by finding `Tainted` attributes in the postcondition of the caller's current state
  - which sinks can be reached during the call by finding `MustNotBeTainted` attributes in the precondition of the callee

Putting these two together gives the bad flows.
By contrast, we have no sink information as we are trying to determine all flows of tainted data, not just those to sinks.
Therefore this first attempt cannot extract sufficient information from the callee's precondition - further thought is required.

Reviewed By: skcho

Differential Revision: D37002325

fbshipit-source-id: d1004962ce00db4a352dc4f53da916ceb057e8db