[cleanup] Move away from block-specific vocabulary when talking about any kind of function variables
Summary:
When developping function specialization over blocks, all the names and
documentation was directly relating to them. Now that the specialization
has been widen to account for function pointers too, the code must avoid
misguiding developers into thinking it is only accounting for blocks and
use a more generic vocabulary. The chosen names are closures and
functions depending on the context

Reviewed By: da319

Differential Revision: D36893501

fbshipit-source-id: dad5aced0628abfa2462314e9b247befccb52297