Faster erlang model for equality by using dynamic type attributes
Summary:
The previous erlang model for equality was building and evaluating
a SIL tree of logical Or operators, that encoded a matching on the
erlang type of values being compared.

We instead leverage `get_dynamic_type` to get this information
"statically" and dispatch to the correct operation at the OCaml level.

On the plus side, the performance of the model is much better, and the
simpler code will make it very easy to add support for other type
comparisons.

However, we do lose a bit of precision on some cases where the
comparison is done within a function that does not have this
information, for instance on the following code:

```
is_zero(X) -> X == 0.

true_negative() -> true = (0 == 0).

false_positive() -> true = is_zero(0).
```

`is_zero`, not having any type information on `X`, cannot use the
integer comparison and therefore has an unknown result in its
summary. Applying this summary on integers doesn't help since the
imprecise choice has already been made (whereas the Or tree made the
previous version have several summaries, including a precise one for
integers). Here, a type spec should solve this.

This also happens when matching against maps, since that involves
invoking equality on the keys and values. However here type specs
currently won't help because we are not able to express the types of
the key and value.

If this becomes an issue, a fix could be to fall back on the old
Or-tree behaviour when some dynamic types are unknown. However the
performance impact should be carefully evaluated.

Differential Revision: D37009049

fbshipit-source-id: 3c53589c9baa8ca750a79dadda8e00619f3201b2