[sledge] Statically compute candidate targets of indirect calls
Summary:
During LLVM translation, keep track of any functions that appear as an rvalue in the program.  Then, tag indirect callsites with the
subset of such functions whose type is compatible with the callsite.  These
"candidate" callees can then be used to approximate dynamic interprocedural
control flow when we compute distance heuristics for goal-directed analysis.

Reviewed By: jberdine

Differential Revision: D37071436

fbshipit-source-id: c8ebad3ee7109832f01b9be6fc9310b2f845c3c3