Split Any cases wrt. integers on Erlang comparison model
Summary:
Rather than using the unsupported comparison, when one value compared is known to be an integer and
one has an unknown type, we build a disjunction that encodes a case analysis on the type of the
unknown value. The integer comparison is used on the case that it is indeed an integer, and the
incompatible comparison is used on the other case.

This yields more precise results, especially through function calls where the type of a parameter of
the function may be lost before it is compared to another known integer value.

Differential Revision: D37241554

fbshipit-source-id: 381cfdf935e7ff68aa56f2ec265bf0c53b38accf