[sledge] Stop at unresolvable indirect calls consistently
Summary:
Currently, `-trace Stop` causes the analysis to stop when any of the
listed error conditions are met except for `on_unknown_call`, which requires
that `Report.unknown_call` also be traced due to the `Dbg.kprintf` there.

This diff makes the behavior consistent across the different error
conditions, so that `-trace Stop.on_unknown_call` does what it says on the tin.

Reviewed By: jberdine

Differential Revision: D37342989

fbshipit-source-id: 7e682cb7f624d089dd2c05b691b8b6fe77c49d60