[build] build a more minimal subset of clang-ocaml
Summary:
There is an issue with building the clang plugin when the version of
OCaml has changed as the build will fail instead of simply rebuilding
the OCaml artifacts. However, we do not even need to build anything in
facebook-clang-plugins/clang-ocaml/ to build infer so we can sidestep
the problem.

The problem remains when running `make clang_plugin_test`.

Also delete an unused variable that made it seem like we cared about
more files in clang-ocaml/ to build infer, and fix a random typo.

Reviewed By: ngorogiannis

Differential Revision: D37348105

fbshipit-source-id: 17a2c75696d01c9c8ef3aa3e41b603cb64c84fa6