Add jar scanner prototype
Summary:
Given a jar file, `scan_jar` outputs the amount of classes and an estimate of the lines of code of the original source. These statistics will be needed when the analysis is performed.

Jar files contain a table that maps each bytecode line to its corresponding source line (for each method): we use this table to obtain the source lines. The mapping is not perfect, therefore one can only estimate the source lines.

Reviewed By: davidpichardie

Differential Revision: D37382117

fbshipit-source-id: e387786487c6361a1798f2788a06583aa925851f