[pulse][taint] Do not report sensitive data flows to certain procedures
Summary:
Do not mark arguments of certain procedures as taint procedures - these are:
- built-ins (according to `BuiltinDecl.is_declared`)
- `dealloc` (according to `Procname.is_objc_dealloc`)

This avoids noisy reports of sensitive data flows that are of no use to the user.

The predicate for testing procedures is `PulseTaintOperations.should_ignore_sensitive_data_flows_to`.
It is expected that this predicate will evolve in future commits after further testing/usage.

Reviewed By: dulmarod

Differential Revision: D37382948

fbshipit-source-id: d2aae89e54a862a81a9e966cee56fcd7a3eacafb