Replace LOr with proper disjunction in comparison models wrt. integer
Summary:
When comparing a value of Any type against an integer, we build a disjunction list of the cases where
the value is an integer or not, instead of computing a LOr disjunction. As the list only contains two
possible cases, this should not lead to too many disjuncts, and is expected to yield more precision
in future developments of comparison models.

Reviewed By: rgrig

Differential Revision: D37525235

fbshipit-source-id: f5d25d48e1c826a4a5e944fffee9e3e0f4ed2aa9