[pulse][taint] Taint object fields when an object is tainted
Summary:
When we have a tainted object, we would like all its fields to be considered tainted too.
Therefore, when we taint a value, traverse memory to find the other values it points to, and taint these as well.

This also fixes a false negative in `cpp/pulse/taint/basics.cpp`.

Reviewed By: dulmarod

Differential Revision: D37555714

fbshipit-source-id: aaf0394810d9f65941656bf0ae2f876b94d3bacc