[erl-lineage] Output edges that differ only in location
Summary:
For code like `f(X) -> g(X), g(X).` there are two edges from f.X to g.X,
one annotated with the location of the first call and one annotated with
the location of the second call. We want both, but previously only one
was output in the JSON report (although both exist in the internal
summary).

Differential Revision:
D36816682

LaMa Project: L1122176

fbshipit-source-id: ed74e1d5123efb7069366d75c4f9f6008b7f3172