[pulse] Add unknown function calls to the decompiler
Summary:
When a known function call is analysed, the decompiler is augmented with information about its return value.
Therefore, when error-reporting looks up the return value, it is given a helpful pointer to the function which
returned the value.

However, for unknown function calls the decompiler is not augmented in this way, leading to `UNKNOWN` results on looking
up return values. This commit attempts to augment the decompiler in the natural place.

Reviewed By: dulmarod

Differential Revision: D37650616

fbshipit-source-id: aecaaacaca3d070c987050ed9783ed6e079f32c8