[reanalyze] Delete summaries to reanalyze from DB/cache
Summary:
This diff deletes the summaries to reanalyze from DB/cache.

Before https://github.com/facebook/infer/commit/6b671c :
When a procedure was chosen to reanalyze, it set a dummy summary to the cache with `status=Pending`, which triggered a new analysis without changing the DB file. However, by the recent change, the `status` field in `Summary.t` was removed, so we cannot use that way.

Reviewed By: ngorogiannis

Differential Revision: D37750015

fbshipit-source-id: 79025d6af5010b1cd2433a79595f837fbb507817