[sledge] Add variable context and fresh variable generation modules
Summary:
Add Var.Context, a module to be used as a uniform representation for a
vocabulary and quantifier prefix across several layers of the code
base.

Also add Var.Fresh which manages generaing variables fresh with
respect to a Var.Context, to be used to unify computations that
generate fresh variables.

Reviewed By: bennostein

Differential Revision: D37593400

fbshipit-source-id: f9e442d2279c30928fb711259c1186f5dd187cb4