[sledge] Change Exec to use Var.Fresh and quantifier-free Sh
Summary:
Symbolic execution explicitly manages the vocabulary and universal
quantifier prefix for generic command triples. It is not necessary to
allow and support quantified formulas as triple pre- or
post-conditions. So change triples in Exec to use quantifier-free
symbolic heap formulas.

Also change the variable context manipulation to use Var.Fresh. This
unifies the treatment of fresh variables generated as part of symbolic
execution with those generated by formula manipulation.

Reviewed By: bennostein

Differential Revision: D37593383

fbshipit-source-id: 920274f8bcb8d0e6060e7448d51c26bb1abe9c5d