[sledge] Change representation of vocabulary from set to int
Summary:
The necessary operations on the vocabulary do not obviously justify
the full set structure of the current vocabulary represenation. So
switch to using just an integer that bounds the identity of the
variables in the vocabulary, essentially representing the full set of
variables with identity less than the bound.

Reviewed By: jvillard

Differential Revision: D37593396

fbshipit-source-id: 604d34921966b8c2a9d6a6d16b40dad9980d7752