[sledge] Vendor iter library
Summary:
The current approach of having a wrapper library can be confusing and
leads to poor documentation and merlin interaction. An alternative is
to vendor the source of the library and adjust and extend its API
directly, which leads to simpler code with better documentation and
merlin integration.

Source and license for iter version 1.5 from
https://github.com/c-cube/iter/tree/v1.5

Reviewed By: bennostein

Differential Revision: D37743446

fbshipit-source-id: d4c5578a94a91cc0d621022a26b56b61f58ea69c