[sledge] Directly correlate variables and registers
Summary:
During symbolic execution, logical variables need to be created to
denote the value of program registers. In the presence of concurrency,
each thread has its own registers, sharing the same register names
across threads.

Currently the association (environment) between registers, threads and
variables is done in a stateful and unpredictable way. This makes
mapping the variables in symbolic states back to progam registers nigh
impossible.

Here we change this to make the identity of variables a function of
register id and thread id. This is done with some bit packing, and
assumes that programs will be small enough that all register ids will
fit in 48 bis, and that we will not need more than 14 bits to
represent the number of threads during symbolic execution.

Reviewed By: bennostein

Differential Revision: D37834860

fbshipit-source-id: 53e49543ce9ebf39c80ebad4afb4c5b092b1210b