[pulse][taint] Fix a hack avoiding duplicate taint errors for a single policy
Summary:
We used a hack to compare taint policies to ensure that we do not report multiple violations of the same policy during
the same call. A comment in `PulseTaintOperations.ml` documented this and suggested an improvement - this commit implements
the improvement. In particular, policies are given IDs, which are used to check equality (instead of using pointer equality).

The hack is in fact worse than it initially seemed: after the introduction of attributes with multiple sources/sinks,
it leads to unexpected behaviour. Consider the following situations:
```
policy1: SourceA or SourceB --> Sink   |   policy2: Source --> SinkA or SinkB
                                       |
obj = sourceA+B();                     |   obj = source();
sink(obj);                             |   sinkA+B(obj);
```
LHS produces a single taint error as desired: we do not want to report multiple violations of `policy1` at the same callsite.
However, RHS produces two taint errors, because policies are stored in a sink-biased way: a hashtable maps sink
kinds to corresponding source/sanitiser kinds. Therefore `policy2` is in fact stored as two sink-biased policies:
`SinkA <-- Source`, and `SinkB <-- Source`. These are non-equal, and so the reports are considered distinct.
This is unexpected behaviour for users, who should not have to consider the internal sink-bias of our taint analysis.
The changes in this commit also fix this issue, by assigning IDs based on user-defined policies (not sink-biased ones).

At the same time, we take the suggested minor performance enhancement of tracking of a set of policy IDs rather than a list.

Reviewed By: jvillard

Differential Revision: D37998822

fbshipit-source-id: e612618b7a9dac9cf885e6a8e656038def876662