[sledge] Substitute functions with model functions
Summary:
The goal of this diff is to allow modeling functions in code.
Modified the front-end to look for functions with the `__sledge_ prefix`.
If `__sledge_foo` exists then foo will not be translated. A use case for
this would be a function crashing the frontend's translation phase.
On the other hand `__sledge_foo` would be translated into llair as `foo`,  which essentially achieves the replacement. Also, any reference to `__sledge_foo` in the code will be replaced by `foo` (to allow, for example, referring to model functions in their own implementation).
Right now, the code does not check whether foo and `__sledge_foo` are type-compatible so care must be taken to ensure this.
A new test `analyze/model_substitution.c` tests substitution.

Reviewed By: bennostein

Differential Revision: D37487167

fbshipit-source-id: d9025e07ed05bfa924f2b7c8d3135d5f3a793b31