[resource][java] Use UnknownEffect attribute to remove Allocated attribute
Summary:
When an unknown call is performed in Java, we do not havoc
but just remove the `Allocated` attribute at call site. But to make the
analysis work inter-procedurally we also add a `UnknownEffect` attribute to
inform the caller.

In this diff, we adapt the caller side to make sure it will use the
`UnknownEffect` attribute in an appropriate manner.

Reviewed By: jvillard

Differential Revision: D38098001

fbshipit-source-id: 04adb2c3d5a11fd5e1ecbbce8a1d0de3d9edfa89