[pulse][taint] Improve "data flow only" taint specification
Summary:
Users can specify a taint source as "data flow only", i.e.  data-flow-to-sink advisories are reported instead of regular
taint errors when data flows from the source to a specified sink.
However, as taint sources are linked directly with procedure names, there is a shortcoming: to specify a given procedure as both a "data flow only" source and a regular taint source, taint config files would have to specify it twice. This could easily lead to copy/paste errors and so on. Worse, it results in multiple entries into the attribute tracking sets of taint sources (`Tainted`) for effectively the same taint source.

This commit improves the situation by associating the "data flow only" property with taint *kinds*, not with the procedures
themselves. Users can therefore specify a taint source as having a regular kind and a data-flow-only kind, and so be able
to report both taint errors and data flows to sinks. E.g.:
```
  taint source: some_procedure()
  kinds: DataFlowToSinkSource (data flow only), SimpleSource
```
The trade-off here is that now a single kind cannot be used for both taint errors and data flows. However, it is much easier/more readable to specify multiple kinds for a procedure than it is to specify multiple entries for a single procedure. It is also more user-friendly: kinds can be used to clearly demarcate data flow reporting vs. regular taint reporting.

Reviewed By: dulmarod

Differential Revision: D38038279

fbshipit-source-id: 928dcc8eaf24e2c5353947d2fc25b9ce84744030