[pulse][taint] Differentiate between block/non-block taint configuration matchers
Summary:
Currently, Pulse's internal representation of the "name" of an Objective-C block is some variation on the name of the
function in which the block is defined. This causes issues for users who want to specify the enclosing function as a taint
sources/sink/sanitizer, but not the enclosed blocks.

This commit gives users the option to specify a `match_objc_blocks` boolean in each matcher (false by default).
The matcher will identify Objective-C blocks if and only `match_objc_blocks` is true.

*It is expected that this functionality will be removed as taint specification of blocks improves - my understanding is that this work is planned as currently the specification exposes Pulse's inner workings.*

Reviewed By: dulmarod

Differential Revision: D38242291

fbshipit-source-id: 3d9fa8b8751abe62470853a8b396e2f87f8a4d57