[pulse] get rid of "known" formula, revisit latent issues
Summary:
This does several things at once because splitting them up would
temporarily make Pulse get out of a local optimum and report worse
results. This change improves the performance of pulse quite a bit,
measurements are included further up the stack as there are other
improvements in the following diffs.

The main change is how we handle latent issues wrt the arithmetic
domain. Before, the arithmetic facts were split in 3: known, pruned, and
both:

```
(** Instead of a single formula, distinguish what we have observed to be true (coming from
    assignments) from what we have assumed to be true (coming from conditionals). This lets us delay
    reporting certain errors until no assumptions are needed (i.e. the issue is certain to arise
    regardless of context). *)
type t =
  { known: Formula.t  (** all the things we know to be true for sure *)
  ; pruned: Atom.Set.t  (** collection of conditions that have to be true along the path *)
  ; both: Formula.t  (** [both = known ∧ pruned], allows us to detect contradictions *) }
```

The "known" part comes from assignments, the "pruned" part from
conditionals:

```
x = 1;  // add v_x = 1 to known and both
if (y > 5) { // add v_y > 5 to pruned and both
  ... }
```

The idea is that when `known => pruned` then we know for sure that the
conditions on the current path have been realised. But, it turns out
that's the same as asking that `pruned` be empty! Why? Because "pruned"
facts are about values in the *pre*condition (or at least we'll remove
facts that make no sense in the PRE from pruned as callers have no way
of influencing them, in particular of "making them true"), while facts
in "known" are necessarily about the post condition by construction.
Ergo, "known" cannot say anything about "pruned"!

That means we need to change how we approach latent conditions.

In the new world, we only have "pruned" and "both" (the names haven't
changed yet to make the diff easier to review) and issues are manifest
when "pruned" is empty. "pruned" in summaries is evaluated when applying
the precondition, to see if the current state implies it.

**CHANGE 1**: get rid of "known": `Formula.t` is now just `{ pruned; both }`

**CHANGE 2**: we incorporate only "pruned" when evaluating a precondition,
 then add only "both" after having incorporated the post-condition

These changes surfaced some pre-existing issues that somehow conspired
with the previously-broken-ish state to yield better results than it
ought to have. This fixes them at the same time.

**CHANGE 3**: free(x) should propagate `x > 0` to pruned, not known! It's a
condition on the input. There was a similar issue in the handling of objc's
nil-messaging.

**CHANGE 4**: now we get a lot of latent issues because of change 3, so we
 introduce one more rule: x > 0 doesn't count towards being latent if x
 is known to be allocated. The reasoning is that we don't want to hold
 off reporting just because the current path requires values to be
 *valid and allocated*. In other words, the normal happy path is always
 expected to be possible.

**CHANGE 5**: the way we used to apply substitutions in post-conditions of
 summaries was subtly wrong and could introduce non-canonicalised values
 into the state. Now we rewrite the call substitution between pre and
 post to take into account the new arithmetic relation induced by the
 post.

*Note*: This change suggests a future change that isn't going to be addressed for
now: instead of letting clients/models decide which conditions to put in
pruned vs which ones to put in the "both" formula only, we can automatically
throw into "pruned" any fact that is relevant to the precondition. Or maybe
even just rely on "both" and detect if any facts are about the precondition at
all in there directly.

Reviewed By: skcho

Differential Revision: D38250210

fbshipit-source-id: 2735ee468ab5d05dc26e91ffa92d1342fd4af1a3