[const refable] prevent const refable issue being raised on moved params
Summary:
**Context**

Due to move semantics, it is common to take an argument directly by value, instead of by const-reference, and then move from it.

```
void foo(std::vector<int> vec) {
  bar(std::move(vec));
}
```

This pattern is usually beneficial when the calling code passes an rvalue and assumes the move construction is a cheap operation.

In these situations, it is common for a variable to be moved into the function, and subsequently for parameter to be moved to a child function, forming a chain of moves. In most cases where a parameter is passed by value and moved in the body of the function, it is intentional.

Therefore, we should not suggest that the function parameter should be turned into a `const&`, even if the parameter is not modified.

**In this diff**
- Check if const-refable function parameters get moved. If so, count these as modified.
- Add test cases for this functionality.

Reviewed By: skcho

Differential Revision: D38623065

fbshipit-source-id: 2f3c48e47f8d3a53e429bd969179f6a1130fb727