[sledge] Add encoding of formulas into Z3, use for alarm filtering
Summary:
Add an encoding of symbolic heap formulas into Z3, and use it to
implement unsatisfiability checking used for alarm filtering.

The sledge internal logic is unsorted, while Z3 requires
sorts. Therefore this encoding treats every element of the universe as
a rational/real number. This causes the interpretation of arithmetic
inequality to be not entirely correct. Also, terms representing
byte-arrays/sequences are left uninterpreted.

Reviewed By: bennostein

Differential Revision: D38617389

fbshipit-source-id: 9aa67c3bb577a4b4babdc67e98418f18af9fc6ed