[sqlite] remove source file column
Summary:
The source file column in the procedures table is only used for filtering. This diff gets the source file from the attributes column and removes the source_file column entirely.

This column was used to break ties when merging different databases. The attributes blob is used to do that now.

Reviewed By: martintrojer

Differential Revision: D37913285

fbshipit-source-id: 9276f39fef8dfbd4ed2c8f292a5aee240af59f5f