[datalog] Scoped variable names and opaque type
Summary:
To name variables, Infer uses a counter `n$0, n$1, ...`. This counter is reset back to 0 for each scope, creating variables with the same name, but in different scopes: this is a problem for the datalog analysis. To fix this, variable names are extended with the scope information (e.g. `n$1@Class.foo():void`).

As part of this commit, the `Fact.t` type is now opaque.

Reviewed By: davidpichardie

Differential Revision: D38829321

fbshipit-source-id: bb8488847294581842a75898bd2e60438be0b894