[pulse][unnecessary copy] Check addresses returned via unknown calls for modifications
Summary:
When we call an unknown function, we ensure to add its actuals with `WrittenTo` attributes so that we don't report noisy results on unknown calls that potentially modify their arguments.

However, we don't track any relation between the actuals and return values. So, if copy is from a source that is returned from an unknown function but we modify one of the actuals later on (as is the case in the example test which was a FP before), we can't currently detect it. This diff aims to solve this problem by keeping an additional atttribute `ReturnedFromUnknown` from ret_val -> actuals so that when we are checking for modifications to the return value, we also check for modifications to the actuals.

Initially, I realized that a similar connection is created for `PropagateTaintFrom` in three instances:
- from return value to actuals (this one)
- from this to other actuals
- from last actual to other actuals

So I tried to split them and reuse the first one for this purpose but it turned out to be not possible/easy: Taint information is all aggregated together into one  `PropagateTaintFrom` which is processed together for taint purposes. Instead, I am creating a separate Attribute for just returned values which is not related to taints.

Reviewed By: skcho

Differential Revision: D38908357

fbshipit-source-id: 8586323458c84cdb89c9aa4a3bcb1fb34d2dda0e