[IR][Textual] Switching from Core.Hashtbl to Caml.Hashtbl
Summary:
We now need to have hash tables with keys a bit more elaborate than plain string. `Caml.Hashtbl` are easier to use with complex keys so I prefer to switch for this implementation.

We then use this new datastructure for building hash tables for `Procname.qualified_name` and symbolic names `TypeName.t` and `VarName.t`.

Reviewed By: artempyanykh

Differential Revision: D39163188

fbshipit-source-id: 4034e1733c3fd2fa7c9e0eda176a9866da23a4b9