[infer-erl][1/n] Validate type specs: reject empty union types
Summary:
Extend the validation of Erlang ASTs to reject type unions where the list of types is empty.

We add a new case within `validate_form` to match `Type` and `Spec` forms, which invoke `validate_type` on the type ASTs occurring within.

For a `Union` type, we simply check that list of types `ts` is nonempty. If it's empty, we write an error message to the logs mentioning the name of the type, and reject the AST (`validate_type` returns `false`)

We also recursively check types that occur within other type constructors (such as `List` and `Tuple`).

Differential Revision: D38314352

fbshipit-source-id: f9d8903b34795fb9c5bae461d20837de8a6c8c0e