[pulse] Improve behaviour of PulseOperations.eval when applied to an exception expression
Summary:
Currently, `PulseOperations.eval` does not evaluate exceptions adequately (it simply returns a fresh abstract value that is unconstrained).

We change this so that an expression of the form `Exn(expr) : Exp.t` is evaluated by evaluating `expr` (which is the exception object) and returning the resulting abstract value.

This is necessary for the proper analysis of exceptions within a single procedure.

When a Java exception is thrown, the exception expression is assigned to the procedure's `return` register before `__java_throw` is invoked (which converts the state into an `ExceptionRaised` state). As mentioned above, currently (prior to this change), this simply assigns a fresh abstract value to the `return` register, meaning that the data in the exception expression is lost.

Reviewed By: rgrig

Differential Revision: D38973652

fbshipit-source-id: 7f7da6bca23b2529d72147c6412a0b61b685efa2