[pulse][refactoring] Introduce remove_by_rank
Summary:
Before this diff, we needed a dummy attribute to remove the same type of attribute from the set of
attributes.  The dummy value was only used for getting its rank, which is abstracted in the
`Attribute` module.

This diff introduces `remove_by_rank` in `Attribute`, which is for generating remove functions for
attributes.  We need to have some more `remove_` APIs, but no dummy value is required now.

Reviewed By: jvillard

Differential Revision: D39261720

fbshipit-source-id: aea86ad858de8cddd6124a85e7bd8ae98c9ff987