[textual] Add a test that finds a basic taint flow in a textual/hack source
Summary:
Although this is a very basic test, it shows that we have enough of the
Textual->SIL translation in place to start finding some taint flows.

This also allows us to experiment with some features in advace that are
to be implemented as part of Hack -> Textual translation.

There's some duplication in the Makefile. I want to clean it up after I
implement --capture-textual with multiple .sil files.

Reviewed By: jvillard

Differential Revision: D39207870

fbshipit-source-id: 2b41b4eab40dcf88b0296eb08f8c8e2ac8c30312