[sledge] Replace distance computation with a tabulation algorithm
Summary:
This removes the whole-program analysis previously used to compute distance
heuristics and replaces it with a compositional/summary-based tabulation
solver.

Procedures are summarized by the length of the shortest path through them,
along with information about their effect on the automaton state of the goal
trace and the newly-visited locations within.

Currently, the compositional analysis is organized in a "top-down" fashion,
building a summary of the `main` procedure by recursively summarizing each of
its transitive callees.  A "bottom-up" approach that summarizes procedures in
reverse topological order is also possible, but may incur unnecessary
analysis work by summarizing procedures in initial abstract states not
reachable from the entrypoint.

Reviewed By: jberdine

Differential Revision: D38941628

fbshipit-source-id: 79d59aa5adc356470ce9f73eaf1b2da651cb2162