[java] add test to document how to run Infer on JAR files (#1673)
Summary:
Running the analysis on a JAR file directly is less precise than running Infer during the build because some useful type information may be missing. However, it is a very useful feature to have when integrating Infer with the build system proves to be too complex. This pull request adds a test to document this feature of Infer and make sure it keeps on running.

All the tests are passing.

Pull Request resolved: https://github.com/facebook/infer/pull/1673

Reviewed By: ngorogiannis

Differential Revision: D39433852

Pulled By: jvillard

fbshipit-source-id: 7bf34b41668a2e07111355ad553af0172f73a92e