[pulse] Remove more unreachable variables in var_eqs
Summary:
This diff removes more unreachable variables from `var_eqs` before making a summary.

Before this diff, a variable was addressed as reachable if its canonical representative in var_eqs was reachable. For example, when `a1 = {a2}` was in var_eqs and `a1` was reachable, it did not remove `a2` even if `a2` was actually not used.

In this diff, it made the relation not symmetric:

* if a variable is reachable, its representative is reachable;
* however the other way is not always true. Even if a representative is reachable, it does NOT imply all of its equivalent class is reachable.

With this, we can remove `a2` from `a1 = {a2}` when `a2` is not actually reachable, i.e. the result becomes `a1 = { }`.

Reviewed By: jvillard

Differential Revision: D39410988

fbshipit-source-id: a575807f2e1c3929f8227c15c31f273e0032ad2c