[cli] --files-to-analyze-index argument
Summary:
In some contexts, it is desirable to analyze a smaller subset of the files modified in a commit. Some analyses, however, such as ConfigImpact, rely on knowing the exact set of modified files, so using `--changed-files-index` to analyze a subset is not optimal.

This diff adds a new, optional, argument, `--files-to-analyze-index`, which allows specifying the set of files to analyze explicitly. To avoid usage errors, using this argument requires also using `--changed-files-index` and that the set of files passed to `--files-to-analyze-index` is a subset of that passed to `--changed-files-index`.

Reviewed By: dulmarod

Differential Revision: D37483288

fbshipit-source-id: 15600ea295614ce5d836a1ebd7ef325747c04cd6