[pulse] fix incorrect `WithSummary (WithSummary ..)`
Summary:
Errors include a special case of an error that comes with a pre-computed
summary, to avoid Pulse doing the work of computing a summary from an
abstract state several times. But, the type allows the non-sensical
`WithSummary (WithSummary (error, summary1), summary2)`, which must be
avoided. Before this diff, we sometimes created it nonetheless because of a poor API choice of returning a
`WithSummary ..` in `of_abductive_summary_error`, which was hidden as a
simple `error` type. PulseReport was then inadvertently creating the
dreaded double summary.

Fix the API to make `WithSummary ..` creation more explicit.

Making the type safer would also be possible but is a large and
cumbersome refactor. Most places where we create an `error` type are
very simple: we create an explicit variant of `error`, so it's obviously
safe. It doesn't seem worth the added complexity for now.

Reviewed By: ezgicicek

Differential Revision: D39540271

fbshipit-source-id: 5cc52349b7e8186389083848b4d479d8f3ef16d3