[pulse] rename some variables in formula.conditions
Summary:
This avoids losing some conditions and incorrectly tagging issues as
manifest as a result in some cases. For example if the path condition
has `v4 = 0` and we have `v4 = v1` and only `v1` should be kept then we
should normalize to `v1 = 0` otherwise we'll drop `v4 = 0` as
irrelevant. If we're lucky it may happen that either `v1` is in the
precondition or is related to precondition variables in some way.

Reviewed By: ezgicicek

Differential Revision: D39579266

fbshipit-source-id: e7ab2c957975363843c1dc3ac7d31f07889f2264