[pulse] handle conjunctions inside terms
Summary:
Adding atoms that contain conjunctions was handled weakly by pulse, as
we didn't go into the individual components. With this diff,
`Atom.atom_of_term` is now `Atom.atoms_of_term` and can return a list of
atoms. We work harder to identify of a term can be transformed into a
conjunction of atoms. This makes us handle certain Erlang examples
better as the Erlang frontend sometimes generates constraints that are
conjunctions.

Reviewed By: fantazio

Differential Revision: D39579268

fbshipit-source-id: 7a6618fbe77fc77b66f08d8d3dcd65f89ac6e60b