[java] some simplifications and optimizations of the frontend (#1681)
Summary:
This pull requests simplifies the code of the Java frontend to make it more efficient when running the capture on JAR files. On the two benchmarks I looked at, running `infer capture --generated-classes file.jar --sourcepath $PWD` was 1.6X faster after these changes in the first case and 2.2X faster in the second case.

These optimizations includes:
- using a hash table instead of a map to collect the list of java classes
- avoiding the repeated creation of strings when matching the packages
- avoid getting the classname from a node when the classname is already the key of the hash table
- changing the order of the source and package comparison to compare file names (cheaper) before comparing the packages (more expensive)

These changes are mostly intended to reduce the memory allocations when running the frontend on a JAR file. The benefit should be lower when running the capture with Buck but I don't expect the frontend to become slower either.

Pull Request resolved: https://github.com/facebook/infer/pull/1681

Reviewed By: ngorogiannis

Differential Revision: D39855153

Pulled By: davidpichardie

fbshipit-source-id: 401fd120a26a8924c114b7ba8ea68936b4ad8f2b