[pulse] raise -> raise_notrace most of the time
Summary:
We use exceptions in a bunch of places to short-circuit the execution
flow but we don't expect these exceptions to escape the module. We also
don't care that they care the current backtrace so we can use
`raise_notrace` instead of `raise`. This avoids polluting the backtrace
of exceptions that reach toplevel and are printed for debug. It's also a
tad more efficient though I haven't measured if that matters.

Reviewed By: skcho

Differential Revision: D39885341

fbshipit-source-id: 177b5d59679207a75e7b53feb4f0e1d13aa9cbbb