[analysis] add a basic CPU-time timeout to the AI framework
Summary:
Pulse sometimes takes too long to analyse a procedure and there's no
easy fix in sight, eg the procedure is just absurdly long and/or keeps
calling other procedures with a lot of disjuncts in their summaries.
It's hard to either make Pulse so fast that it doesn't sweat it, or to
make sure to catch all such edge cases and refuse to analyse them a
priori. A timeout is a simple solution to catch many of these cases.

Unlike the biabduction implementation in terms of Unix signals, which
guarantees a tight bound on the maximum time spent on one procedure, this
timeout is simply periodically checked by the AI framework. As such,
procedures can outlast the maximum time by quite a bit if it takes too
much time between two timeout checks. We do observe that running on
OpenSSL, eg with a timeout of .4s most procedures time out at almost
exactly 400ms but some go for several seconds.

This can be improved by either checking in more places
(`Timer.check_timeout` calls can be inserted anywhere, which is flexible
but this sort of whack-a-mole never ends...) or by adding another
signals-based timeout (risky, suddenly an exception can be raised in any
piece of code!).

The `Timer` module could also be extended to measure the time taken by
various checkers in the future.

Reviewed By: ngorogiannis

Differential Revision: D39770941

fbshipit-source-id: 24b1b3a12abd9dea6953ff8944a47728f581c8e2