[IR][Textual] an expression can be a type
Summary:
we now allow types to be normal expressions.
To ease parsing, types are enclosed in a `< typ >` syntax.

We expect this type-as-expr to appear in only 3 builtins:
* `__sil_cast(<T>,exp)` (this a side-effect free expr)
* `__sil_allocate(<T>)`
* `__sil_allocate_array(<T>,dim)`

We will add `instanceOf` test in a next diff.

Since we do not expect type-as-expr in user-defined functions, we don't need type of type... (We will check that during type verification T126360471)

Reviewed By: artempyanykh

Differential Revision: D40024049

fbshipit-source-id: 55ee610346b950877879eca0f5945e3ed270b34b