[sledge] Derive constant-prop information from branch conditions
Summary:
When a conditional branch is based on some integer constant, equality or
disequality, we can potentially learn some integer-constant information at
the relevant jump targets.

For example, in the following LLAIR block (terminated by a conditional
branch)

```
    %call!19 := nondet "__llair_choice";
    if (((i32)(s32) %call!19) = 12) goto %entry.jmp.return;
    else goto %if.end;
```

we can infer that `%call = 12` when control goes to `%entry.jmp.return`.

Reviewed By: jberdine

Differential Revision: D40101612

fbshipit-source-id: eb7d869af11f29917c7b8ef6aafe1dac8f324361