[pulse][taint] allow taint_targets to be fields
Summary:
In addition to specify some or all arguments or the return value as
taint target, one can now specify only part of their content by
specifying which fields should be tainted.
Rather than selecting multiple fields, we only allow selecting one field
at a time and that field must belong to a taint_target. This implies
that we can select the same field for multiple arguments as well as
select the field of a field of the returned value or any combination of
those.
If multiple fields must be selected, then multiple specifications will
need to be written. This makes sense if we assume that in most cases
their won't be a need to be very specific and when that need arises then
only a field or two would need to be selected.
If e.g. all the fields but one must be seleted, then they all need to be
specified.
We also need to do some type checking to verify the field exists in the
taint_target's type. Doing otherwise (just assuming it exists) could
lead to weird FP

Reviewed By: dulmarod

Differential Revision: D39847627

fbshipit-source-id: e071d2e508c44be39afa4d5bac2a2d45608ede52