[sledge] Validate witness histories via SMT encoding
Summary:
This diff adds a validation step to reporting that a goal trace has
successfully been followed to the end. This validation step involves
an SMT encoding of the semantics of the instructions in the witness
history, and a simplistic search through the paths represented by the
history DAG.

This diff also adds a `-dump-witness` flag to save the witness found
by symbolic execution to a file, and the `sledge validate` subcommand
for subsequently solving this for a valid concrete execution.

Reviewed By: bennostein

Differential Revision: D40077548

fbshipit-source-id: e21fd16f31203c255bb452acb7662319b41347a2