[pulse][taint] Record a false positive that implies lazy field creation
Summary:
Because fields only exist in the memory after they have been accessed
(read or write) and sources taint everything that can be accessed in the
memory from a certain value, fields that have not been accessed before
the structure they belong to is tainted will not get the taint. This
leads to false negatives as shown in the test.
The same situation is also true for pointers and the values they are
pointing to if the pointers has not been dereferenced before being
tainted.

We need to propagate the taint to the content of a tainted value at the
moment we add that content in the memory

Reviewed By: skcho

Differential Revision: D40339907

fbshipit-source-id: 99a9de78abc592599e1ca08a3a568cafc5bc1087