[pulse] Create a new CFG module ExceptionalNoSinkToExitEdge
Summary:
We define a new module ExceptionalNoSinkToExitEdge, which implements the ProcCfg interface.

The purpose of this is to (later on) create a view of the CFG where there is no edge from the exception sink to the exit node, and where summaries are extracted directly from the exceptions sink node and included in the summary (with type `ExceptionRaised`).

Reviewed By: rgrig

Differential Revision: D39258471

fbshipit-source-id: 553af7b02628114f691a760d2bea12ee767b66d5