[I] Skip instructions for bot states.
Any instruction should only produce bottom state out of a bottom
state so simply use identity as a transition function if a pre
state is bottom.

This fixes assertion failure on analysis of codepeer - it happened
that a transition function of some instruction assumed non-bottom
pre state.

Diff to be viewed with whitespace changes hidden.

# Conflicts:
#	infer/src/bufferoverrun/bufferOverrunAnalysis.ml