[pulse] Generate a proper summary from ExceptionRaised instead of just returning Unsat
Summary:
We modify `exec_summary_of_post_common` to extract a summary from an `ExceptionRaised` instead of just returning `Unsat`.

To do that we call the `summarise` function introduced in the previous diff.
We also add a new `exception_raised` argument to `exec_summary_of_post_common` for wrapping the summary in a suitable execution domain.

However, at the moment this does not get invoked because the postconditions extracted from the exceptions sink node are of type `ContinueProgram` (and we do not expect `ExceptionRaised` states to appear at the procedure exit node either).

Reviewed By: rgrig

Differential Revision: D39498564

fbshipit-source-id: d042b575449ac2984e9f170f18cc95b82a864aca