[pulse] Postconditions of exceptions sink get turned into ExceptionRaised
Summary:
We turn `ContinueProgram` postconditions at the exceptions sink into `ExceptionRaised`.

This is necessary for the proper modelling of inter-procedural exceptional control-flow. This change allows procedure summaries to contain postconditions of type `ExceptionRaised`, which will cause a exceptional control-flow edge to be followed in the calling procedure when the callee throws an uncaught exception.

Reviewed By: rgrig

Differential Revision: D39499482

fbshipit-source-id: b6c8d8db5d1ca17525d65a7d0ab1cafae1592c2c