[pulse] Change behaviour of Sil.Call when the call returns an ExceptionRaised state
Summary:
We modify the symbolic execution of the SIL Call instruction, so that when a procedure calls returns an `ExceptionRaised` state, the returned value of the call gets copied into the `return` register of the calling procedure.
This is needed since the exception handler (in the calling procedure) expects the exception object to be present in the calling procedure's `return` register.

We also rename the Jave pulse test case `FN_callSinkWithSourceInsideExceptionObjectBad` to `callSinkWithSourceInsideExceptionObjectBad` to indicate that this is no longer a false negative.

Reviewed By: rgrig

Differential Revision: D39305655

fbshipit-source-id: 4f61a678e497a56659b2b227151f2eea064e8de4