[F] Fix strong updates when handling subprogram calls.
When updating the state of a caller with information (about updated out
parameters and global variables) from a callee, strong update of caller
locations was always performed. This is not correct in particular
if a location from callee maps to more locations in the caller (for
example a dereference of a pointer with more possible targets is passed as
out parameter) or if a location in the caller represents more values (for
example data stored in an array).

Fix the problem by checking whether a strong update can be performed.
This comprises of:
1) assume that a single allocsite cannot represent more allocated objects
(this is currently not ensured by BO, but can be ensured by a frontend)
2) check whether the locations imported to the caller can be strongly
updated or the location in the callee map to a single location in the
caller and the following assumption holds: A location corresponding to an
input of a subprogram (global variable or subprogram parameter) has either
value top or all the objects it represents were updated. This assumption
holds when default value for a location is top and such location is always
weekly updated:

    Example 1:
       procedure Call_Weak_Update_Arr (B : Boolean) is
          A : Arr;

          procedure Callee_Strong_Update (Arr_Par : in out Arr) is
          begin
             Arr_Par (2) := 2;
          end Callee_Strong_Update;
       begin
          A (1) := 1;
          Callee_Strong_Update (A);  --  did strong update of the content of A
          pragma Assert (A (1) = 1); -- A (1) evaluated to 2
       end Call_Weak_Update_Arr;

    Example 2:
       procedure Call_Weak_Update (B : Boolean) is
          Ptr : Int_Ptr;
          I, J : aliased Integer := 1;

          procedure Callee_Strong_Update (O : in out Integer) is
          begin
             O := 2;
          end Callee_Strong_Update;
        begin
          if B then
             Ptr := I'Unchecked_Access;
          else
             Ptr := J'Unchecked_Access;
          end if;
          Callee_Strong_Update (Ptr.all);  -- did strong update of I and J
          pragma Assert (I /= 2);  -- I evaluated to 2
          pragma Assert (J /= 2);  -- J evaluated to 2
       end Call_Weak_Update;

Add identification of the procedure where the allocsite was created. That
is, the procedure which called the builtin creating the allocsite and to
which state the mapping "allocsite -> value" was inserted. Without this,
it can happen that two identical allocsites for unrelated allocations are
created in two different procedures. This can lead to incorrectly overwriting
the value of such allocsite when handling calls between such procedures.
Example:

   procedure A is
      A := malloc;
      B := malloc;

      function Callee return Something is
         --  Identify the allocsite created by malloc also by Callee.
         --  Otherwise, it would be identical to the allocsite created
         --  by malloc in A and the value of this allocsite would be
         --  overwritten when processing the call to Callee leading
         --  to incorrectly reporting the assertion failure.
         C := malloc;
      begin
         C := 41;
         return C;
      end Callee;
   begin
      A := 40;
      B := Ret_Array;

      pragma Assert (A (1) = 40);  -- @assertion:PASS
   end A;

* infer/src/bufferoverrun/bufferOverrun.Analysis.ml
(instantiate_mem_reachable): When handling subprogram calls always strongly
update locations with values from calees (even locations representing multiple
values). Do weak updates when a location from the callee maps to more locations
in the caller.

* infer/src/bufferoverrun/bufferOverrun.Domain.ml
(MemReach/Mem.update_mem): add an optional parameter to force strong update.

* infer/src/bufferoverrun/absloc.{ml|mli}
Add an identification of the procedure where the allocsite was created.