[IR] Simplifications before removing the root_typ load/store field
Summary:
This field is only used by biabduction and its role is often confusing. In this diff we make simplifications that preserve biabduction test sets. We will remove the field in a following diff.

The main change is in SymExec.ml where we are able to reconstruct the root_typ from the regular field type.

Reviewed By: jvillard

Differential Revision: D40546893

fbshipit-source-id: d6f12991f4190c6a735266ea3f7317c4fdf4f4f7