[pulse] Modify the option pulse-model-return-first-arg to be able to model functions as well, add option pulse-model-return-this.
Summary: The option `pulse-model-return-first-arg` was being used as a `return-this` behaviour for Java. Now that behaviour is handled by the option `pulse-model-return-this` and `pulse-model-return-first-arg` returns the first argument of either C functions or ObjC instance or Java methods, and its meaning is to return the first argument of the function in terms of the source language semantics.

Reviewed By: skcho

Differential Revision: D40591234

fbshipit-source-id: 59722b41a6f754d629cfba0da5bb028ea6b23250