Add a custom model for `make_tuple` in SimpleShape
Summary:
As tuple construction in Erlang is represented by a builtin call, we need to provide a custom model
of that function to the SimpleShape analysis to make it able to correctly infer fields on tuple
construction (and not only on tuple pattern matching).

We implement a simple "custom model" functionality for this purpose. Future extensions could include
specific models for "fake structures", such as integers or atoms.

Differential Revision: D40686791

fbshipit-source-id: 8d63d7c253afcafa6b0eda8547dede23768e553c