[pulse] Fix infinite loops in the retain cycles
Summary:
We had a problem in the retain cycle check where we would loop forever in the case such as

```
      mem  ={ v1 -> { * -> v4 },
              v2 -> { * -> v6 },
              v4 -> { _replyMediaInfo -> v3911, _pk -> v3806, * -> v5 },
              v10 -> { _replyMediaInfo -> v3918 },
              v3806 -> { * -> v10 },
              v3911 -> { * -> v10 },
              v3918 -> { * -> v10 },
              v3945 -> { * -> v10 } };
```

The original address had the `is_ref_counted` attribute so we started the check, but other addresses in the cycle didn't have the attribute, so we looped forever. This is now fixed.

Also made `seen` and `checked` sets to avoid duplications and faster membership check.

Reviewed By: jvillard

Differential Revision: D40938217

fbshipit-source-id: 97ab0f29e7fd6ea4e0f07f8fb1d153b8b8637b9d