[buck] rename ClangFlavors to Clang
Summary:
Since `BuckMode` `ClangFlavors` is used to select the buck/clang integration depending on buck version (and since there is no such thing as buck2 flavors) rename the variant to just `Clang`.

Also, rename `Buck2` module to `Buck2Clang` since it's not going to be the only buck2 integration when the java one arrives.

Reviewed By: jvillard

Differential Revision: D41121125

fbshipit-source-id: b0e56b8bda337060d09410c5dc16cfb43c59514c