[java] add tests documenting the dependencies analysis mode (#1679)
Summary:
This pull request adds a test to document the `--dependencies` options triggering the analysis of the classes appearing in the Java classpath. This feature is especially useful when analysing JAR files packaging all the code of a Java application including the dependencies.

The examples added in the test are outlining how running the Pulse taint analysis with the dependencies mode enabled can eliminate cases of false negatives and cases of false positives.

All the tests are passing locally.

Pull Request resolved: https://github.com/facebook/infer/pull/1679

Reviewed By: bennostein, ngorogiannis

Differential Revision: D39433881

Pulled By: jvillard

fbshipit-source-id: bacdfcfd1a257750ad99ed1838182ffb82fd8a65