[hack] Use --keep-going flag to determine how to deal with errors
Summary:
:
Without the flag any capture error will cause the process to die with a
non-zero exit code. With --keep-going capture errors will be ignored.

I've also updated the config and manuals wrt. --keep-going being used in
capture. It's been like that for a while in buck and clang drivers, just
not explicitly exposed in the manuals.

Reviewed By: geralt-encore

Differential Revision: D41218011

fbshipit-source-id: 48e9e80e66977b2a75b14ca574905fc7dc588bff