[pulse] Add formula that constant closure is not nil
Summary:
This diff adds the formula that constant closure addr is not nil.

Problem:

When an unknown message is sent to an object, the abstract semantics (always) makes two disjuncts: (1) where the object is nil, (2) where the object is not nil.

However, when the object is a constant closure, it should not make the former, because which makes a contradictory state that an address is null and closure, at the same time.  Even worse, by unification of null value, the closure is propagated to other non-function typed null pointers, introducing broken closure inlinings.

Reviewed By: dulmarod

Differential Revision: D41240348

fbshipit-source-id: 11acffa129bf7206898c3d438ba765939a101a1f