[erl-frontend] Add support for named lambdas
Summary: Add support for named lambdas by creating a closure and binding it to a variable with the same name as the lambda in the beginning of the lambda's procedure. Note that this is not very useful in Pulse because named lambdas are mostly (if not always?) recursive and Pulse doesn't support recursion. However, it is useful in lineage.

Reviewed By: thizanne

Differential Revision: D41539672

fbshipit-source-id: bfa357bdb99f211b48b96f3585846224fc8d8bd0