Unifying Doli and Textual -- 2nd part
Summary:
`Doli` and `Textual` should share the same description of the bodies of the functions/procedures.

We made one large parser called `CombinedMenhir`  with two entries, one for `Textual` and one for `Doli`. And we unified the lexer of `Doli` and `Textual` into `CombinedLexer`.

We needed to make it possible for identifiers to also play the role of keywords, e.g. `final` is both a keyword and an identifier.

*NEXT STEPS*

(A): Make the `Doli`-part in `CombinedMenhir` read `Textual` bodies.

(B) If easy, refactor the various sub-parsers into different files.

(C) For all Doli keywords, check whether they should also be accepted as identifiers.

Reviewed By: jvillard

Differential Revision:
D41522665

LaMa Project: L1122176

fbshipit-source-id: 12a162a4a7b2f2946f494909a194003dc9bab061