[unnecessary copy] Report on intermediate copies if the source is never used later
Summary:
If we copy from a source var into a temp where the temp is modified later on, we currently do not report. However, this is too strict if the source variable is never used later on.

In this diff, we report on such cases. To do that, even if the temporary is modified, we need to keep track of the copy data so that we can report on it later on.

Due to infer's clang frontend (where we first dereference parameters before passing them to callees but don't do the same for locals), there are two cases that needs to be handled:

- parameters where it suffices to just check loads to see if the source is accessed
- local variables where we also have to check if we call source var on functions after the copy. For this, we extend the `passed_to` to include info on source variables (that occur in the source code) along with their timestamp.

Reviewed By: skcho

Differential Revision: D41612623

fbshipit-source-id: 5e32b0cb2535ad9bdd6ab1d8aa4577f3259e7f8e