[pulse] Fix a bug in post calculation of non-disjunctive domain
Summary:
In the previous implementation, when analyzing a node in a loop, it ignored the non-disjunctive
domain value in the `old_state`, i.e. non-disjunctive domain value is always newly calculated by
disjunctive domain values.  However, the problem is that disjunctive domain values can be skipped
when they were already visited before, which made the non-disjunctive domain value smaller than
expected, and it became the bottom in the worst case when all disjunctive values were visited.

Reviewed By: ezgicicek

Differential Revision: D41625328

fbshipit-source-id: b893b613c49112a68323d034e087c21142929609