[pulse] Call initializer for global constant function argument
Summary:
This diff calls global constant initializers when they are used as
function argument.  There have been a mechanism calling initializer
when a global constant is `Load`ed.  This diff applies the same
mechanism to function arguments.

One minor fix is that not calling an initializer when it is analyzing
the initializer body.

Reviewed By: fantazio

Differential Revision: D41523313

fbshipit-source-id: e0862e76dc4b51594e6f1b8b97375c86bd0b390d