Add caller_procname to ModelEnv and remove proc_desc from OnDemandEnv.
As suggested by the maintainer of BO from FB:

"It could make more sense that adding caller's proc name to ModelEnv,
rather than OndemandEnv (which is in Dom.Mem), since the ondemand env is
the data for evaluating parameter values on-demand."

Proc_desc in OnDemandEnv is only used when creating a new Allocsite to
file the procname of the caller of the function creating the Allocsite.
This is needed in order to not end up with two identical allocsites
created in two different subprograms by the call to the same builtin. In
these situations, we either have already the correct procname (we don't
need to get it from OnDemandEnv) or ModeEnv is available and we could add
the procname to ModeEnv.