Refactor SimpleLineage assignments
Summary:
This is a no-op change. The processing of assignments has been inlined up to `exec_instr`, to allow
matching as early as possible on the different cases and exclude Prune, Metadata and the like from
the beginning.

This will allow later writing easily some specific matching on the assignments to manage fields.

Reviewed By: Rhapsod

Differential Revision: D41273030

fbshipit-source-id: 9c9c3c5e00ea3b0fdc22aee8f08e7fda2c35e2a2