Shape analysis: give a shape to the args and returned values of a unsupported function.
Summary:
The previous model for unsupported-expressions-as-functions and unknown functions did not update the
abstract state, especially, it did not unify the shape of the ret_id special variable, used to store
the result, with anything.

This is sometimes wrong when the ret_id is not used elsewhere on any intruction that would give
a shape to it, as it can have no known shape (which will be an issue for lineage).

We fix that by always giving a shape to the returned value and the parameters, even for unknown
function calls where this shape will therefore have no known field (which remains a sound
abstraction).

Reviewed By: rgrig

Differential Revision: D41848206

fbshipit-source-id: 80f40b83a0949538b5a629434fc9851c84d16d13