SimpleShape fields are now in syntactic order
Summary:
This is a purely documentation change on a feature not-yet-implemented. Fields were originally
"reversed" in draft implementation but in the end they will be in syntactic order. As there already
was a comment for the order we push a commit to fix that comment, to make it clear that it has no
effect and make the ulterior change semantics clearer.

Reviewed By: rgrig

Differential Revision: D41883674

fbshipit-source-id: 887c7d4548c587e0eb82d06eb7f3486b1643f11a