[scope-leakage] An interprocedural version of the scope leakage analysis
Summary:
Basic plumbing to make the analysis interprocedural.
The summary contains only the scopes inferred for the parameters and return value and applying the summary copies the scopes to the actual arguments and to the return variable.

Currently, the analysis doesn't usually reduce the false negative rate. This is because it requires a slightly stronger pointer analysis.
Also, the abstract transformer omits a case, since it would result in showing users 'Unknown' scopes, which stand for top. This is because currently only one scope is allowed per variable. The next diff would solve this by allowing to associate more than one scope and also a detailed trace log.

Also, fixed a bug in must_not_hold, which needed to use existential quantification for top arguments, otherwise it would report, for example, the an object of scope bottom must not hold an object of scope top.

Reviewed By: bennostein

Differential Revision: D41529461

fbshipit-source-id: 23af4cb8d76a0c491fa16d06ee2a3ec38620e17e