[doli] first step toward writing models with the Doli langage
Summary:
A Doli program should generate a matcher that converts
an opaque procname (for wich we don't have actual implementation or do
not want to look at it) into a concrete procname that is attached to
a procedure implemented in the Doli program.

In this diff, we are missing two important features:
1) we need to convert a Doli program into a list of Textual procdesc
2) we need to implement the matching between a Doli matching and a
procname.

Differential Revision: D42193883

fbshipit-source-id: 75ec47ce5b20904163d7480770ff998319add416