Add an option to limit cyclic types fields distinction in lineage
Summary:
We add an option `--lineage-prevent-cycles` that will have SimpleLineage break the distinction of
variable fields when they enter a cyclic type.

That is, if [X#a#b] has the same type as [X], the terminal fields of [X] with a depth limit of [5]
will stop at [X#a#b] and not include, for instance, [X#a#b#a#b#a].

This is especially useful to abstract some lists into the only [L#head] and [L#tail] nodes (and
possibly the sub-nodes of [L#head]), instead of recursing into [L#tail].

Added to the fact that, due to abstraction, [L] in the previous example could also have some other
discovered fields, this prevents a lot of combinatorial explosion due to recursive+wide structures,
which in turn should allow supporting wider and/or deeper structures without compromising the
scalability of the analysis.

Reviewed By: rgrig

Differential Revision: D42152601

fbshipit-source-id: 68b9701282cb6096f52c52bfcb480eb832307bca