[pulse] Gather taint dependencies as graph
Summary:
This diff gathers taint dependencies as a graph, instead of a set, so that it avoids false positives
on tainted sources when the sinking value is actually sanitized.  For example,

```
x = source();
y = sanitize(x);
sink(y);
```

we should not report a taint issue on `x`, because the sinked value is `y` and it is sanitized. By
this diff, it addresses `x` as if it is also sanitized when checking tainted-ness for `sink(y)`.

Reviewed By: fantazio

Differential Revision: D42483292

fbshipit-source-id: 0ce7fcbfb0fb974163b00d07a2d7d800db86bc1a