[erl-frontend] Fix location for some artificially inserted nodes
Summary: Sometimes we insert extra nodes in the CFG that do not directly come from the Erlang AST. One example are guards, where we have to do some extra bookeeping (e.g. unbox Boolean to SIL integer, or do branching based the result). Previously some of these nodes' locations were not updated and were pointing to the beginning of the function (default). This diff fixes this.

Reviewed By: rgrig

Differential Revision: D42547188

fbshipit-source-id: 90b985ce1401b2cb5bda2f4406e5992b0dcadb3b