[erlang] Added "ArgumentReturn" custom model
Summary:
This is a new type of custom Erlang model behavior. The existing one,
`ReturnValue`, lets one override the behavior of a function to be that a
specific value (possibly containing non-deterministic components) is
returned, *irrespective* of the argument values. The new behavior,
`ArgumentsReturnList`, allows one to specify a list of argument&return
value pairs. Arguments can also have nonspecified (sub)values, which
means "any".

Reviewed By: mmarescotti

Differential Revision: D41740135

fbshipit-source-id: 3e50765600ec82c9c672a4595c4e510cc17f48b3