[frontend] Translate static_cast to rvalue-ref
Summary:
This diff translates `static_cast` to the type case expression in SIL. To avoid too much of change in analysis results, it translates an expression only when it is a cast to an rvalue-ref, i.e.

* `static_cast<T&&>(exp)` ===> `(T&&)exp`
* `static_cast<T>(exp)` (the other cases) ===> `exp`

The translated result will be used to avoid FPs in const refable issues in the following diff.

Reviewed By: ezgicicek

Differential Revision: D42675760

fbshipit-source-id: 6a92ed6097ac2f466b6addd350ecb3d6ca72573e