[pulse][array] canonicalize arrays when applying pre and post in calls
Summary:
We only applied the canonicalization after applying the callee's pre. At
this point, we already know some equalities and can deduce more via the
canonicalization but don't know all the equalities known by the callee.
The missing information is gathered and compiled with the known one
after applying the post. Therefore we would need to canonicalize after
applying the callee's post. However, applying it only there is not
sufficient because then we would have applied the callee's summary in
incomplete or incorrect information before trying to canonicalize the
memory representation.

What we want to do is canonicalize anytime we normalize the subst map.
Indeed, when we normalize, we rely on canonicalized representation of
values, so this is the right time to actually canonicalize our memory
representation. This is done twice: once before applying the post to
ensure we propagate information from the post to the right value and
once after applying it because the application of the post and the
propagation of the callee's equalities may have revealed some new
equalities in the caller.

Reviewed By: dulmarod

Differential Revision: D41948533

fbshipit-source-id: 0a12f7cd20a4dbdbf709afad12ed104d9442f2ed