[incr] refactor and unify dependency-tracking data structures
Summary:
This diff introduces a basic HashSet implementation to the infer stdlib, and replaces existing
dynamic dependency tracking structures to use that (instead of mutable references to functional data
structures).

In order to avoid serializing and persisting unnecessarily-large hashsets, we also bundle the
various types of summary dependencies into a `Deps` module and add a notion of "partial" (i.e.
tracking dependencies of a summary in the process of being computed) and "complete" (i.e. for a
fully-computed summary, to which no more dependencies will be added) dependencies.

Reviewed By: ngorogiannis

Differential Revision: D42744915

fbshipit-source-id: a25848b1526c1575c58a1df2a3e14b2bfd9a02f0