Remove also frontend generated vars from the scope.
The instruction ExitScope is generated for all variables
(both Var.LogicalVar and Var.ProgramVar) which become dead
according to the Liveness preanalysis.

BO implements this instruction by removing all variables of the
type Var.LogicalVar out of the abstract state and doing nothing
for variables of the type Var.ProgramVar. The rationale is that
the Liveness analysis doesn't take aliasing into account and
program variables can be aliased. That is, they can be accessed
after the instruction ExitScope through a pointer and removing
them from the abstract state is thus undesirable.

However, keeping all program variables in the abstract state
poses a performance problem - an example is an analysis of
a file gen_il-gen-gen_nodes.adb from GNAT sources. So this commit
removes the frontend generated variables from abstract states.
Such variables cannot be aliased so normally, it should be safe to
remove them. However, it seems that the content of these variables
is used by cost analysis so this is why this is put into a flag.
Perhaps, it would be safe to remove the variables also with
respect to the cost analysis when we are not in a loop.