[inferbo] Fix an ignored case when the upper bounds of offset is +oo, and the upper bound of size is smaller than +oo
Summary:
The buffer overrun checker misses to handle the case when the upper bound of offset is +oo, and the upper bound of array size is less than +oo, which will causes false negative in some test cases.

For instance, for the following program,

```
int a[1];
for(int i=0; a[i]; i++) {}
```

The variable `i` will eventually be equal 1 and causes an overrun error within the loop statement. However, this error was missed by buffer overrun checker.