More precise handling of symbolic intervals.
Handle symbolic intervals in more precise way. In particular (1) improve
the conversion of (symbolic) intervals to boolean values, and (2)
improve the equality of (symbolic) intervals. This is important to
make evaluation of expressions more precise.

* infer/src/bufferoverrun/itv.ml
(is_false): instead of comparing the interval with False (zero) directly,
return True if the interval is <= and >= zero. As the result, is_false
returns True also for [B(u), 0] and [max (I(s), 0), min (I(s), 0)].
(eq_sem): Return True also in the case when the bounds of two intervals
itv1 and itv2 are not equal and it holds that itv1 <= itv2 <= itv1.
(ne_sem): Use eq_sem instead of duplicating the logic of eq_sem.