[clang] Store attributes even if the function has no body
Summary:
For implicitly defined methods (e.g. ctors/operator=), our frontend didn't create procdesc (and consequently attributes) hence it was not possible to query the attributes for these types. This diff fixes that by just storing the attributes for these.

See D42997333 for test reading from these attributes.

Note: this doesn't work for library calls yet because they are not translated at all (unless `--header` is passed). Ideally, we should store their attributes (only for the types that are used in the file, not everything otherwise we will have a huge attribute store) but it looks complicated so I will try to do that later.

Reviewed By: skcho

Differential Revision: D42927488

fbshipit-source-id: 86ab0ce151c41a8814045ebbc85dda225a1b86e5