[misc] Flush formatters in the parent to avoid double printing
Summary:
In ForkUtils.protect we already do Logging.reset_formatters which
flushes formatters, but this happens in children. We need to flush *in
the parent* before we fork to avoid carrying over stale buffers.

This diff adds a call to Logging.flush_formatters to ProcessPool before
it forks children. This is not a bullet proof fix (ideally, we should
flush in the parent before *every* call to fork), but it works fine and
gets rid of some double printing I saw.

Reviewed By: rumster

Differential Revision: D43452027

fbshipit-source-id: 4e303110ee88a3bdd758feb0840c9485bbf33459