[pulse] Fix a bug in calculating captured actuals in closure call
Summary:
This diff fixes a bug in binding captured values to corresponding
variables in closure calls.

- From procedure attributes, it constructs `captured_formals`.
- Using the `captured_formals`, it constructs `captured_actuals`.
- Then, actual binding is done with `captured_formals` and `captured_actuals`.

A problem was that when the callee is a closure, it sometimes constructs `captured_actuals` the order of which is different to `captured_formals`. The mismatch is coming from:

- `captured_formals` are made in the order of fieldnames, https://github.com/facebook/infer/blob/main/infer/src/absint/ClosureSpecialization.ml#L59

* `captured_actuals` are made in the order of edges,
https://github.com/facebook/infer/blob/main/infer/src/pulse/PulseClosureSpecialization.ml#L208

To avoid the mismatch, this diff changes the type of `passed_closure` (which is the source of `captured_formals`) from a map to an assoc list, then makes sure that it is constructed with the same order with `captured_actuals`, i.e. the order of edges by `BaseMemory.Edges.fold`.

Reviewed By: ezgicicek, dulmarod

Differential Revision: D43437889

fbshipit-source-id: 93029bd40cdb56e736c67ec38ecd1f4b40ab6974