First step of interprocedural Lineage: support structured function arguments
Summary:
When a function argument is known to be a structure (from the analysis of that function), upon callin
the function with concrete arguments, we only derive flow from the components of these arguments that
are known to flow into the result, instead of the whole concrete argument. For instance, upon calling
a `fst` function that returns the first element of a tuple as `Y = fst(X)`, only `X#1` flows into
`Y`.

The generation of the graph will still involve one `fst/1:arg0` node (the fields are not splitted),
only the `Derive` edges will be more carefully generated.

Reviewed By: rgrig

Differential Revision: D42998266

fbshipit-source-id: 64324cfb9e4a183bfc4bd7f875df48b5caab6570