[unnecessary copy] Prevent double reporting
Summary: We were double reporting when there is a conditional in a user defined copy assignment operation. This was because in one branch, the source was coming from the arguement to the `operator=` but in the other branch it was coming from `this`. However, for the latter case, the decompiler kept track of the `operator=` as the source expression because the order of the evaluation of source and copy was wrong. This diff fixes that.

Reviewed By: skcho

Differential Revision: D43570510

fbshipit-source-id: ed2e6ebb754733542731fc66abe9c71264c2aecf