Extract/unify dependency tracking into its own module under IR
Summary:
Instead of tracking dependencies separately in `Summary` and `Tenv` (and potentially soon-to-be
`Attributes`), keep it all in one place under the IR module.  Though not "IR" technically, this is a
good place for dependency-tracking to live so that it has access to Procnames and SourceFiles but is
maximally available in other places where dependencies need to be recorded.

Reviewed By: ngorogiannis

Differential Revision: D43569089

fbshipit-source-id: bc7823b407211826c7a05c8ef5fe33c198cc1775