[clang] Simplify adding an ObjC class to the tenv
Summary: In ObjectiveC we can find info about an ObjectiveC class in the ObjcInterface node or in the ObjCImplementation node, and there could be more than one ObjcInterface node for a given class. For that reason, we need to collect this information in different places to add to the tenv. This diff tries to simplify this and aims to ensure that we don't lose previously added information to the tenv. It does this by unifying how we add the info to the tenv from the ObjcInterface and the ObjCImplementation nodes.

Reviewed By: ngorogiannis

Differential Revision: D43574297

fbshipit-source-id: 9be7fe118075b43d79227dde2bc50b9178c0f116