Make the import of dereferences more robust.
When importing a location representing a dereference from a callee
to a caller, fallback to attempting to create a symbolic location if
the target of the dereference is unknown in the caller.

Note the standard mechanism for creating symbolic locations can
fail because of the lack of type information. For example, in the case of
Ada, when calling a method on an abstract type that modifies a field of a
concretization of this type, we don't have type information for
the field in the caller.

Lift the condition on using the type passed to
BufferOverrunDomain.Valon_path when creating a value of a location
to always use this type when the type cannot be induced from formal,
actual parameters and type environment
(computed by BufferOverrunOnDemand.typ_of_param_path).
The rationale is that some instructions give type information
which cannot be computed by the mechanism above and it is too
bad to not use it. Since having type information means creating
a symbolic value, the consequence of this change can be that
a symbolic value can be created even for local variable. This is
no that bad since such value behaves as top and won't be imported
to callers of the current subprogram.