[config] remove "language" from matchers
Summary:
This was unused in the sense that only one language, Java, was supported
anyway. It was also pretty hard to actually implement the language-based
matching given the current state of the code. I suspect it's not that
useful anyway so in order to use this beyond Java let's just remove it.

Reviewed By: ngorogiannis

Differential Revision: D43667359

fbshipit-source-id: f94a5d7d005e1d1b524bc4525276f52c67ba7351