[clang] New preanalysis to substitute away mock pointers
Summary:
This aims to substitute away mock pointers from the code. This consists of a programming pattern where developers use global variables that contain a function pointer, where the function pointer points to the real function that we want to call in the production code, and gets swapped by a dummy function for testing.

Before, we had some support for this pattern in Pulse, but now we add support in the frontend: we identify such global function pointers, substitute their use with the actual function and remove the initializer code for such global variables.

The result is simpler code where Pulse analysis is easier to debug, and the analysis is also more performant, since fewer global variables are convoluting the state, and the method calls don't involve an extra step of calling the global initializer.

I left the Pulse support around because that's not specific to mock pointers, but works for any global function pointers.

Reviewed By: ezgicicek

Differential Revision:
D43838894

Privacy Context Container: L1122176

fbshipit-source-id: 3d3cbe9e2295b183b8ceeb26b5537dfbdeb4f979