[unnecessary copy] Revise domain to efficiently handle timestamps
Summary:
In the previous domain, `Timestamp.t` was included in `CalleeWithUnknown.t` as a field value. While
this is not incorrect in its design, introduced some duplications; for the same callee, there
could be multiple `CalleeWithLoc.t` values, e.g.,

- `{callee={copy_tgt= None; callee= foo; timestamp= 42}; loc=a.cpp:30}`
- `{callee={copy_tgt= None; callee= foo; timestamp= 52}; loc=a.cpp:30}`
- `{callee={copy_tgt= None; callee= foo; timestamp= 62}; loc=a.cpp:30}`
- ...

To avoid the duplication, this diff revised the domain as a map locations/timestamps, i.e.

`{copy_tgt= None; callee= foo} -> { a.cpp:30 -> { 42, 52, 62, ... }, ... }`

Reviewed By: ezgicicek

Differential Revision: D43881421

fbshipit-source-id: ac2501a3ce591ab9abd33bd3d50df3c1a9971abc