[clang] refine logic for skipping unimplemented features
Summary:
When we encounter an AST node for which we have no translation, we translate it
as `__infer_skip(args)` where `args` are themeselves the translations of the
sub-statements of the AST node. But, this only works when said sub-statements
are expressions. Otherwise we get a crash and the entire function is skipped.
This was the case for coroutines for example: sub-statements of
`CoroutineBodyStmt` are statements that are not expressions.

Instead, only inject `__infer_skip(args)` when all the sub-statements are
expressions.

Reviewed By: skcho

Differential Revision: D44063787

fbshipit-source-id: 78cd76aed79eeca48a7952bc92634ab4b5106571