[taint] Specify a block that is passed to a given method as a source or sink
Summary:
Now it is possible to specify the parameter of a block as a source, when specifying which method the block is passed to rather than the block name, which is what is needed in practice.

No other way is provided to specify blocks at the moment because it's not clear that it will be needed but can be added later.

Reformatted the error message for parsing, see screenshot.

Also this is only supported in `taint_initial` rather than when blocks are used in calls. Again, this is the only use case known to me, can be extended later.

Reviewed By: geralt-encore

Differential Revision: D44219563

fbshipit-source-id: be4f6ecf77f5429554a367bf388533a73ff0eac3