[ondemand][1/16] global state mechanism
Summary:
Global (mutable) state used during analysis needs to be saved and restored by
ondemand when pausing the analysis of the current procedure to go and analyse a
dependency. Currently all global state is registered in ondemand.ml.

With this diff, global state is registered at the point it is defined instead,
using a generic mechanism to register save/restore/init functions. This saves
on boilerplate code and makes it easier to add more global state safely, and
sometimes avoids exposing internals of modules just for the sake of saving and
restoring global state. On the other hand, making global state easier to add
can also be considered a negative, but overall as long as global state exists
we might as well have a safe and convenient way to define it.

We start with some global state in Pulse. The following diffs get rid of each
piece of global state one by one.

Reviewed By: ngorogiannis

Differential Revision: D44291215

fbshipit-source-id: e68037bb1cdc75ef50313f2589c13aa4ac7d0e95