[taint] Do not taint null pointers
Summary:
Once we know a value is null, then it doesn't make much sense to report a taint error on that value.

Moreover, because of how Objective-C analysis works, adding taint attributes to null pointers was causing us to miss steps in the traces.

In every method call in Objective-C, there is a split of the state, one case is that the receiver is `nil`, in which case `nil` is returned, and the actual method call is the second case.

So, when we had a method call with `b = foo(a)` , and `a` was tainted and `foo` being unknown, we had a state where `b = a = nil` and both `a` and `b` had the Tainted attribute that included both an item with the value history of `a` (which didn't include the method call `foo`, and the one with the value history of `b`, which did include the method call `foo`). Then we would report the taint error on this branch, and the value history that included `foo` was missing from the report.

So this change improves the traces for Objective-C reports.

Reviewed By: geralt-encore

Differential Revision: D44505635

fbshipit-source-id: c4f1d2bd045ec628c4b6f852802d6aaec1112dc9