[pulse] Don't report on movable/intermediate copies from sources that are coming from lvalue refs
Summary:
It is not safe to suggest moving from parameters that are lvalue references because if we move an object that is referred to by an lvalue reference, we risk leaving the original object in an invalid state at the caller, as the lvalue reference still refers to the original object. If the caller then tries to access the original object through the lvalue reference, we will get undefined behavior or crashes.

```
void bar(Obj o);

void foo(Obj& obj) {
  bar(obj); // stop reporting here
}
```
Let's stop reporting copies from objects reachable from lvalue-refs but ensure to keep reporting on rvalues. We only do this for copies into intermediates/assignment/optional because we still wanna report on vanilla copies that are fixable by adding a reference as in the example test :

```
void foo(Obj& obj) {
  auto f = obj; // still report here as the fix is to do auto& f=...
}
```

This change also means we stop reporting copies that are also `const &` parameters. I think this is fine as the fixes for these are non-trivial (as similar to this case): one needs to change the type of the functions' parameter, ensure the caller moves to the arguments as well as move inside to fix the original copy.

However, this doesn't mean that we should get rid of the `*_CONST` type of issues because we can still have local copies as follows that might be fixed more locally:

```
void foo() {
 const auto& obj = ...;
  bar(obj);
}
```
We can revisit this decision in the future if needed.

Reviewed By: skcho

Differential Revision: D44533841

fbshipit-source-id: 9325cd93f0b03dcf33e91a30efd217c4e6c24bcb