[infer/python] First attempt at keeping line information around
Summary:
Trying to get line info back from bytecode: each instruction in the
Python bytecode comes with an option line information. Meaning that if a
single line of python leads to multiple bytecode instructions, only the
first instruction will have the line information.

To try and use them as much as possible, I'm keeping the last line
information we spotted in the env, and use that whenever possible

Reviewed By: jvillard

Differential Revision: D44786967

fbshipit-source-id: 3ecc708daca5ce8a3d484437b81a8169c5ec96fd