[build] use @alert to forbid Tenv.load during analysis
Summary:
The OCaml `alert` feature allows turning using a definition into a warning or error, and control over when these are enabled.

This diff introduces a symbolic category `analysis` which can be used to group alerts over functions/values which must not be used during analysis. This alert is enabled as a fatal error in all dune libraries implementing analysers.

Reviewed By: bennostein, jvillard

Differential Revision: D44957005

fbshipit-source-id: 36f6694f84c8d2d1e6e0aed71bd62b9e02631551