[explore] fix bug where we would block on user input without displaying anything
Summary:
`infer explore` should display the list of issues and prompt the user to input
which one to display. However, in some cases the list isn't printed because
it's staying in some buffer. In thoses cases infer is still waiting for user
input but from a user perspective infer looks stuck as there's nothing printed
on the console.

The list is printed using `L.result`, which uses `F.kasprintf`. I'm not
entirely sure why but I did not manage to make it flush to the console. Use
`F.printf` directly instead. Add boxes as all output should be in some box (I
thought that could be why it didn't flush before but that wasn't enough).

Reviewed By: artempyanykh

Differential Revision: D45081452

fbshipit-source-id: d045bbcdee6b0671ef33ecc969e00ca154aa299e