[clang][pulse] Check const-member function for modification
Summary:
This diff allows us to make sure that the argument corresponding to `this` parameter of unknown functions are not modified if they have const versions available.

The reasoning for this is Clang's const-correctness.

Let's take a look at this example

 ```
void source_dispatch_const_fcn_bad(const std::set<int>& source) {
  auto cpy = source;
  auto it = source.find(3); // const find is dispatched since argument is const
}
```

 Clang dispatches `std::set::find`'s const version (which returns a const-iterator). so Infer know's that source is not modified.
However, if we called `std::set::find` on a non-const source as follows,

```
void source_dispatch_non_const_fcn_bad_FN(std::set<int> source) {
  auto cpy = source;
  auto it = source.find(3); // Infer misses this
}
```
Clang dispatches `std::set::find`'s non-const version that returns an iterator that can be used to modify the elements of the container. This means that Infer cannot know that find is a function that doesn't get its this argument modified even though we are pretty sure that source is not getting modified here.

To deal with this, we come up with a heuristic that looks up in the `ProcAttributes` of the class for const versions of the same method. If such a const version exists, then we assume that the argument corresponding to `this` is not getting modified.

To do this, this diff extends the `ProcAttributes` with whether the function is a const one and the uses that information to not modify the corresponding argument.

Reviewed By: skcho

Differential Revision: D44957511

fbshipit-source-id: 49adde1f3a2f5c5a16544fd5b70edf13b0e08339