[restart] refactor main flow
Summary:
There is a bug in the current use of the restart scheduler in on-demand: if a recoverable exception (eg a time-out) other than that used by the restart scheduler is thrown and terminates analysis, then a lock will remain in place on-disk for the procedures under analysis, messing up things and leading to non-determinism.

This can be fixed in `Ondemand` but it's also bad API design of the restart scheduler. This diff removes `lock/unlock` functions and exposes a single `with_lock ~f` function that takes care of all the tricky sequencing and finalisation, as well as accounting properly for time stats.

Reviewed By: skcho

Differential Revision: D45116740

fbshipit-source-id: b0a6ea1fbc8b63e20825495451042afc46d3493c