[db] fix canonicalize operation
Summary: Checkpointing is used before caching capture artefacts to ensure minimal size (this integrates the WAL log into the main Sqlite database and allows deleting the WAL log). We were instead checkpointing the analysis DB, which means we can lose information stored in the WAL log as it's deleted afterwards.

Reviewed By: martintrojer

Differential Revision: D45231051

fbshipit-source-id: eca3a281c62891262d7f56330ba97f6ef24f2f35