[java] sanity checks on files we add as source files from the CLI
Summary:
Users can specify what java/kotlin source files separately from a provided .jar
we are capturing in two ways:
- `--sources`: specify each file individually
- `--sourcepath`: specify a directory containing source files

In the case of `--sourcepath` we only capture paths that end in either `"java"`
or `".kt"` (notice the inconsistent inclusion of `.` in the suffix). In the
case of `--sources` we do no validation at all.

This changes the behaviour for both options to perform the same on every path:
- match it against a `".java"` or `".kt"` suffix
- verify that it is indeed a file, or at least it is not known not to be one

Reviewed By: ngorogiannis

Differential Revision: D45270056

fbshipit-source-id: 543b0253e3fc6f804d73d08da02b150d4b9ecb0e