[infer/python] Wrap booleans
Summary:
Wrap boolean into a "python" type as we can't decide yet if a value is a
Textual boolean or not, as we don't have type information.  Consider
```
def coin():
    return False

x = coin()
```

We currently can't tell that `x` is a boolean, so we wrap the Python booleans into a Textual type, and always call `is_true` on anything we want to prune.

Reviewed By: jvillard

Differential Revision: D45269125

fbshipit-source-id: 33594d98491f45c5e013300ae2d29e5ae3aa84a2