[pulse] recording invalidation should not always check validity
Summary:
It's often the case that `PulseOperations.invalidate` and variants are called
merely to record that a particular value is not a valid memory address, but we
check that the value was previously valid every time. This only makes sense for
operations like `free` or `delete`, to check for double frees, but for null
dereferences it doesn't make sense as we don't expect NULL to be valid before
invalidating it.

So far this hasn't caused a lot of issues in our tests because we often create
fresh values before assigning them to 0 then invalidating them. It's still
wrong and causes issues in future changes that are more careful with always
normalising abstract values.

Reviewed By: dulmarod

Differential Revision: D45401813

fbshipit-source-id: b46ef9798ad58d03027205d00d049331fa76afb4