[pulse] get invalidation trace from the access trace
Summary:
As written in the added comment:

```
 (* [invalidation_trace] comes from the [Invalid] attribute and may not be about the exact
    thing we are looking at but about some other value that happens to be equal to it (as we'll
    keep only one such attribute per value). [access_trace] has the most accurate info and if
    all goes well it should itself contain a sub-trace leading to how the value became invalid
    prior to being accessed (see also [PulseReport]. *)
```

With the previous behaviour the error message sometimes wrongly referred to
another value, as seen in the test changes. For example:

```
int* p = NULL;
int* q = NULL;
*q = 42;
```

We could get a message saying `q` was invalidated in the first line because its
value is equal to `p` and we'd get the invalidation line from the value's
attribute instead of from the trace of what happened to `q`.

Reviewed By: dulmarod

Differential Revision: D45402585

fbshipit-source-id: 0f7312d8df1ac315be5b6ff36b030289afed7d16