[analysis] track missing procedures and sourcefiles
Summary:
In order to support distributed analysis, we introduce missing dependency tracking during analysis. This is meant to be used in an iterative fashion, where the capture database is progressively completed to achieve an analysis run without any missing dependencies.

This diff introduces the tracking, following diffs will implement the actions of extracting a capture database and completing one using a record of missing dependencies.

Reviewed By: jvillard

Differential Revision: D45359066

fbshipit-source-id: 0c3dce0a84522f6702ad048565ec86dff302b6cd