[infer/python] Adding minimal support for classes
Summary:
Add a first minimal support for class declarations. We don't support fields or methods, really, just the class itself. In the bytecode, we don't know if a code object is a method, a function or a class, we have to infer this information from the code itself. This is why we try our best to trace this in our env (see the new `is_class` information).

A simple python program like
```
class C:
  pass
```

compiles to
```
  1           0 LOAD_BUILD_CLASS

              2 LOAD_CONST               0 (<code object C at 0x7fec74b84870, file "order.py", line 1>)
              4 LOAD_CONST               1 ('C')
              6 MAKE_FUNCTION            0

              8 LOAD_CONST               1 ('C')

             10 CALL_FUNCTION            2

             12 STORE_NAME               0 (C)
```

This first creates a function object (anonymous) storing the name "C" and the "code object C", and then call the builtin `LOAD_BUILD_CLASS` passing it this code object and the name "C", storing the result in a module variable named "C". TL;DR: it sets up an instance generator for the type C. This is abstracted away in `python_class/python_class_constructor`.

This is the kind of complexity we don't want to see (at most, we should make infer aware of `__new__` and `__init__` but not much I guess). We use this pattern to detect the declaration of class, and treat the class object C as a class rather than a function.

Reviewed By: davidpichardie

Differential Revision: D45739067

fbshipit-source-id: df0c1c34a7f782af49e2659696fca8c3c30c4180