[ondemand] fix timeout and missing-deps bad interaction
Summary:
For logging missing deps, an exception may be raised in `Exe_env.get_proc_tenv`. When this occurs after global state has been saved in ondemand, it leads to a crash due to the timer being suspended.

At a logical level, we should not throw an exception (such as `MissingDependencies`) after saving global state without the `raise` being inside a `protect` call which will restore the global state on `finally`.

Reviewed By: davidpichardie

Differential Revision: D45807245

fbshipit-source-id: b474ba568b99ebc9cae2cf96a1095e7d7c646575