[pulse] Refactor a bit and add extra logging for pulse-force-continue
Summary:
D43734567 introduced extra logic to treat function calls as unknown calls in case the original post
lacked a ContinueProgram disjunct.

Unfortunately, there were not enough pointers in the code or in the logs to understand WHERE the
unknown call comes from and WHY.

This diff does a little refactoring in PulseCallOperations.call, adds a comment explaining WHY we're
re-dispatching the call as uknown call and adds extra logging so that the log entries `skipping
unknown procedure` are not totally confusing.

Reviewed By: davidpichardie

Differential Revision: D45852569

fbshipit-source-id: b23d27434258a00ff53c86769ac3b7936476ff08