[debug] Output more state in collapsible blocks
Summary:
There have been a few more places where we would output state as-is which defeated the purpose of
having the execution log available at a glance.

I don't want to go over-board with collapsing every block, but I feel the current state is a
reasonable tradeoff between brevity and accessibility.

Reviewed By: jvillard

Differential Revision: D46029919

fbshipit-source-id: 79d352ef83c7e5859ca743b1eb82f94025111393