[hack] Add a model to support nullsafe field accesses
Summary:
Nullsafe access a-la `$a?->field` is translated using `$builtins.hack_dim_field_get_or_null` builtin. This diff adds an OCaml model for this builtin.

NOTE: We already have an OCaml model for `hack_dim_field_get` and ideally would write a textual model that delegates to `hack_dim_field_get` in the non-null case. But we can't. The problem is that `hack_dim_field_get` is special in a sense that it needs the field to be a *literal value* which is not true when we try to call it from a textual model. This is one case where we can't freely compose OCaml and Textual models.

Reviewed By: davidpichardie

Differential Revision: D46193078

fbshipit-source-id: 0a2fcb56b04d05da15a518c34a1e7c33c6828c69