Detect async functions via attributes not naming
Summary:
We previously just looked at the function/method name to see if it
started with "gen" to decide if we were calling an async, which is
clearly suboptimal. This diff instead looks up the decl attributes of
the thing we're calling. It depends on another diff to make hackc pass
these attributes through from IR to Textual. Here we pass them from
Textual through to Sil and look for them in call

Reviewed By: jvillard

Differential Revision: D46109528

fbshipit-source-id: 1ead7540568f9319f052b2ecb8aaf8b4fafc1248