[scheduling] fix bug in is_empty
Summary:
Before this it could be that `is_empty ()` was true but `next ()` could return
more work to do because `next ()` calls `fill_queue ()`. It was observed on
OpenSSL that two procedures were left in the graph at the end of the scheduling
(not completely sure how or why always these two!) with the replay scheduling
in the next diff.

Reviewed By: bennostein

Differential Revision: D46227130

fbshipit-source-id: 062569c3d0ee8b4887c2f08da1a994480337640e