[infer/python] Introduce proper return type for "next" iterator operation
Summary:
When I introduced iterators, I warned that having one builtin to "peek" if a value was available and another one to actually get the value would be difficult/impossible to implement. While working with `LOAD_METHOD`, I faced a similar situation, so I decided to try and fix the iterator situation to assess if that was easy to do.

In this diff, I introduce a new builtin type for iterators that bundle the "do we have a value to read" and "the value read, if any" in a single struct

Reviewed By: jvillard

Differential Revision: D45908971

fbshipit-source-id: 5d692e68fad95587d0b880b489942af68ee3b8c7