[infer/python] preliminary support for method calls
Summary:
First attempt at supporting method calls. Most of it is currently hidden away in builtin, to avoid its over-complex logic (see [0,1, 2]). TL;DR is that each method call introduces a test+branch because the runtime doesn't know if it is an actual method call, or a member access that ends up being a callable (like a lambda).

We'll put this logic in a model, so we can leverage the types present in the codebase to try and avoid branching.

[0] https://docs.python.org/3.8/library/dis.html#opcode-LOAD_METHOD
[1] https://docs.python.org/3.8/library/dis.html#opcode-CALL_METHOD
[2] https://github.com/python/cpython/blob/3.8/Python/ceval.c#L3423

Reviewed By: jvillard

Differential Revision: D45945174

fbshipit-source-id: 8f8bb3da64ab53ca352b69e12b5cc087d3ce141c