[pulse] canonicalise values coming out of the state on the fly
Summary:
Values, or logical variables, or `AbstractValue.t`, are pervasive in pulse.
They are stored in the stack and heap and are used as keys in the
stack/heap/attributes. However, some of these values can be equal and then the
domain is not well equipped to give a consistent view of equal values. As a
result we sometime manipulate non-canonical values which lead to
inconsistencies (we may fail to read the right attributes, for example).

There are also performance issues in the mitigations used against that, namely
eagerly rewriting memory edges to normalise them each time a new equality is
discovered. This incurs a linear scan cost plus some reallocations if we find
places to normalise.

In this commit and the next, we strenghen the invariants of Pulse so that:
- keys of the heap and attributes are always normalised (was more or less
  already the case before)
- values taken out of the domain are normalised on the fly (was not the case
  before)
- these invariants are enforced at the type level

As a result we are more correct (we were missing some normalisations before)
and (not yet in this commit) we can avoid eagerly rewriting the state and claim
back some perf.

The main challenge is how heavy a change this is. This is the first brick:
ensure values coming in and out of `AbductiveDomain` are properly normalised,
without leaking heavy type machinery to users. The following commits will
tighten the invariants by gradually removing unsafe interfaces of lower-level
modules from the high-level interface of `PulseDomainInterface`, eliminating
the last sources of non-normalised values.

This commit introduces the type machinery itself in PulseCanonValue and uses it
to make PulseAbductiveDomain safe wrt normalisation.

Reviewed By: skcho

Differential Revision: D45917442

fbshipit-source-id: f83e65ee237c644cac754c3ab5f9cddbee12c0a8