[pulse] detect same-stack-address contradictions when removing vars
Summary:
This is needed by the next commit where we'll stop normalising in stack
co-domains eagerly. The issue is then that we'll miss when several stack
variables are assigned the same address (which is inconsistent). Instead we can
detect these contradictions a different way: since we record an attribute for
stack variable addresses each time they go out of scope we can detect if we are
about to add the same address as a previous stack variable.

Reviewed By: skcho

Differential Revision: D45917429

fbshipit-source-id: df21015f5fc5ef37d801da4d4dc4d8453f6fc691