[topl] take get_reachable as a callback
Summary:
The goal is to not depend on the graph visiting functionalities of
PulseDomain.ml as these are unsafe as they are now (they do not take into
account normalization of abstract values) and will shortly move into
AbductiveDomain. However PulseTopl cannot depend on AbductiveDomain as that
would create a circular dependency. So instead make it take a callback.

Reviewed By: rgrig

Differential Revision: D46075541

fbshipit-source-id: e9bb306f9f959d6af8169db627eff433237dd78f