[shared ptr param] Revise error message
Summary:
The previous error message was like, "a shared pointer parameter is passed by value".  The intention
of which was to note that the parameter is not passed by reference, but some developers were confused
by the message with thinking, "No, it is not passed as a value. It is passed as a pointer."

This diff revises the error message more explicitly to change a shared pointer to a raw pointer.

Reviewed By: geralt-encore

Differential Revision: D46440036

fbshipit-source-id: a1e0c8b3de22cf99f461c486181cdfefe62beeaa