[topl] Don't work on an empty state
Summary:
The Topl state is a list, and most functions in `PulseTopl.mli` have the
form "preprocess, then do something for each element of the list".
Usually, "preprocess" is cheap, but turns out not always: sometimes it
traversed the whole Pulse state.

One solution is to do the preprocessing lazily. Another solution is to
add a fast path for the empty list case. This code change does the
latter. And it adds a stat (topl_reachable_calls) that counts how many
times we call a recently introduced callback (get_reachable), which was
one of the expensive preprocessing steps.

Reviewed By: skcho

Differential Revision: D46396474

fbshipit-source-id: 572709adfd07fe241bc5b91d7f7f95abe5fbfd9c