[taint] Add support for exclude_in for regex matchers
Summary:
Add support for `regex_exclude_in` for regex taint elements, like `procedure_regex`, etc.

The syntax support is added for all the elements, but the implementation of the check is only added for `procedure_regex` for now, and will add the next ones in following diffs.

This needs the location of the procedure to work which we get from the proc attributes. We also required the proc attributes in other areas of the taint analysis, so I refactored this so we only loaded once and pass it around, conveniently as an optional argument that catches both the fact that we may not have the attributes and the fact that for fields we don't need this.

Reviewed By: geralt-encore

Differential Revision: D46596509

fbshipit-source-id: 0c049c6b994132c1ac863950f63e3d2c24d3de00