[orchestrator] kill all children when the main process crashes
Summary:
If the tasks generator raises an exception and wants to crash, the behaviour of
infer was really strange as it seemed that nothing happened at all and all
workers were stuck. The user then had to kill infer to get the error message
and make infer stop. This is because the children are running and the epilogue
of the main process was to wait for them to finish. But if we haven't sent them
work to do they'll also be waiting for the main process to contact them.

The solution is simple enough: change the epilogue to kill any remaining
children.

Reviewed By: ngorogiannis

Differential Revision: D46731865

fbshipit-source-id: 6d6f4e37450f8c993ad72c9114dde56bc7cd03d6