[hack] Make Pulse propagate taint from unknown calls and fields
Summary:
Technically, handling of unknown calls (havoc or not) and fields (follow or not) are separate things
but we need both to make the examples work. The change is quite local so I hope this is fine.

1. Unknown calls: previously we'd havoc the memory reachable from an argument to an unknown
function. This would cause taint propagation to propagate taint from new fresh values rather than
from potentially tainted old values. A better fix would be T138610370
2. Follow field accesses when gathering taint improves the situation around shapes where the lack
the usual load/store to fields but instead the fields are created via a model.

Reviewed By: davidpichardie

Differential Revision: D46806045

fbshipit-source-id: 1d3e7782a3ab50f05159963d71b2f21852f1c544