[infer/python][3/3] Add more structure to global/local names
Summary:
Qualified names are now structured as a prefix + name. The prefix is the
path from the toplevel module to the definition, like
`module.class_name` and the name is the actual source identifier (`function_name`).

Since the module is always present, the prefix is never empty.

This enable us to stop mixing short names with long name as strings or textual qualified names.

Reviewed By: davidpichardie

Differential Revision: D46762021

fbshipit-source-id: 294a87a96175b8650db75514f758db05cb4e3385