[specialization] iterative communication between caller and callee
Summary:
When a caller provides additional calling context (by specialization)
to the callee, it could allow exploring more execution paths from the callee and
these paths could themself generate a need for more context. We hence add a kind of
ping-pong (with a threshold) to allow such an iterative refinement.

Reviewed By: dulmarod

Differential Revision: D46715833

fbshipit-source-id: 98caa45690296b93e00039f56765ecdd44328a3c