Lineage: add a Builtin internal edge variant for supressed Infer builtins
Summary:
Unless infer gets the `--lineage-include-builtins` option, in which case Infer they will be modelled
as usual function calls in the graph, builtins coming from the Erlang frontend would be suppressed
from the graph and, before this diff, replaced by `Direct` edges (exported as `Copy`).

Instead of using `Direct`, we add a new `Builtin` edge type instead to encode the information that
those are not standard copying edges. The graph simplification is adapted to keep Builtin edges over
Direct ones (but any other edge over Builtin).

Builtin edges are still exported as `Copy` ones in the json output. There is currently no observably
different behaviour on this new variant, but later developments will make use of the additional
information.

We also include a trivial debug printing change (print NONE vs. raising) when some invariant in
unexpectedly broken in shapes (that was discovered through an issue with builtins, but is
independently simply more robust).

Reviewed By: rgrig

Differential Revision: D46573620

fbshipit-source-id: a9cfcab292872c273c347bc5f7615f190fbf436d