Lineage: Compute Tito arguments before reducing the graph.
Summary:
Upon finalisation of a procedure analysis in Lineage, two things happen (in that order before the
current change):

1. The graph is simplified by removing temporary nodes that have only one (Direct) source or
   one (Direct) destination : upon these conditions, edge structures of the form `(A) -[Direct]-> (temp
   var) -[Edge]-> (C)` are reduced into `(A) -[Edge]-> (C)`.
2. The TITO arguments of the function are computed by walking through the (reversed) graph and
   collecting all co-reachable argument fields.

The current change makes the analysis compute the Tito argument on the complete graph first, then
only reduce it. Thus the computation has access to information hold by those temporary
variables. This is required for later developments on shape-polymorphic functions: these temporaries
could be abstract variables (summarising several fields), even if no source program variable is
considered abstract. We will need to know if a path goes through abstract variables or not.

We checked that the overall analysis is not impacted by this change.

Reviewed By: rgrig

Differential Revision: D46573621

fbshipit-source-id: 4f9ca58687610bfef3c3dd6396330370b910dc35