Lineage: first-step precise support for shape-polymorphic functions
Summary:
"Shape-polymorphic" functions are functions whose one of the return fields is obtained by direct copy
of a parameter's field, without any assumption on its structure, the archetype example being the
identify function `id(X) -> X`.

Within the analysis of this function (and similar one), no field is known for the Tito argument
`X`. Therefore the Tito map can only record that `arg0` flows into the return, and when `id` is
called on a tupled argument, eg. `{A, B} = id({a, b})`, every field of the argument therfore flows
into every field of the return, imprecisely tainting `A` with `b` and `B` with `a`.

This commit improves this situation by recording, for every Tito path, if it is "shape preserving",
meaning the source argument field path is directly copied into the return field path. Upon calling
a function, shape-preserved argument are written "in parallel" (field-by-field) into the return field
path instead of every-field-into-every-field.

The shape-preservation condition, informally, is described by "the return field path has only one
co-reachable argument field path, and this argument is co-reached by going only through Direct edges
and no abstract graph node". This is sound, but sometimes imprecise. In particular, the "only Direct
edges" condition prevents shape-preserving values that go through a function call (Summary edge),
even if that inner function preserves shape. This is planned future work. Note that this requires
enriched information, since currently there is no way to tell if those Summary edges are indeed
shape-preserving or not.

Reviewed By: rgrig

Differential Revision: D46573619

fbshipit-source-id: a021df3ac6d8e75369d50506686b613066d2d564