Support closures in summary marshalling
Summary:
Summaries generated by some analysis are marshalled, using the
[`Marshal`](https://v2.ocaml.org/api/Marshal.html) Ocaml module, so they can be shared amongst
several workers.

Prior to this diff, that mandated that these summaries contain no function/closure, otherwise the
marshalling function would raise an exception.

This diff adds the [`Closure` flag](https://v2.ocaml.org/api/Marshal.html#VALto_channel) to the
arguments of the marshaller: it allows closures to be exported as their position in the code.

This allows in particular using a lot of Core data structures in summaries, instead of having to make
a choice between a complete standard library and function-free data structures. In particular, Core
hashtables contain the hash/comparison functions as part of the structure, and do not provide any
closure-free equivalent (which could be similar so `Set/Map.Tree`).

Reviewed By: ngorogiannis

Differential Revision:
D46901041

Privacy Context Container: L1122176

fbshipit-source-id: a1d91914b131d6d7c8e99fab60819337c46769fc