Change the hashing function of erlang atom to the first 7 bytes of md5
Summary:
Erlang atoms are represented internally by a structure that has their `name` as a string field, and
a `hash` of that name in an integer one. That `hash` is what is used by most operations as it makes
the analyser job easier to manipulate integers.

This diff changes that hashing function from using (some bits of) the JaneStreet Core standard one on
strings into the first 7 bytes of the MD5 sum of the atom name.

The rationale is to make it easier to quickly generate Topl code that matches on atom
hashes. However, we do not make any promise of forward compatibility, thus this hashing function must
not be considered as production-usable but rather as a better way to do quick experiments with easily
generated hashes (see the code diff for python and shell ways of computing this hash).

Note that this "unstability" is not really a regression as neither Core nor the OCaml stdlib itself
do not make any promise regarding the hash implementation (and as a matter of fact, it has changed in
the past and it was explicitly stated that forward compatibility was not a goal).

Differential Revision: D47157305

fbshipit-source-id: 4262749d9e23f36c1f67641f1b498e55c05fe90f