Flow queries on Topl
Summary:
The idea here is to have a simple way to find data flows using different
analyses: (native taint analysis of) Pulse, Topl, and Lineage. That is,
using exactly the same config, one will be able to use either of those
analyses. This should give us flexibility.

This language of the config is minimal, because it acts as a sort of
common factor of what all three analyses can do. If one wants to exploit
a feature that is available in only one of the analyses, then one would
have to use that analysis directly.

The plan is to use Pulse by converting to a PulseTaint config, and to
use Topl by converting to a Topl property. Lineage would first needs
some path-finding capabilities to be implemented, and then it may use
the config directly.

In this commit, we have only the conversion to Topl.

Reviewed By: jvillard

Differential Revision: D45048344

fbshipit-source-id: 741208c63f3131dcf50cfc129d562a72f32d0005