[infer] Fix issue about pattern matching with method's parameters
Summary:
Because we were using `pp_verbose` to find the procedure name to match against, and in C++ compilation that string includes the types of the parameters of the method, we were matching wrong names. See example in the test that we fixed now.

I changed the code to using `pp` instead for clang methods where it seems safe to me, I don't have enough context about other languages.

Reviewed By: ezgicicek

Differential Revision: D47260651

fbshipit-source-id: cd524623d6acb20dbab95325da48d07581d8b56e