Lineage shapes: introduction of the Structure module instead of fields
Summary:
Lineage shapes used to only have one representation, a (possibly empty) set of known sub-fields. To
make this extensible to more expressive representations (such as "atom within a known set"), we
introduce a Structure module, which for now is mostly another level of indirection through these sets
of fields.

We also introduce an explicit `bottom` structure, which currently does not really have an observable
difference to vector without known fields. This makes the distinction between structures and
sub-fields (when the structure is known to be a vector) slightly easier in the code and prepares the
way for introducing other shapes (that would be less precise than bottom but more precise than
zero-length vectors) later.

Reviewed By: rgrig

Differential Revision: D47519764

fbshipit-source-id: 5083410789b3705faa69d9aeb8972fa1b8edae46