[infer/hack] do the plumbing for self in traits
Summary:
In this diff, we implement the logic behind the additional `self` argument present in static methods of traits. It works in two steps:

First, at a call site of `C::f(...)`, if method resolution finds that `f` is implemented in a trait `T` used by the ancestor `D` of C, we create a reference to D's static companion `D$static` and pass it as the additional argument to the call.

Second, when we encounter a call `__self__$static::g(...)` while being in a trait, we read the content of the `self` variable (that must be there, by construction) to get back this dynamic type, and use it to do further method resolution of `g`.

Reviewed By: davidpichardie

Differential Revision: D47514745

fbshipit-source-id: 29e807ee31667776d2811d76cebd2ca55f02c852