[pulse][refactor] Introduce ValuePath and use it instead of value * hist pair when possible
Summary:
ValuePath.t wraps AbstractValue.t * ValueHistory.t and augments it with the "access path" to the
value, e.g. if it's a variable on the stack or an edge in the heap.

We need this extra information to be able to apply taint to the ValueHistory of a particular memory
location rather than as an attribute on an abstract value itself.

NOTE: there are a few places, particularly in models, where I used `ValuePath.Unknown`. This is to
limit the scope of the change while retaining the existing behaviour. This can be improved on later.

Reviewed By: vsiles

Differential Revision: D47591787

fbshipit-source-id: 1052abbad614c9be1f917d8150c10fea5396a556