[infer/textual] Improve error messages when a function is not declared
Summary:
If a textual file is missing a function declaration, it is failing hard with the following message:

```
[%expect.unreachable]
[@expect.uncaught_exn {|
  (* CR expect_test_collector: This test expectation appears to contain a backtrace.
     This is strongly discouraged as backtraces are fragile.
     Please change this test to not include a backtrace. *)

    ("IBase.Die.InferInternalError(\"procname dummy::IntBox.f should be user-declared or a builtin\")")
    Raised at Stdlib__String.index_rec in file "string.ml", line 128, characters 19-34
    Called from Pyutils.split_left_on_char in file "pyutils.ml", line 17, characters 31-62
    Re-raised at IBase__Die.raise_error.do_raise in file "src/base/Die.ml", line 26, characters 8-56
...
```

An unwanted side effect is that code output is skipped, so it is quite complex to understand what went wrong when developing a new front-end.

With this diff, it will now report a Textual type error and go on instead of crashing and so will display the full code output and a list of type errors.

```
dummy.py, line 17, column 0: textual type error: procname dummy::IntBox.f should be user-declared or a builtin
```

Reviewed By: ngorogiannis

Differential Revision: D47712997

fbshipit-source-id: 240c4eccec606cf463598ba206049588f4fe0324