[jsonbug] Split suggestions from qualifier
Summary:
Infer's error messages (qualifier) sometimes contains how to fix the issue. Recently, we got feedback that it would we good to split the messages in Phabricator inline messages so that

- we first show the issue message
- then give details on whether it is frequent/P0/
- and finally, at the end give suggestions on how to fix the issue.

This requires splitting the qualifier that currently contains both error message and suggestion to two parts. This diff is the preparation diff that does that but extending jsonbug.atd with a new field `suggestion` which is currently empty.

Unnecessary copy issues have such suggestions. Next diff will fill them in. This will require further www changes so that we display the suggestions as well.

Will do an audit for all other issue types to fill this field too.

I find this split also a good idea since we make clear for each type how to fix it.

Reviewed By: skcho

Differential Revision: D47795611

fbshipit-source-id: 04ab0fba128d3133b9b8d725185479fed04447a5