[infer/python] revamp class constructor in a more usable way
Summary:
Python doesn't have a `new` statement like most OOP languages. The call to a constructor is similar to a regular function call, but the name of the function is the name of the class:
```
class C:
   def __init__(self, x, y):
      self.x = x
      self.y = y

# There is no c = new C(0, 1), only
c = C(0, 1)
```

We used to generate a special builtin when we could detect such cases, but this doesn't work with multiples files as we can't detect if a call is a function call or a constructor call if the class definition is not local to the file.

This diff gets rid of the special built by generating instead a reasonably working constructor function, so other files can call it as if it was a normal external function:

```
        define C(x: *PyObject, y: *PyObject) : *C {
          #entry:
              n0 = __sil_allocate(<C>)
              n1:*PyObject = load &x
              n2:*PyObject = load &y
              n3 = n0.C.__init__(n1, n2)
              ret n0

        }
```

We use `sil_allocate` to create a new instance (Python allows the dev to override `__new__` to do so, but we don't support it at the moment) and then call the `__init__` class method, if it was present. The signature of this constructor functions is derived from the signature of `__init__`.

Reviewed By: ezgicicek

Differential Revision: D47145100

fbshipit-source-id: 5cedbd5d9037d30f746afa46a37f00c21999cfe5