[infer/python][1/3] support for __init__ when a class has a parent
Summary:
When a class D inherits from C, the way it's constructor `__init__` is
declared is slightly different. The runtime encodes it as a closure
instead of a regular method, and we don't care about that.

This diff cleans up / ignores the bits of the bytecode we don't need.

In the next diff, we'll support properly the call to
`super().__init__()` which is not correct at the moment

Also, depending on the presence or not of `__init__`, we have to generate empty ones and call them correctly, if the call chain is implicit. This will be done too in a subsequent diff.

Reviewed By: ezgicicek

Differential Revision: D47791882

fbshipit-source-id: d96f9905b8ea4d08e59aa8517c55d69c21ce4971