[hack][pulse] Add FN test for shape field lookups
Summary:
It seems like for `hack_array_get` model, if we can't find a dynamic type, we end up returning 0 disjuncts, meaning that we don't analyze the rest of the program. This seems wrong.

https://www.internalfb.com/code/infer/[352998a9eb6cd6b400ea2ff89bdc88abe8770470]/infer/src/pulse/PulseModelsHack.ml?lines=424-433

Adding a test showcasing this issue. Next diff proposes a fix.

I also add log message for the exact codepoint.

Reviewed By: davidpichardie

Differential Revision: D48031946

fbshipit-source-id: e1eae8b5d58fa4c8ba1dd8d1d9a70f9d2a4105c7