[hack][pulse] special virtual dispatch for Hack function references (1/2)
Summary:
function references are transformed into singleton objects by HackC but the
corresponding __invoke function may not have the correct arity.

On a Hack program like
```
$f = Main::foo<>;
...
$f(args);
```
Hackc generates a special class `FunctionReference::Main$static_foo$curry` and put an instance in `$f`.

The calls `$f(args)` is turned into a Textual virtual call like
```
  n2: *HackMixed = load &$f
  n3 = n2.?.__invoke(args)
```
We intercept this call in `Pulse.ml` and replace it by
```
  n3 = FunctionReference::Main$static.foo(n2, args)
```
using summary specialization on the dynamic type of `n2`.

This diff is the part1 only. We need more work to pass the correct class object argument because `n2` is wrong here.

Reviewed By: vsiles

Differential Revision: D47998275

fbshipit-source-id: 88cd884e70b9117ad6d863d6f5264450292d33ca