[infer/python] Keep more type information about class fields
Summary:
Store class field types in the Env, to have more precise `LOAD_ATTR`.
One problem we currently have is top level code and class ordering:
because we process the toplevel blob first, and then the content of the
class, we don't have the new field information stored during the process
of the toplevel code. In a similar manner, depending on the ordering of
the class in a file, we might not have the information handy.

We might try to address this in a future commit but it will be quite
tricky as Python can interleave definitions in a crazy way

Reviewed By: ngorogiannis

Differential Revision: D48189114

fbshipit-source-id: af241c7d0cb06ecdded9bac164fc54bd482c0cbd