[clang] Post-process translation of calls with lambda parameters
Summary:
When lambdas are passed to a function as an argument, and the closure itself is passed, then the analysis doesn't work so well. We need to first evaluate the closure, which basically "initialises the closure object" and sets values of the captured variables as fields. So here we add a transformation

```
    f(closure) ---> temp = closure; f(temp)
```

for general function calls because they expect a pointer to a closure

```
    f(closure) ---> temp = closure; n = temp; f(n)
```

for `operator()` and `std::function::function`

because they expect the actual closure.

The first transformation will improve the analysis: f(closure) was not working properly. The second transformation will improve readability of the translation. Better to evaluate the closure first before passing it to a function.

By the way, often the frontend was doing this already, but it depends on very subtle changes whether a `MaterializeTemporaryExpr` was in the AST or not.

Reviewed By: skcho

Differential Revision: D48189083

fbshipit-source-id: e33dbf694a8e3ba15a4e5bdabe0d1349a7c165db