[frontend][clang] Translate coreturn opreand always
Summary:
Before this diff, the frontend did not tranlsate `coret_operand` when
there was some `coret_promise_call` which is about returning
`__promise` local vairable. Due to this, the liveness checker reported
false positives for the variables that were used in the operand. For
example,

```
co_return co_wait foo(x);
```

it reported `x` was dead.

To avoid the issue, this diff translates both of `coret_operand` and
`coret_promise_call` always.

Reviewed By: ezgicicek

Differential Revision: D48688728

fbshipit-source-id: b2339193d73b5143004b0f34eaeca031caff785c