LineageShape: track "variant" shapes for constant atoms
Summary:
We add a `Variant` shape that allows tracking variables whose value is one of a constant known set of
atoms. This information has no impact on Lineage right now but will be used on a later development to
precisely support maps whose keys have a variant type.

We add a maximum variant width for good measure, with a default at 20 (preliminary tests suggest that
this could be increased a lot without much impact on performance, if useful for improved precision).

Reviewed By: rgrig

Differential Revision: D47594996

fbshipit-source-id: 644569f79039f3e816aeaca6aef1241c8c15a04e