[infer/py] add support for STORE_SUBSCR and BINARY_SUBSCR
Summary:
Python bytecode to access arrays/lists/types or class defined `[]` is quite generic, so we parse the getter/setters as simple builtins, and the logic will have to be implemented in the models.

The rough idea is to dispatch on the runtime tag: if it's a class and the `[]` operator is defined, call the user  code, otherwise fail. If it's a primitive type (list, set, tuple), call the relevant std code. Models will be implemented in a future diff

Reviewed By: ngorogiannis

Differential Revision: D48825678

fbshipit-source-id: d7b5526c0cd8ea859b5269d0ee5d3acd82cb92d5