[uninit] Add Initialized attribute to initialize value inside callee
Summary:
Before this diff, there was only `Uninitialized` attribute and when an address was written in a
callee, i.e. `WrittenTo`, it removed the `Uninitialized` attribute. While this worked on most of the
cases, there was a corner case where the initialization is done in a low-level.

```
void init_int(int* p) {
  p[0] = 42;
}

int foo() {
  int x;
  init_int(&x);
  return x;
}
```

* Pulse knows `p[0]` is written in the callee, but not `*p` is written. In Pulse's abstract state,
`p[0]` and `*p` are not the same address unfortunately.

* When `p[0]` is written, Pulse tries to remove the `Uninitialized` attribute from `p`
conservatively.  However, this does not help in the above case, since the paramter `p` does not have
the `Uninitialized` attribute initially.

To avoid such false positives, this diff introduces `Initialized` attribute, then passes it to
callers.

Reviewed By: dulmarod

Differential Revision: D48908109

fbshipit-source-id: 1cae942577ac8bcaab11ec0065d2e4694e8ff3d6