[hashnorm] align record normalization with that of tuples
Summary:
Normalizing variants involves normalization of (inline) tuples and records. The current implementation of tuple normalization is easy to reuse for inline tuples in sum types.

The current record normalization is not usable with inline records (`Constructor of {a:int}`) because inline records are not free-standing types and thus we cannot use `{t with a=5}` notation.

This diff aligns tuple and record normalization code: it destructs the record, normalizes those fields that admit normalization, and stores the results in primed variables. Then the record is constructed explicitly.

Thus, for `Location.t` we have
```
type t =
  { file: SourceFile.t  (** The name of the source file *)
  ; line: int  (** The line number. -1 means "do not know" *)
  ; col: int  (** The column number. -1 means "do not know" *)
  ; macro_file_opt: SourceFile.t option
        (** If the location is coming from macro expansion, the name of the file macro is defined in *)
  ; macro_line: int  (** If the location is coming from macro expansion, the line number *) }
[@deriving compare, equal, sexp, hash]
```
and the generated code is

```
[*ERROR**][1856298] +let normalize t =
[*ERROR**][1856298] +  let { file; line; col; macro_file_opt; macro_line } = t in
[*ERROR**][1856298] +  let macro_file_opt' = SourceFile.Normalizer.normalize_opt macro_file_opt in
[*ERROR**][1856298] +  let file' = SourceFile.Normalizer.normalize file in
[*ERROR**][1856298] +  if (phys_equal macro_file_opt macro_file_opt') && (phys_equal file file')
[*ERROR**][1856298] +  then t
[*ERROR**][1856298] +  else
[*ERROR**][1856298] +    { file = file'; line; col; macro_file_opt = macro_file_opt'; macro_line }
```

Reviewed By: skcho

Differential Revision:
D49013981

Privacy Context Container: L1122176

fbshipit-source-id: fc66deea0263db81a29fe385042eb3fda6f9ab6b