[frontend] Fix assert false in list-initialization translation
Summary:
During the translation of list-initialization, this diff checks if the super class is found from tenv. If it is not, it does not try the translation of the initialization for the super class fields. There was `assert false` due to this.

Background: Current frontend design for list-initialization is not ideal in terms of that it requires all fields of transitive super classes, which may be unavailable from a compilation unit. Instead, it could be a function call of the list-initializer that gets `std::initializer_list` as an argument.

Reviewed By: ngorogiannis

Differential Revision: D49151177

fbshipit-source-id: 00a87f537b90b5e7e07a54b18d5d9504907a9e47