Lineage: infer map shapes.
Summary:
We add support for map shapes in LineageShape. This is done by processing constant atom keys as if
they were record fields. A map only used with atom keys will thus be processed exactly as if it was
a record with those keys as fields.

When a map is used with a key that cannot be statically inferred as an atom value, this
representation mandates that we consider every past and future binding present in the map to be
accessed. We therefore add a specific `map_value` shape information to the records, which, when
present, forces every *map field* of the record to have that shape for the associated value. Note
that this does not affect usual record fields, eg. for values that can be either a map or a record.

Lineage does not fully use this information yet: it will see the map fields as individual cells, but
does not implement precise support of the map functions yet.

Reviewed By: rgrig

Differential Revision: D49066704

fbshipit-source-id: 514481f9ab649a5b1fbda243f818d28c9fc79963