[infer/py][1/N] Start to replace usage of `L.die` with proper result reporting
Summary:
Stop raising exceptions on errors, report the message and let infer live. For now, we never "die" at all. The last diff in the stack will introduce a call to `exit` in a more central location.

For this stack, the pattern is to replace all `Logging.die FooError "...."` with `Logging.foo_error "..."; Error ()`

Reviewed By: ngorogiannis

Differential Revision: D49309011

fbshipit-source-id: fac83f7ad937feb0c4a7db1d286544ce5f3e9418