[infer/py][2/N] "toplevel" function of PyTrans moved to Result
Summary:
Stop raising exceptions on errors, report the message and let infer live. For now, we never "die" at all. The last diff in the stack will introduce a call to exit in a more central location.
For this stack, the pattern is to replace all Logging.die FooError "...." with Logging.foo_error "..."; Error ()

Reviewed By: ngorogiannis

Differential Revision: D49309007

fbshipit-source-id: 86f208243e430c6792bdd892d2bcc34689590aaf