[infer/py][3/N] Removing L.die from PyClassDecl
Summary:
Stop raising exceptions on errors, report the message and let infer live. For now, we never "die" at all. The last diff in the stack will introduce a call to exit in a more central location.
For this stack, the pattern is to replace all Logging.die FooError "...." with Logging.foo_error "..."; Error ()

Reviewed By: ngorogiannis

Differential Revision: D49309012

fbshipit-source-id: 3eba9c732ee93203a44a171cf731cfc1c6c475a2