[Doli] Removing the Doli project
Summary:
The purpose of this project was to enable writing models in textual form.

The first part involves defining function signatures to intercept function call, following the syntax of the source language.
We are removing this first part here, as it is not in use and not fully completed. Instead, we can continue using
the DSL in ProcnameDispatcher.mli.

The second part of the Doli project is about writing body of models in textual form. It carries on in the Textual
project.

  viking_goodbye

Reviewed By: geralt-encore

Differential Revision:
D49537532

Privacy Context Container: L1208441

fbshipit-source-id: 0b2f3336b7e53aaa52b134b174de933cb65e0b93