[hack] adding a new builtin get_lazy_class
Summary:
This builtin is used on Hack expression like C::class. Like lazy_class_initialize, it takes
a type as argument. Its model is the same for now, but we we may want to distinguish them later, because
they don't have exactly the same semantics in term of HHVM errors.

Reviewed By: artempyanykh

Differential Revision: D49579240

fbshipit-source-id: d94e727968ca8f73d200d0ed8ab45bfda5a594d2