[test] Do not overwrite test file in clang_compilation_db_* tests
Summary:
This diff avoids overwriting test files in the clang_compilation_db_escaped/realpath tests. Both Makefiles of the tests are running below:

```
$(COPY) "../codetoanalyze/path_with_spaces/hel_lo.c" "../codetoanalyze/path with spaces/hel lo.c" \
|| diff "../codetoanalyze/path_with_spaces/hel_lo.c" "../codetoanalyze/path with spaces/hel lo.c"
```

If they run in a bad timing, at the same time:

* one is trying to copy the target file (write)
* one is trying to diff the target file (read)

it can introduce the `diff` running failure. This diff tries to avoid the overwriting at the copy to avoid the timing issue.

Reviewed By: geralt-encore

Differential Revision: D49597444

fbshipit-source-id: d6ef35b9246230dbaa0b32b8a30ee504bf1c4c4e