Lineage: Use WTO-based iteration strategy
Summary:
Infer provides two iterations strategies to walk through a CFG, `RPO` (Reverse Post Order) and
`WTO` (Weak Topological Ordering, from Bourdoncle's work). The goal of the iteration strategy is to
ensure a fixpoint is reached (every node data has been computed with the most up-to-date information)
and to avoid useless recomputation (on `A -> B <- C`, `B` should be computed last to avoid doing it
twice or thrice).

Switching to `WTO`, which aims to be smarter especially in the presence of loops, happens to make
Lineage faster (~30% observed increase on a sample codebase).

Reviewed By: rgrig, jvillard

Differential Revision: D49461102

fbshipit-source-id: 898a9058d9fcf2762c4eef9569c6ccd59aefdd46