[infer/py][2/3] Introduce support for function with default arguments
Summary:
In this diff, we generate specialised versions of a function that has default arguments. We can do that since the runtime expect every argument with a default value to also have a default value.

As in the previous diffs, objects used as default values for argument won't be dealt with correctly.

Reviewed By: davidpichardie

Differential Revision: D49494734

fbshipit-source-id: 63f72b1bf7c946767b6e5b1d6029e7cd790464a2