[infer/py][IR] Introducing an IR between Python bytecode and Textual
Summary:
This diff looks big, but doesn't introduce much novelty besides the actual IR between Python and Textual.
The goal of this IR is to make the translation to Textual less complex, by getting ride of the first two complex pass:
- naming things correctly, without ambiguity
- reconstructing the CFG

PyIRTest.ml is just a duplicate of PyTrans.ml (without all the modules) to make sure I reach the same level of support that PyTrans currently provides.

PyIR.ml has inlined comments to help the reader. For now, it supports most of what PyTrans currently support, without the support for default arguments, or the CFG reconstruction. They will be part of the next diffs in the stack

Reviewed By: ngorogiannis

Differential Revision: D50083408

fbshipit-source-id: 3d671398ad8c49ec66a11f4de0a5504c1b4c5bac