[sqlite] clear bindings after execution
Summary:
The OCaml bindings to SQLite ensure that when a value is bound to a query parameter it is copied to a value that is memory-managed by SQLite. This means that any value bound occupies memory until the bindings are cleared, and this memory lies outside of the OCaml heap. When writing out summaries, this memory can potentially amount to a lot.

This diff clears bindings and resets the prepared statement right after query execution, meaning that memory for parameter values is released immediately.

Reviewed By: dulmarod

Differential Revision:
D50126133

Privacy Context Container: L1208441

fbshipit-source-id: ea7e9dfb2272ca78a4db935ef9a277571663d112