[pulse] count summary contradictions and total number of summary disjuncts
Summary:
This gives a rough idea of when we probably did some useless work because we
were working on an unsatisfiable disjunct right until we tried to make a
summary out of it. If the disjunct limit had been reached for that procedure
then we also missed paths due to that.

Reviewed By: skcho

Differential Revision: D50225984

fbshipit-source-id: 202bb9874120613dc3c959118b20abd0dbf4181d