[pulse] stop recording MustBeValid in the post
Summary:
Now that precondition values are normalized we don't need the hack that
consisted in storing MustBeValid attributes in the post (D41438420).

This gets rid of a bunch of junk latent issues about values that cannot
actually be reached from the precondition. Usually it went like this:

```
foo(x) {
  // *(&x) gets value v_x
  y = unknown(x);
  // v_y = unknown(v_x) is stored in the path condition
  *y = 42;
  // MustBeValid(v_y) is recorded in the post even though v_y is not in the PRE
  if (y == NULL) {}
  // the null test creates a latent invalid access on v_y
}
// when creating the summary we don't notice v_y is unreachable because the
// unknown(v_x) = v_y fact keeps it alive (this could also be addressed as a
// separate fix)
```

Now we don't record `MustBeValid(v_y)` in the first place.

Reviewed By: skcho

Differential Revision: D50229927

fbshipit-source-id: fb3a9e893fbc224fce5c86a63d87589cab2b2a9d