[pulse][interproc] no need for the full power of PulseInterproc.visit when applying the post
Summary:
`visit` does some extra work compared to `subst_find_or_new` to figure out if
we discovered new aliasing. This makes sense in the precondition, where the
same value could appear several times and will have to be matched against
potentially different values in the caller, but when applying the post this
cannot happen as all remaining values are fresh as far as the caller is
concerned.

Reviewed By: dulmarod

Differential Revision: D50087877

fbshipit-source-id: 0bad4811ab48eab5063f5593a894672a644f244a