[infer/py][IR] Starting to keep track of exception handlers info
Summary:
To correctly skip/parse exception handlers, we need to remember some information to tag blocks that might raise an exception, and which handler will be called in this case. This diff starts to store this information but does nothing with it at the moment.

Minor changes to the `*_FINALLY` opcodes are done to correctly scope this information

Reviewed By: ngorogiannis

Differential Revision: D50269403

fbshipit-source-id: 56821c351f7554f2b432f286000a4015e50f739a