[specialization] adding a FN test that demonstrates too strict requirements on specialization
Summary:
the summary computation is too much black/white and we need a grey option.
When a function foo needs dynamic informations about the caller site, either we don’t have
them and we compute a default (black) summary (as Infer was doing before adding Specialization),
or we have all required dynamic informations and we compute a very precise summary for foo.
But if we have part of this dynamic informations, we can still compute a (grey) summary that is
better than the default (black) one. I will implement this idea in a next diff.

Reviewed By: skcho

Differential Revision:
D49638288

Privacy Context Container: L1208441

fbshipit-source-id: 0f90471fcccd69e32084b3b657e2b1979609254e