[dead store] Ignore the variables for structured bindings
Summary:
This diff ignores the variables for structured bindings in the dead store checker. Currently, the
frontend does not handle the structured binding correctly, esp. when it is done for reference:

```
auto& [x, y] = ...;
x = new_value; // dead store FP
```

For example, `x` and `y` are bound as references, but they are used as if non-reference value in SIL, which
introduced dead store false positives.

Reviewed By: jvillard

Differential Revision: D50643042

fbshipit-source-id: 7896ff0036e68303a2ed15b7be38fa25bf064f65