[taint] Make the MustNotBeTainted attribute a map to avoid duplications
Summary:
Previously this attribute contained a set, and so it accumulated sometimes many elements. But we only need, for each value, one sink information per kind, since that's all we want to report. So a map kind -> sink makes more sense.

In the case of sinks that are derived because of being the field of a sink, we want to make sure that if we have both an original and a derived one, we keep the original. This happens when we add a new element, we use the compare function with prefers Basic over FieldOf over PointedToBy.

I expects this will bring time and memory wins.

Reviewed By: jvillard

Differential Revision: D50413327

fbshipit-source-id: aaf11c8599710ab2edaa98934180d3e5f21fbd41