[infer][taint] Apply sanitizers on proc entry (initial taint) to avoid FPs
Summary:
It's possible that a taint sanitizer may in turn call taint sinks. Then, whenever the sanitizer is
called with tainted values, it will report taint errors. This is arguably not what we want. A more
reasonable behavior is whenever something flows into a sanitizer, it gets sanitized regardless of
what the sanitizer does internally.

This diff applies sanitization at proc entry during initial taint which is enough to improve the
behavior of sanitizers.

(See the previous diff for a specific example)

Reviewed By: geralt-encore

Differential Revision: D50788213

fbshipit-source-id: 8187ad04a8d00db94d2becdee1046fff85fa9a36