[unnecessary copy] Join non-disj astate when dropping disjuncts
Summary:
HACK: When it drops disjuncts due to the disjunct limit, it also loses some information on the
non-disjunctive abstract state, i.e. [pre_non_disj], which resulted in false positives. To mitigate
the problem, it joins [pre_non_disj] when needed, while it is incorrect to join without running
[exec_instr] precisely speaking.

Reviewed By: jvillard

Differential Revision: D50638427

fbshipit-source-id: ace6c1982984af92b270de1b88492b741da6a8e1