[unnecessary copy] Address unknown source as unowned
Summary:
The unnecessary copy checker has suppressed the issue when the source is *unowned*. Because, when
the source is unowned, we should not suggest `std::move` it.

The problem was that the unowned-ness was determined in a passive way, i.e. when the source was
unclear, it addressed the source as "owned". Thus, this introduced false positives.

In this diff, it changes unclear source is addressed as "unowned".

Reviewed By: jvillard

Differential Revision: D50845216

fbshipit-source-id: 812ea40f0214ee0b9ba25f078b9d250d62ae5f99