[infer/py] extend support for RAISE_VARARGS
Summary:
Adding minimal support to re-raise exceptions ([RAISE_VARARGS 0]).
For the time being, we just introduce a builtin function to keep track
of this behavior, but we really should use the 3 values left on the
stack from the exception handler to deal with this correctly.

Unclear if there's any need to do so at the moment. I left a comment in
[raise_varargs] in case it becomes a future need. The options I see are:
- pop the 3 values and pass them to the builtin fun: unclear if we can have anything else on the stack at that point
- duplicate the 3 values and pass them to the builtin fun: same, unclear these are exactly/the only values we need to unwind.

Reviewed By: jvillard

Differential Revision: D50919227

fbshipit-source-id: 9d36a6c6e7d6d91e25892f1d6017e1025d52a1db