[pulse][formula] make Formula.t private
Summary:
This will become useful in the following diffs. We'll want to do something
extra when adding or removing elements to each domain so making the type
private lets us control all the places where that happens ("something extra"
will be: adding all the variables from the term into an occurrence list to be
able to do incremental normalisation).

This accidentally fixed a bug in intervals where we were using the initial
version of `phi.intervals` instead of the current version during a fold
operation in `Intervals.and_binop`, hence the unit test change.

Reviewed By: skcho

Differential Revision: D50556742

fbshipit-source-id: 86712379e92086f8d6a9d3511d3ad5c05b098a32