[pulse][formula] improve intervals based on previous knowledge
Summary:
When we add an interval fact to the formula `x -> interval` (eg `x∊[i,j]`) it
could be the case that we already knew another interval fact about the same
`x`. In that case we should conjoin the information instead of discarding the
existing knowledge.

It turns out the `intersection` function on intervals was pretty bad so replace
it with a better implementation based on existing code.

Reviewed By: skcho

Differential Revision: D50556730

fbshipit-source-id: 1532bf73362db9897fc03dc004c3a64b1b1433f7