[pulse][formula] detect new atoms during incremental term normalisation
Summary:
Before we were calling `Atom.eval_with_normalized_terms` during equality
propagation in term_eqs but only to detect `Unsat`. We can use its full power
to detect if some terms have become atoms, eg `x≠0 -> y` and `y=0` should
remove the `x≠0` *term* and add the *atom* `x=0`.

To make this complete we also need to detect new atoms when `y≠0` is added but
currently only equalities are propagated so this is done later in the stack.

We also need to detect that `y` is in the *range* of term_eqs in the example
above, which we currently don't do either (next diff) so the only changes for
now are when the variable appears on both sides of the equality, eg `x≠0 -> x`.

Reviewed By: skcho

Differential Revision: D50699923

fbshipit-source-id: 54f824bac03704383a5b598196ffbd3b0c6c967c