[pulse][formula] do not save redundant information in summaries
Summary:
A fact like `v1 = 0` is currently stored in three places in the formula:
- in linear equalities as `v1 -> 0`
- in term equalities as `0 -> v1`
- in intervals as `v1 -> [0,0]`

This is needed for completeness but when storing the formula in a summary we
can be more careful and store it only once since the caller will redo the work
of propagating the fact to all appropriate domains. We favour term equalities
for simplicity as they are more general.

Reviewed By: skcho

Differential Revision: D50843685

fbshipit-source-id: c620260d254c3e11f86bc8a04b93efde483e73dc