[frontend][cpp] Revise translation of switch and try/catch
Summary:
This diff revises the translations of switch and try/catch:

* Before this diff, the `switchStmt_trans` returned a control with its
leaf node as `trans_state.succ_nodes`. However, this is incorrect,
because the leaf node should include the nodes that are generated by
the translation.

* In the try/catch translation in C++, it has been designed to insert
the excption edge only at the end node of the try's body,
a.k.a. `try_exit_node`. The problem was that it tranlsated the try's
body first, then connected it with the `try_exit_node` in a broken
way. Instead, this diff passes the `try_exit_node` as
`trans_state.succ_nodes` when translating the try's body, so that
their connections are done inside the translation naturally.

Reviewed By: artempyanykh

Differential Revision: D50984754

fbshipit-source-id: 0c7d953e5f158570d5ce47498c62d332a031f433