[hack] new Transitive Access issue
Summary:
: we add a new Pulse analysis that track spurious loads transitively reachable from specific functions.

Main limitation: we raise an alarm even if the result of the load is not used. It can be enhanced later with a specific value attribute, if we really need this granularity.

The targets and source will be customizable through a Json file in the next diff.

Facebook: this new client analysis detect Falco listeners that contain (or transitively call a function that contains) a global access.

Reviewed By: artempyanykh

Differential Revision:
D50644590

Privacy Context Container: L1208441

fbshipit-source-id: f277d1de74d76748a534c6abe280ae3cf5ecc0ff