[taint] Cut off taint traces as soon as the sink is hit
Summary:
One _problem_ with taint traces is if there's a chain of callees all matching a particular sink
kind, the trace will go on and on as long as the value flows through the chain. This can get very
confusing and a shorter taint trace could be an improvement.

This diff scrubs the callee_summary of taint attributes if the callee is itself a sink (taking into
account the kinds of sinks). This results in traces getting cut off early as soon as the sink is
hit.

One possible improvement is to scrub the taint from the summary when the summary is created rather
than on every application, but this requires cutting the dependency cycle between modules, which can
be done separately.

Reviewed By: geralt-encore

Differential Revision: D51261889

fbshipit-source-id: cd2b2672fe4534a807789394afb046aa2125e38b