[reference-stability] Do not report issues on double lookups
Summary:
This keeps track of the last lookup of the map with an attribute. When there is a double lookup on `operator[]`, do not report a reference stability error. This covers the big majority of existing false positive cases. It does introduce some false negatives, see example.

Note: This can probably easily be adapted to issue a warning on double lookups, a relatively small performance issue, but possibly significant over all services.

Reviewed By: skcho

Differential Revision: D51263391

fbshipit-source-id: e13bb206ee2f5e77420734253abbf32e2d27b9c6