Lineage: disable field sensitivity on huge procedures
Summary:
The last developments on Lineage have made the analysis more precise at the cost of sometimes
painfully degrading its performance, making a run on a typical Erlang codebase more than 10x slower.

The issue have been pinpointed to accelerating performance hits on large functions: precision
features used to be manageable through simple sensitivity uniform limitation but it seems to no
longer be the case.

We solve this by disabling field sensitivity altogether for big functions through the
`--lineage-field-max-cfg-size` option. It defaults to the same value as `--pulse-max-cfg-size`,
considered as a good estimate for "we won't put too much effort into that one".

Note that Lineage itself usually still needs to run on large functions, as they may hide either
sources or sinks that we want to be aware of.

Reviewed By: rgrig

Differential Revision: D51315857

fbshipit-source-id: efbc23160f2e4f99e7083181fba0cca781512c43