[clang] Add cleanup attribute to local variables
Summary:
This adds the cleanup attribute to the `var_data` of local variables.

> GCC’s C compiler allows you to define various variable attributes. One of them is the cleanup attribute (which you can also write as __cleanup__) which allows you to define a function to be called when the variable goes out of scope (for example, before returning from a function). This is useful, for example to never forget to close a file or freeing the memory you may have allocated.

Example use is something like

```
char *buffer __attribute__ ((__cleanup__(free_buffer))) = malloc(20);
```

In the AST we get the cleanup attribute, but not the function attached to it, but we will assume it will be about freeing memory in the next diff (we will avoid reporting a leak in that case) which should be the most common case.

Reviewed By: skcho

Differential Revision: D51498997

fbshipit-source-id: 8c58293d97137cc432f6c41d2320541560cf403c