[pulse] Refactor DynamicType attribute
Summary:
DynamicType attribute contains the type but also the source file which is needed for Objective-C classes. The way the code was before had the main functions to deal with dynamic type ignore the source file, and have separate functions deal with this.

However, this leads to code duplication given how the attribute get/set methods get added through the whole Pulse stack. More importantly, I'm planning to add another element to this attribute: the closure procname to do specialization on closures, so the previous approach wouldn't scale in this case.

Now the callsites receive all the data from the attribute and deal with it.

Reviewed By: davidpichardie

Differential Revision:
D51588549

Privacy Context Container: L1227730

fbshipit-source-id: d4299979fdd53db9717dcff2cc29fe52893f9113