[pulse] add ability to track several precondition cells
Summary:
This fixes a previous limitation of the cell ids mechanism: a value history may
depend on several memory cells in the precondition but we would only remember
at most one. Now we remember a set of them and prepend the corresponding caller
histories as appropriate.

Also document a bit more of `ValueHistory`.

Reviewed By: skcho

Differential Revision: D51157587

fbshipit-source-id: a075a5d4e36a1dfc6c5b651b8914255bfd944a69