LineageShape: promote boxing fields as a special field label
Summary:
The Erlang frontend models integers and atoms as SIL structures. Integers thereby have their actual
value boxed into a `value` SIL-integer field, and atoms have a hash of their name into the `hash`
SIL-integer field (they also have their `name` in a string field, although most operations only use
the hash).

So far, Lineage shapes used to be unaware of these fields until the very end. That is, they were
processed as any structure field throughout the shape analysis, only `fold_cells` would explicitly
remove them from generated field paths to avoid cluttering the output with implementation
details. Concretely, when `X#foo` was known to be an integer with a `value` field, LineageShape would
only generate `[foo]` as a known field for `X`.

This had an unfortunate consequence when `X#foo` could be either an atom with `hash` or an integer
with `value`. The shape analysis would indeed generate two identical `[foo]` field paths for `X`,
corresponding to the removal of two different boxing fields. Upon assigning something to `X#foo`,
this would therefore generate two edges and have some undesirable interaction with the one-source
rule shape-preservation of arguments.

We fix this by using a specific `BoxedValue` field that will represent `hash` or `value` -- that is,
any field that would ultimately be removed from the output of `x#thefield` to be identified with
`x`. Since this is a unique variant, this will prevent the duplication of paths due to boxed fields
removal (there will be only one such path).

**Expected impact.** Most of the time, this will have little observable effect on Lineage. Some edges
will be generated redundantly, but the Json export will deduplicate them. However, since the shape
preservation of arguments for inter-procedural precision relies on the property that each node on
a precise path may have only one predecessor, this will improve the precision on the specific case
where a variable may be either an atom, an integer or a tuple (as known by information not on the
argument-to-return flow), and is copied directly from argument to return. The test plan shows such an
example.

**Side note.** There is also a slight awkwardness in that the analysis will be aware of those boxing
fields upon pattern-matching on integers/atoms, but not when creating them. Although this would
appear in shape environment debug printing, this should have no observable impact later. We do not
change this behaviour to minimise the current diff (this can always be considered later).

Reviewed By: rgrig

Differential Revision: D51622864

fbshipit-source-id: 67d6fbe8aa9db17c3cf31540f453e19ab81e6471