[pulse] Replace calls to lambda copy ctor and to
Summary: Lambdas are special immutable objects that don't need to be copied. So replace the calls to the ctors with assignment and removed the ctors. We get specialization working now, see test. Moreover, debugging examples with lambdas becomes simpler. This also saves storage for the ctor methods, and perf for analysing those methods.

Reviewed By: ngorogiannis

Differential Revision: D51983877

fbshipit-source-id: a73469b0839d23ee02375866ff33f51023456813