[pulse] record uninit FP
Summary:
returning structs by value is broken in pulse: `struct foo p = struct();` and
`struct foo p; p = struct();` have different behaviours, the latter being
incorrect. This is possibly a frontend issue since the code emitted by the
frontend for the later case assigns to `*&p` whereas pulse expects the fields
to be in `&p.x` etc.

Anyhow, this is surfaced in an unrelated test so changed that test to the good
case and recorded the FP separately.

Reviewed By: ngorogiannis

Differential Revision: D52294276

fbshipit-source-id: 0c8e233fb5a332872c8acc2987a71f1806c9b784