[uninit][hack] Add a false positive due to confusion of string
Summary:
This diff adds a false positive of the uninitialized const checker.

Details: When there are multiple enum fiels in the class, they are initialized via `cinit` as
follows.

```
enum E: int {
  S1 = 1;
  S2 = 2;
}

abstract final class C {
  const E F1 = E::S1;
  const E F2 = E::S2;
}
```

```
void cinit($constName) {
  if ($constName == "F1") {
    return E::S1;
  } else if ($constName == "F2") {
    return E::S2;
  } else {
    assert(false);
  }
}

void sinit($this) {
  // initialize F1
  $name = "F1";
  $v = cinit($name);
  set_static_prop($this, $name, $v);

  // initialize F2
  ...
}
```

The problem was that the string value is analyzed as PulseAttribute in Pulse, unlike integer value,
thus it was incorrectly overwritten at function summary application. In the above example, after the
function call of `cinit`, `$name` was changed to `"F2"` incorrectly.

Note that the summary of `cinit` may have three disjuncts (in theory):

* one disjunct with `$constName == "F1"`,
* another disjunct with `$constName == "F2"`,
* and the other.

Reviewed By: davidpichardie

Differential Revision: D52629396

fbshipit-source-id: f85da7cc4275f766618430f7173ac6cb96003568