[objc][frontend] Fix translation of special message expression
Summary:
In the frontend translation, `class`, `alloc`, and `new` message expressions are translated
specifically. The problem was that it missed the translated sub-expressions in the result. For
example,

```
Class c = [(sub-expression) class];
```

here the `sub-expression` did never appear in the translated result, thus which introduced a false
positive of dead code.

This diff fixes it to include the translated code for evaluating the `sub-expression`s.

Reviewed By: ngorogiannis

Differential Revision: D52662398

fbshipit-source-id: 915b62574d9ba4dca06d62c6ffe14bbaed12b5f6