[pulse] Revise F14 iterator operator++ model as unknown call
Summary:
This diff revises the F14 iterator's `operator++` model as unknown call, so that the iterator points
to a different value after `++`. Before this diff, the semantics was `skip`, which introduced a FP
of use after delete because the iterator pointed to the same element in the loop deleting elements.

Reviewed By: rgrig

Differential Revision: D52659360

fbshipit-source-id: adcf71792f8835688cc1a9f3f72dec51d9786a1c