[transitive-access][test] multiple accesses in a same function
Summary:
this test shows a lack of signals related to dropped disjuncts.

It seems the problem is not really the transitive-access analysis itself. But the way we report on traces. It seems we won't report 2 traces if they start from the same location, while they do not end the same way.

Reviewed By: geralt-encore

Differential Revision:
D52722206

Privacy Context Container: L1208441

fbshipit-source-id: f4e0669a3ddce20116ba9496abca5d4de4e4a094