[pulse] new non-disjunctive field to record calls resolutions
Summary:
given a specific function summary, we record all the
calls that were resolved during its analysis and transitively during the
analysis of its transitive callees.

In this first version, we only record closure resolution. The key we use for the recording is a bit naive: the source location of the call site. This can lead to overlap between several resolution. We will improve that later.

Reviewed By: skcho

Differential Revision:
D51946210

Privacy Context Container: L1208441

fbshipit-source-id: f57108586af7ceeb903d33f958b3b6f2f52651c3