Lineage: light refactor of tito extraction
Summary:
We change the type of TITO extraction to make it less ad-hoc/implementation-leaking. Instead of
explicitly taking the already computed return paths it will now take the procdesc and shapes and
compute them internally.

This will make easier and more natural to eg. change the implementation to go forward from arguments
instead of backward from the return. With this diff the API won't change.

Reviewed By: rgrig

Differential Revision: D51716500

fbshipit-source-id: 54de4e80acf334e83c31a38413fa07b5e1ad83ca