Lineage: use ocamlgraph for the full graph structure
Summary:
The Lineage analysis generates a "partial graph", which essentially is a list of edges, for each
traversed instruction. Once it has finished analysing a procedure it consolidates these partial
graphs and some incoming and outgoing edges into the complete Lineage graph of the procedure.

Before this diff the complete graph was another list of edges, basically the concatenation of all the
partial graphs for each instruction. Then it was transformed back-and-forth as an adjacency
representation in order to compute both the graph simplification and the TITO summary of the
procedure.

We switch from this edge list representation to a proper graph structure, that doesn't require these
transformations and allows for better abstraction of the relevant API. We currently use ocamlgraph as
a ready-to-go graph implementation, leaving some opportunities to change it into a more optimised
structure should the need arise.

We do not change the partial graph structure as the current architecture (fast generation of local
edges then consolidation in the end) is the fastest and simplest we tried to eventually build the
graph.

The two main graph traversal algorithms, namely simplification and TITO extraction, remain mostly the
same. They just use the new graph API to fold over successors and/or predecessors rather than going
through the adjacency list transformation. Having a more structured graph could make it possible to
define better implementations later on (for instance, TITO generation could likely work over a single
graph traversal rather than proceeding for every return field).

Reviewed By: rgrig

Differential Revision: D51950373

fbshipit-source-id: 14c1b98937b1db28c331d50f6a6c5e511d5431fd