[transitive-callees] improving issue.exp format
Summary:
using absolute line number is not great in issue.exp files because
they generate bigger diffs. Here we prefer to keep both relative and absolute location
in the JSON format. This extra information is only generated for experiments purpose so
we don't care too much about the size of the JSON file.

Reviewed By: geralt-encore

Differential Revision: D52840408

fbshipit-source-id: 8f51c47e99918dc9219f15e82bd8d994b9cbc794