[analysis] set max_jobs to 40 and remove worker number heuristic
Summary:
Previously, it was believed that Infer analysis does not benefit from running one worker per CPU thread, but one worker per CPU core (estimated as half the CPU threads).

Experiments show that another bottleneck is at play (possibly I/O) and that performance is maximal at 30-40 workers on 40-core systems as well as on systems with >> 40 cores.

Thus, we remove the previous heuristic which set the number of workers to half the CPUs (assumed threads) in favour of directly using as many workers as CPUs or threads, and sets `max_jobs` to 40, so that we normally never use more than 40 workers.

Reviewed By: skcho

Differential Revision:
D52899849

Privacy Context Container: L1122176

fbshipit-source-id: abb0a182cdcf7406ecc69a6efa2e5596afa22272