[pulse] Option to skip argument havoc
Summary:
Pulse has some heuristics about when to havoc an argument of a call to
an unknown function. For example, in C, after a call `f(x)` where `x` is
a pointer it assumes that x can point to something else.

With this diff, the default behavior remains the same, but it is also
possible to make Pulse always assume that the arguments are not changed
by unknown functions, by saying `--no-pulse-havoc-arguments`.

Reviewed By: kren1

Differential Revision: D53183246

fbshipit-source-id: d289c06c8022994ba4a6e00f62ffcfceb6c277af