[hack] capture variadic functions with default params
Summary:
In Hack frontend we deal with default params at the `hackc` level by generating a textual implementation for each possible combination of configuration (0 given, 1 given...). We deal with a variadic param by intercepting the call inside Pulse resolution.

But now we want to handle a combination of defaults and variadic param.
```
class C {
  function foo(T1 $opt1 = ..., T2 $opt2 = ..., T ...$args): T { ...}
}
```
The hackc tool generates 3 implementations:
```
define .wrapper foo($this: *C): *T {...}
define .wrapper foo($this: *C, $opt1: *T1): *T {...}
define .wrapper foo($this: *C, $opt1: *T1, $opt2: *T2, $args: .variadic *T): *T {...}
```
When we record implementation in `TextualDecl` we store the variadic version of `foo` (the last of this list) in the special hashmap that we query with function `get_variadic_procdesc`. At call site `foo(receiver,exp1,..., expN)` we must be able to understand which of the 3 versions is used. So we add a test between N and the arity if the variadic version in order to recognize properly which implementation is concerned.

This diff only deal with Textual verification. The capture of `variadic.hack` file here would fail before this diff.

We need to also adapt Pulse analysis in an other diff.

Reviewed By: geralt-encore

Differential Revision:
D53048397

Privacy Context Container: L1208441

fbshipit-source-id: 18974880adf2348d73cf18f8a7da8d1714f70daa