[clang] Add summary specialization for function pointers
Summary:
Just like in previous diff for Objective-C blocks, we want to use summary specialization for function pointers. For that, we create a builtin to deal with the logic of calling the function pointer: if we know the function we call it, if we don't, we request dynamic type specialisation. When we evaluate a constant function, we treat it like a closure without captured variables: we add a DynamicType attribute to it with the name of the function. This can then be used for the dynamic type specialization.

This also removes the old closure specialization (the syntactic one, there is more to delete in future diffs) and a couple of other bits that were now deadcode.

Reviewed By: davidpichardie

Differential Revision: D53134252

fbshipit-source-id: d779c01d814bef55db0d9482dfa9822110084db4