[clang] Collect temporary variables to destruct as a set
Summary: In the clang frontend, it collects temporary variables and injects destruction statements for them. The problem was that the temporary variables were collected as a list, which means it included duplicated variables for some cases, injected multiple destructions for the same variable, then introduced use-after-delete false positives. To avoid the issue, this diff uses a set instead.

Reviewed By: ngorogiannis

Differential Revision: D53350750

fbshipit-source-id: b15692822149bdc7545dd997d1587080757eaf6e