[pulse] Adding static types for the Objective-C classes
Summary: Just as we do with Hack and Python classes, we want to add the StaticType attributes to the parameters of a function which are Objective-C classes. We can use them to check for retain cycles later. Retain cycles are only an issue for ObjC classes, so we need a way to tell that they are, and at the moment we use attributes added at allocation time, which means this restricts the check to only classes that we know have been allocated.

Reviewed By: davidpichardie

Differential Revision: D53530266

fbshipit-source-id: abad6a141e8277a813bc3eebadeec86d73a3c4d6