[hack] Add trait sinit calls in class sinit
Summary:
When a trait `T` initializes a static field and the trait is used from
a class `C`, the static field should be initialized when calling
`C::sinit`, which was not done before this diff.

This diff adds method calls of `T::sinit` at the beginning of
`C::sinit`, when `C` used `T`.

Reviewed By: davidpichardie

Differential Revision: D53579369

fbshipit-source-id: 2d85256d0f4b12f1e535ae3031cfdcc9e89ccd83