[pulse][specialization] refactoring alias specialization
Summary:
we want alias specialization to follow a same approach than dynamic-type specialization. Here we make alias specialization use the iterative process `iter_call`, event if we only expect at most 1 round of iteration for aliases.

It generates a few (surprising) changes on alias+dyntype tests. But this is not really a regression in my opinion since we will not handle them until the end of this diff stack.

Reviewed By: dulmarod

Differential Revision:
D53644639

Privacy Context Container: L1208441

fbshipit-source-id: ad897c170acf4d9eb9e8cce1c1257d837d300d9b