[pulse][specialization] merging control flow of the specialization modes
Summary:
we now merge the control flow between alias and dynamic-type specialization. This is not sufficient to make the combination of specialization happen but the control flow is almost there.

This change flip some FN/FP tests but only on function that requires a combination of alias and dynamic-type specialization.

Reviewed By: skcho

Differential Revision:
D53676051

Privacy Context Container: L1208441

fbshipit-source-id: d410eca3fbaf460ee9dfc7f7edb9c9188373381b