[pulse][specialization] main type becomes a struct
Summary:
A specialization key is not a struct with an `aliases` and a `dynamic_types` field. But we don't roll out the combination yes, because at most one side is not empty for now. We will improve that later.

Again, we see some FN/FP flips in tests about dynamic/alias specialization combination.

Reviewed By: skcho

Differential Revision:
D53676065

Privacy Context Container: L1208441

fbshipit-source-id: da1923c9b1be26624945f9907925b067410c3899