[pulse][specialization] final refactoring to combine alias/dyntype specialization
Summary:
This last touch deals with PulseCallOperation. We make sure the specialization key is enriched layer by layer.

This is the end of refactoring. It solves Hack tests that combine virtual call and aliases, but not Objc tests that combine closures and aliases. This is expected because alias specialization is not strong enough for dealing with alias on heap paths, and closures captured values are accesses through heap paths.

Reviewed By: skcho

Differential Revision:
D53611680

Privacy Context Container: L1208441

fbshipit-source-id: 65d31bc39edd5b25e482070a8635869aa2096793