[pulse][specialization][test] more tests about aliases
Summary:
We reproduce in Hack the problem we observed on ObjC when a combination of alias and dyntype specialization is required.

The primary problem concerns alias specialization which is currently not possible when aliases deal with heap paths. This is a limitation inherited from the the initial design of alias specialization. I would to solve this problem, but the creation of the precondition is not easy to get right.

Reviewed By: skcho

Differential Revision:
D53712200

Privacy Context Container: L1208441

fbshipit-source-id: 0d12e2550e48aa9c51f114d3e9938c87c847767e