[dbwriter] avoid race on deleting socket file
Summary:
When the main Infer process shuts down it sends a terminate command to dbwriter. However, the shutdown of the dbwriter process is asynchronous, so it may happen that the main infer process has terminated, while dbwriter is shutting down. This can lead to a race with another infer process starting up and using the same results directory (eg, capture then analyze), where the socket file is used by the dbwriter of the 2nd run but then deleted by the epilogue of the 1st one.

This diff avoids the race by making the main infer process wait until
- 5 seconds have passed, while checking every second if the socket file has been removed.
- if not, removing it forcibly (assuming dbwriter has crashed)

Reviewed By: skcho

Differential Revision:
D53613615

Privacy Context Container: L1208441

fbshipit-source-id: 9dbbdb3ffcc3aa09d3d9e10b182500bfd4491f4d