[pulse][alias] collecting aliases during summary application
Summary:
Previously alias specialization was performed in 3 steps
 STEP#1: during summary application at call site, throw an exception if one alias is detected (2 different caller adresses associated to the same callee adress)
 STEP#2: catch the exception and compute the alias equivalence classes (between parameters only)
 STEP#3: ask for a specialized summary

Our ultimate goal is to allow alias equivalence between heap paths. Here, we start a a fusion between step#1 and #2 by collecting **all** alias equivalence classes we detect during summary application, so that we will not have to redo the work during step#2.

In this diff, we only do the collect and print it.

During this collection, we may encounter heap path that deal with array indexes. This is not something we can deal with (quite hard...). If an alias occurs on such heap path, we will give up. We have to be careful to no give up too early so we introduce a notion of lazy heap path.

Reviewed By: ngorogiannis

Differential Revision:
D53809166

Privacy Context Container: L1208441

fbshipit-source-id: b1e52969b264e805d5237e8f7c75cb1ef0d847ef