Interprocedural analysis with dependency: retrieve the dep payload from callee
Summary:
The "interprocedural analysis with dependency" mechanism can be used to define an interprocedural
analysis that needs to read the result of another base analysis, in particular on the procedure that
is currently being analysed.

So far, it has only been used in Lineage to base it on the result of LineageShape.

We change its interface so that the `analyze_dependency` function can be used to also retrieve the
payload from the base analysis on dependency procedures. This is similar to what the
`interprocedural2` analyses already do (although those variations cannot be used to retrieve the base
analysis result on the current procedure, only on its dependencies).

This has no actual impact included in this diff. The goal is to be able to use shape information of
callees in Lineage to generate more precise call nodes.

Reviewed By: jvillard

Differential Revision: D53132596

fbshipit-source-id: 3080fc0b4f2ec5e968ae515278ec79e80299fe18