[restart] also schedule static callees
Summary:
When analyzing a relatively small number of procedures, long tail effects show up in the form of a few workers analyzing when the rest do nothing.

To alleviate this, change the restart scheduling so that the scheduling targets are augmented by the level-1 static dependencies of the top-level procedures. This means that if a top level procedure is requeued because one of its dependencies is being analyzed, its other (static) dependencies can still be scheduled.

Reviewed By: jvillard

Differential Revision:
D54005785

Privacy Context Container: L1208441

fbshipit-source-id: 6437394726fb9b9adf6def79581e40b45cb62837