[retain cycles] Refactor into separate module
Summary:
We wish to move the retain cycle checker to a separate module. This requires a few changes because we won't be inside AbductiveDomain anymore but calling things from the outside.

Another change is that we now not check for cycles when computing the post but after assignments and method calls. We want to find not leaked cycles but any strong cycle as soon as it is created. This is expected to find more cycles.

Reviewed By: jvillard

Differential Revision: D54062768

fbshipit-source-id: 2fb8f78c34192bb1874a07dd52e3f8010681978a