[pulse] Fix the decompiler for captured variables
Summary:
This got broken when we revamped the closure attributes and fields for captured variables. Fixing it now.

Before, in the field to store captured variables, we were storing the position of the captured variables in the `captured` attributes of the procname, to be able to find the variables again. Now, the name of the captured variable is part of the field name, so we don't need to go via Pulse attributes and procedure attributes to be able to just print the captured variable name.

Reviewed By: jvillard

Differential Revision:
D54195143

Privacy Context Container: L1208441

fbshipit-source-id: 29b444fa7cf4b23379e3f8726fb80a6d8941f97c