[pulse] abort formula computations when running out of fuel
Summary:
When we run out of fuel, we should abort the whole computation. The previous
behaviour was to abort only the current function and return the current
formula, but it could be that the callers of that function would behave
incorrectly in that case, wrongly assuming the work had been fully done.

Reviewed By: davidpichardie

Differential Revision: D54015160

fbshipit-source-id: 7ddcebdfc3c239b2c591199a50b7515d52d2de92