change textual typing of throw again to accommodate unit test
Summary: A previous diff changed the textual typing of throw to expect a PtrStruct (which includes Void, and seems appropriate for the way we compile Hack). But that broke a textual unit test (TextualTransformTest.ml) that passed a `*int`, so here we use Ptr instead, and tweak its definition to include Void. (Since Ptr isn't used elsewhere, changing the definition should be safe.)

Reviewed By: geralt-encore

Differential Revision: D54450646

fbshipit-source-id: ec21a523b82743928dc7f3f361a924bc9a595d3f