[hack][capture] add a test about optional arguments in traits methods
Summary:
We were previously failing on this kind of pattern because the
extra self parameter was not inserted in all versions of a trait function with
optional arguments. The Textual type checker was detecting something strange in
the situation were the function has a variadic arg.

Reviewed By: geralt-encore

Differential Revision:
D54527470

Privacy Context Container: L1208441

fbshipit-source-id: 3c752994a6531d32f6b853f1187b5c418dd7f986