[pulse] Improve the decompiler
Summary:
when we were trying to add an access source, we were giving up if the source is not in the list. But sometimes we do not know the source, when the source is a variable, but we do know the access, when that is the dereference of that variable. In that case, we can retrofit the information about the variable in the decompiler.

This will help printing values in some retain cycles examples.

Reviewed By: skcho

Differential Revision: D54484969

fbshipit-source-id: e4b9cf52b66c40a682615a13d702219897c9a9a5