[pulse] Add blocks as a source to the decompiler
Summary: Currently the only prime sources we add are variables or values returned from a call. But blocks are also values that have a specific name. So we add them to the decompiler when they are added to the heap. This helps with error messages in cases where the block is only passed to a method without being stored in a variable first.

Reviewed By: skcho

Differential Revision: D54486886

fbshipit-source-id: 009400c5eca85cc72a37b38fa5f8793ec63746b5