[pulse] bugfix in pulse_force_continue mode
Summary:
when refactoring this action a few weeks ago, I introduced
a branch condition that was far too strong.

Facebook:
The impact of the transitive-access checker is HUGE. We
were poorly collecting ~70 signals on infer-smoke-linux-www-static-checkers,
and now we get ~1500!

It indicates that the situation were we meet an empty list of disjuncts is quite frequent. We need to work more on that.

Reviewed By: geralt-encore

Differential Revision:
D54581977

Privacy Context Container: L1208441

fbshipit-source-id: a171f88ab68edcf43cda3e9cc43a847e337262fb