naive early evaluation of instanceof
Summary: This is a simple-minded change to `Formula.and_equal_instaceof` that evaluates eagerly if the dynamic type is known at the point we make the assertion. It doesn't integrate with the propagation of equalities etc, as I'm now going to leave that until I've got a proper instanceof domain instead. But this should suffice to check the effects of eager solving on coverage etc.

Reviewed By: skcho

Differential Revision: D54537978

fbshipit-source-id: bb590fd76ac707c84163da14e527cfe091321ef6