[pulse][solver][2/??] better detect when to substitute variables
Summary:
Previously we could miss substituting variables hiding in linear terms like
`[x]`. We also eagerly normalised variables `x` to linear terms `[x]` (since
they are a special case of linear terms). This is inefficient (a singleton map
in a pair instead of a primitive int!) and a bit confusing/verbose. This
changes normalisation to leave single variables alone in general (except
temporarily when it's more convenient not to).

Reviewed By: nbenton

Differential Revision: D54317268

fbshipit-source-id: ce0fb6cfd06d0abc794e5e073b4495aa4ff764de