[infer] Ignore hash of lambda name when deduplication
Summary:
Currently the name of lambda class is like,

```
lambda_<filename>:<line>_<hash>
```

Here, the `<hash>` includes the line number.

OTOH, when deduplicating issues, it tries to make a
location-independent procedure name, since the location, esp. line
number, is very easy to change by irrelevant code additions/deletions.

The problem is that it did not remove the `<hash>` in the lambda class
name. Thus, when a line number is changed, it changed the procedure
names of lambdas, so that,

* all pre-existing issues were addressed as "introduced",
* and at the same time, the all pre-existing issues were addressed as "fixed".

Due to this, the following issue reporting logic was confused.

To fix the issue, this diff removes the `<hash>` when calculating the
location-independent procedure name.

Reviewed By: jvillard

Differential Revision:
D54814044

Privacy Context Container: L1208441

fbshipit-source-id: 4d6bdbca4592810af53b826dffbcd53604a24831