[retain cycles] Improving the traces
Summary: We want to print more in the traces of the retain cycles to make the reports more readable. Now we add each assignment trace for each object involved in the cycle with a title that this is is the part of the trace that corresponds to the assignment of said object. See some examples below, looks more readable.

Reviewed By: ngorogiannis

Differential Revision: D54591839

fbshipit-source-id: 58afc6e17f157e773d1d5a2d47361000095db873